% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-labels.R
\name{adjust_label}
\alias{adjust_label}
\title{Change text alignment in row/column labels in Clustering View.}
\usage{
adjust_label(x, pattern = "(.*? )(\\\\(.*? at .*?\\\\))")
}
\arguments{
\item{x}{character vector}

\item{pattern}{character string specifying the places in which to intervene; see \code{Details}}
}
\value{
A modified character vector.

This is an internal function that inserts whitespace into strings.
In cases where a label contains a drug/cell line name with co-treatment information
added in parentheses, some non-breaking spaces are added to separate the two.
Additionally, a number of non-breaking spaces is added to each string so that
they are all of the same length.

The whitespace is inserted in places identified by \code{pattern}.
This must be a regular expression consisting of two parts (delimited by parentheses)
and the non-breaking spaces are inserted between the two.

Alignment will be perfect only if a monospaced font is used.
}
\description{
Change text alignment in row/column labels in Clustering View.
}
\examples{
adjust_label("Cell Line Name: X Drug Name: A (untreated at 0 &mu;M) Title: 1.00")
# Cell Line Name: X Drug Name: A &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(untreated at 0 &mu;M) Title: 1.00

}
\seealso{
\code{MetricClustering}, \code{plotlyMH}
}
\keyword{utils_label}
