% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-fun-colors.R
\name{assert_RGB_format}
\alias{assert_RGB_format}
\title{Assert whether number may code color in rgb}
\usage{
assert_RGB_format(x)
}
\arguments{
\item{x}{numeric vector describing rgb color}
}
\description{
Assert whether number may code color in rgb
}
\keyword{internal}
