% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-labels.R
\name{build_label}
\alias{build_label}
\title{Build labels for plotly tooltips}
\usage{
build_label(data, view)
}
\arguments{
\item{data}{a data table in which the labels are to be constructed}

\item{view}{which view to prepare label for; character string}
}
\value{
A character vector the same length as the row number of \code{data}.
}
\description{
Build labels for plotly tooltips
}
\examples{
SE <- gDRutils::get_synthetic_data("small")[[1]]
dt <- gDRutils::convert_se_assay_to_custom_dt(SE, assay_name = "Metrics")
build_label(dt, "grid")

}
\keyword{utils_label}
