% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{build_label_clustering}
\alias{build_label_clustering}
\title{Build labels for plotly tooltips for Metric Clustering heatmap}
\usage{
build_label_clustering(data)
}
\arguments{
\item{data}{a data table in which the labels are to be constructed}
}
\value{
A character vector with labels
}
\description{
Build labels for plotly tooltips for Metric Clustering heatmap
}
\keyword{utils_label}
