% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{buildLabel}
\alias{buildLabel}
\title{Build labels for plotly tooltips}
\usage{
buildLabel(data, view)
}
\arguments{
\item{data}{a data table in which the labels are to be constructed}

\item{view}{which view to prepare label for; character string}
}
\value{
A character vector the same length as the row number of \code{data}.
}
\description{
Build labels for plotly tooltips
}
\examples{
SE <- gDRutils::get_synthetic_data("small")[[1]]
dt <- convert_se_assay_to_custom_dt(SE, assay_name = "Metrics")
buildLabel(dt, "grid")

}
\keyword{utils}
