% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-associations.R
\name{calc_assoc}
\alias{calc_assoc}
\title{Calculate linear associations}
\usage{
calc_assoc(X, Y)
}
\arguments{
\item{X}{\code{matrix} dependent variables data matrix (rows are samples, columns are features).
Must have the same number of rows as matrix \code{Y} or equal to length of vector \code{Y}}

\item{Y}{\code{vector} or \code{matrix} experimental response data (rows are samples).
When \code{Y} is a matrix must have the same number of rows as matrix \code{X};
when \code{y} is a vector - its length has to be equal to number of rows in matrix \code{X}.}
}
\value{
\code{data.table} with calculated linear associations
}
\description{
Calculate the linear model associations between dependent variables and response variable(s) of interest.
}
\note{
inspired by the \code{calc_assoc} function written by James Hawley
}
\examples{
X <- matrix(rep(1:13, length.out = 42), nrow = 6,
            dimnames = list(sprintf("row_\%s", 1:6), sprintf("feat_\%s", 1:7)))
Y <- matrix(c(10:15, 110:115, 210:215), ncol = 3,
            dimnames = list(sprintf("row_\%s", 1:6), sprintf("met_\%s", 1:3)))
tab_assoc <- calc_assoc(X, Y)

}
\keyword{internal}
