% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{calc_up_limes}
\alias{calc_up_limes}
\title{Calculate superior limits for series of sorted values}
\usage{
calc_up_limes(x, cnames = c("ux", "ulx"))
}
\arguments{
\item{x}{numeric vector}

\item{cnames}{charvec names of the columns in the output data.table}
}
\value{
data.table with superior limit for series of sorted values
}
\description{
Calculate superior limits for series of sorted values
}
\keyword{internal}
