% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-colors.R
\name{change_color_to_hex}
\alias{change_color_to_hex}
\title{Change color name into hexadecimal string}
\usage{
change_color_to_hex(col_name)
}
\arguments{
\item{col_name}{string representing a valid color listed by \code{\link[grDevices]{colors}}}
}
\value{
hexadecimal string representing color
}
\description{
Change color name into hexadecimal string
}
\examples{
change_color_to_hex("darkblue")
change_color_to_hex("indianred2")
change_color_to_hex("seagreen")

}
\keyword{utils_color}
