% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{change_NA_into_char}
\alias{change_NA_into_char}
\title{Change NA into given string}
\usage{
change_NA_into_char(x, lbl_NA = "NA")
}
\arguments{
\item{x}{vector with items suspected of being NA}

\item{lbl_NA}{string - replacement for NA - as default "NA"}
}
\value{
character (for NA -> given string)
}
\description{
Change NA into given string
}
\examples{
change_NA_into_char(c(1, NA, 3))
change_NA_into_char(c("a", NA, "b"), lbl_NA = "missing")

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
