% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-internals.R
\name{coerce_cotreatment_data}
\alias{coerce_cotreatment_data}
\title{internal function}
\usage{
coerce_cotreatment_data(data)
}
\arguments{
\item{data}{data.table potentially containing cotreatment data}
}
\value{
side effect of coercing any relevant cotreatment variables to the appropriate 'type'
}
\description{
internal function
}
\examples{
# lets prepare some dummy date which will mimic combination data
dt <- data.table::data.table(x = c("1","0.1"), y = c("a", "b"))
names(dt) <- 
   gDRutils::get_prettified_identifiers(c("concentration2", "drug_name2"), simplify = FALSE)
dt
coerce_cotreatment_data(dt)
dt

}
\seealso{
\code{MetricClustering}, \code{buildLabelClustering}
}
\keyword{utils}
