% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{colorToHex}
\alias{colorToHex}
\title{Change color name into hexadecimal string}
\usage{
colorToHex(col_name)
}
\arguments{
\item{col_name}{string representing a valid color listed by \code{\link[grDevices]{colors}}}
}
\value{
hexadecimal string representing color
}
\description{
Change color name into hexadecimal string
}
\examples{
colorToHex("darkblue")
colorToHex("indianred2")
colorToHex("seagreen")

}
\keyword{utils}
