% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{convert_factor_to_character}
\alias{convert_factor_to_character}
\title{Convert data.table factor columns to character}
\usage{
convert_factor_to_character(tbl)
}
\arguments{
\item{tbl}{data.table}
}
\value{
data.table with converted columns
}
\description{
Convert data.table factor columns to character
}
\examples{
dt <- data.table::data.table(
  col1 = c("1","2","3"), 
  col2 = c("a", "b", "a"), 
  stringsAsFactors = TRUE)
dt
convert_factor_to_character(dt)

}
\keyword{internal}
