% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaply-helpers.R
\name{create_color_palette}
\alias{create_color_palette}
\title{Create color palette that takes into account 0 as a mid point with neutral color}
\usage{
create_color_palette(colors_vec, limits, breaks = 0.05)
}
\arguments{
\item{colors_vec}{colors to interpolate; must be a valid argument to `col2rgb()`. 
Color in the middle of the vector will be used as a neutral color that represent zero value}

\item{limits}{a two dimensional numeric vector specifying the data range for the scale 
to generate vector of colors}

\item{breaks}{a number of breaks used to divide limits into part for which colors will be generated}
}
\value{
vector of RGB colors
}
\description{
Please note that for a completely positive or completely negative range, the neutral color 
will be applied to the value closest to zero,
}
\examples{
test_colors <- c("blue", "white", "red")
create_color_palette(colors_vec = test_colors, limits = c(1, 4), breaks = 1)
# "#FFFFFF" "#FF7F7F" "#FF0000"

create_color_palette(colors_vec = test_colors, limits = c(0, 3), breaks = 1)
# "#FFFFFF" "#FF7F7F" "#FF0000"

create_color_palette(colors_vec = test_colors, limits = c(-1.5, 2), breaks = 0.5)
# "#0000FF" "#5555FF" "#AAAAFF" "#FFFFFF" "#FFAAAA" "#FF5555" "#FF0000"

}
\keyword{heatmaply}
