% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{create_formula}
\alias{create_formula}
\title{Create formula}
\usage{
create_formula(cell_id, drug_id, cotrt_drug_ids)
}
\arguments{
\item{cell_id}{String of cell identifier column.}

\item{drug_id}{String of drug identifier column.}

\item{cotrt_drug_ids}{Character vector of any cotreatment drug experimental design identifier columns.}
}
\value{
formula accounting for any cotreatment columns
}
\description{
Create formula
}
\keyword{internal}
