% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{do_show_legend}
\alias{do_show_legend}
\title{Helper function to evaluate if legend should be added to the figure}
\usage{
do_show_legend(var_col, data)
}
\arguments{
\item{var_col}{name of column used for coloring dots}

\item{data}{\code{data.table} prepared by \code{prepareDataMC}}
}
\value{
logical value
}
\description{
Current logic is as follows: show legend only if:
(1) user colors by "Drug MOA" or "Primary Tissue" or
(2) there is a codrug data with at least two unique values
}
\examples{
dt <- data.table::data.table(
  x = seq_len(10),
  y = seq_len(10) + 0.5,
  `Tissue` = c("lung", "brain")
)
do_show_legend("none", dt)
do_show_legend("Tissue", dt)

}
\keyword{utils}
