% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-associations.R
\name{.calc_assoc_matrix}
\alias{.calc_assoc_matrix}
\title{Calculate the linear model associations between dependent variables and response variables of interest.}
\usage{
.calc_assoc_matrix(X, Y)
}
\arguments{
\item{X}{\code{matrix} Dependent variables data matrix (rows are samples, columns are features).
Must have the same length as \code{Y}.}

\item{Y}{\code{matrix} Experimental response data (rows are samples).
Must have the same length as \code{X}.}
}
\value{
\code{data.table}
}
\description{
Calculate the linear model associations between dependent variables and response variables of interest.
}
\seealso{
\code{\link[cdsrmodels]{lin_associations}}
}
\author{
James Hawley
}
\keyword{internal}
