% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{.calc_dose_rank}
\alias{.calc_dose_rank}
\title{Calculate Dose Ranks}
\usage{
.calc_dose_rank(vals)
}
\arguments{
\item{vals}{\code{numeric} vector of concentration values.}
}
\value{
A \code{character} vector of ranks.
}
\description{
Helper function to rank unique positive concentration values.
Zero values or NAs are assigned a rank of "-".
}
\author{
Bartosz Czech \email{bartosz.czech@contractors.roche.com}
}
\keyword{internal}
