% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.fill_pheatmap_annotation}
\alias{.fill_pheatmap_annotation}
\title{Prep annotation data.table acc to metric matrix for pheatmap::pheatmat}
\usage{
.fill_pheatmap_annotation(
  dt_anno,
  mat_with_metric,
  anno_var = gDRutils::get_env_identifiers("cellline_name")
)
}
\arguments{
\item{dt_anno}{\code{data.table} that specifies the annotations shown on left side of the heatmap
or shown above the heatmap - depending on the \code{anno_var}.
Each row defines the features for a specific row. The rows in the data and in the annotation
are matched using corresponding names from the required \code{anno_var} column.}

\item{mat_with_metric}{numeric matrix with metric values; must have named rows and columns}

\item{anno_var}{string with variable describing annotation dimension:
one of: \code{CellLineName} for rows or \code{DrugName} for column.}
}
\value{
\code{data.table} with annotation updated to \code{mat_with_metric}
}
\description{
Prep annotation data.table acc to metric matrix for pheatmap::pheatmat
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{internal}
