% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{.format_dose_list}
\alias{.format_dose_list}
\title{Format Dose List for Legend}
\usage{
.format_dose_list(dose_vec, n_per_line)
}
\arguments{
\item{dose_vec}{\code{numeric} vector of unique positive doses.}

\item{n_per_line}{\code{integer} number of items to display per line.}
}
\value{
A single \code{character} string with HTML line breaks.
}
\description{
Helper function to create a formatted string of dose ranks and values 
for the plot legend, breaking lines based on `n_per_line`.
}
\author{
Bartosz Czech \email{czech.bartosz@external.gene.com}
}
\keyword{internal}
