% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism-plotlist.R
\name{.get_info_from_names}
\alias{.get_info_from_names}
\title{Extract drug grid and feature metadata from file names}
\usage{
.get_info_from_names(file_names, normalization_type = "RV")
}
\arguments{
\item{file_names}{character vector of file names following the
pattern \code{prefix__feat_drug_NORM_...}}

\item{normalization_type}{string normalization type ("GR" or "RV")}
}
\value{
named list with \code{drug_grid} and \code{feat_meta}
}
\description{
Extract drug grid and feature metadata from file names
}
\examples{
\donttest{
gDRplots:::.get_info_from_names("plt__metricA_DrugX_RV_extra", "RV")
}

}
\keyword{internal}
