% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.get_pheatmap_fontsize}
\alias{.get_pheatmap_fontsize}
\title{Get fontsize for rownames or colnames in pheatmap::pheatmap}
\usage{
.get_pheatmap_fontsize(
  matrix,
  dimension = c("row", "col"),
  threshold_count = 40L
)
}
\arguments{
\item{matrix}{numeric matrix with metric values.}

\item{dimension}{character value, either "row" or "col", indicating whether to compute fontsize for rows or columns.}

\item{threshold_count}{integer value of the number of rows/columns for which the font size remains standard.}
}
\value{
numeric value of font size.
}
\description{
Get fontsize for rownames or colnames in pheatmap::pheatmap
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{internal}
