% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricContrast.R
\name{.get_symbol_list}
\alias{.get_symbol_list}
\title{List of symbols for plotly}
\usage{
.get_symbol_list()
}
\description{
Since the markers will be colored, it is important that they are shapes with a area.
\code{plotly} has own supported list and markers with area are type \code{-dot},
but not \code{-open-dot}

vals <- plotly::schema(FALSE)$traces$scatter$attributes$marker$symbol$values
vals <- vals[!grepl("-open-dot", vals)]
vals <- grep("-dot$", vals, value = TRUE)

Note: item "hash-dot" was removed from list as one without area.
}
\seealso{
[Working with symbols](https://plotly-r.com/working-with-symbols)
}
\keyword{internal}
