% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_tile_size}
\alias{.get_tile_size}
\title{Calculate size of tiles based on pos_x/pos_y values}
\usage{
.get_tile_size(pos_vec)
}
\arguments{
\item{pos_vec}{numeric vector with pos_x or pos_y values}
}
\value{
size of tile in \code{ggplot2::geom_tile}
}
\description{
Since \code{ggplot2::geom_tile} uses the center of the tile and its size (x, y, width, height),
x and y are given as pos_x and pos_y, it is required to calculate the width and height.
}
\keyword{internal}
