% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-associations.R
\name{.order_assoc_result}
\alias{.order_assoc_result}
\title{Order and rename columns in associations results}
\usage{
.order_assoc_result(res_dt)
}
\arguments{
\item{res_dt}{\code{data.table} of associations results}
}
\value{
\code{data.table} of associations results with reorder and rename columns
}
\description{
Order and rename columns in associations results
}
\keyword{internal}
