% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.pheatmap_cluster_param}
\alias{.pheatmap_cluster_param}
\title{Compute value of param cluster_rows or cluster_cols in pheatmap::pheatmap}
\usage{
.pheatmap_cluster_param(
  mat_to_cluster,
  cluster_flag = TRUE,
  distfun = stats::dist,
  additional_condition = TRUE
)
}
\arguments{
\item{mat_to_cluster}{numeric matrix to be clustered; cluster dimension must be named}

\item{cluster_flag}{logical flag whether rows/should be clustered;}

\item{distfun}{function used to compute the distance (dissimilarity) between rows;
defaults to \code{\link[stats]{dist}} using euclidean euclidean.}

\item{additional_condition}{additional logical flag whether rows/columns should be clustered (will be included 
as the logical sum of \code{cluster_flag} and \code{additional_condition})}
}
\value{
logical flag determining if rows should be clustered or \code{hclust} object.
}
\description{
The \code{cluster_rows} and \code{cluster_cols} parameters  pheatmap::pheatmap may take values:
- boolean determining if rows/columns should be clustered
- \code{hclust} object,
this function allows to calculate proper value taking into account matrix, additional condition and 
selected function used to compute the distance in \code{hclust} object
}
\details{
To compute the correct value when clustering columns - use the transposed source matrix as \code{mat_to_cluster}
}
\examples{
\dontrun{
mat <- matrix(1:24, nrow = 4)
rownames(mat) <- sprintf("row_\%s", 1:4)
colnames(mat) <- sprintf("col_\%s", 1:6)
.pheatmap_cluster_param(mat)
.pheatmap_cluster_param(t(mat)) |> plot()
.pheatmap_cluster_param(t(mat), distfun = compute_distances) |> plot()

mat[2,2] <- NA
mat[2,1] <- Inf
.pheatmap_cluster_param(mat)
.pheatmap_cluster_param(mat, distfun = compute_distances)
.pheatmap_cluster_param(t(mat), distfun = compute_distances)
add_cond <- NCOL(mat) > 10
.pheatmap_cluster_param(mat, distfun = compute_distances, additional_condition = add_cond)
}

}
\seealso{
\code{\link{compute_distances}}
}
\keyword{internal}
