% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{.prep_dt_depmap}
\alias{.prep_dt_depmap}
\title{Load DepMap merged data and metadata}
\usage{
.prep_dt_depmap(
  feature_sets = feature_sets <- c("CRISPRGeneEffect",
    "OmicsExpressionProteinCodingGenesTPMLogp1", "OmicsExpressionSignatures",
    "OmicsSomaticMutationsMatrixHotspot", "OmicsSomaticMutationsMatrixDamaging",
    "OmicsCNGene"),
  prefix = c("KO_", "GE_", "SG_", "HM_", "DM_", "CN_"),
  metadata_cols = c("OncotreeLineage", "PatientRace", "CCLEName")
)
}
\arguments{
\item{feature_sets}{character vector with the molecular feature set(s) to load from DepMap.}

\item{prefix}{character vector with the prefixes to use for the each feature set in \code{feature_sets};
has to be the same length as \code{feature_sets}}

\item{metadata_cols}{character vector with the metadata columns to load for DepMap cell lines}
}
\value{
\code{data.table} with merged data and meta
}
\description{
Load DepMap merged data and metadata
}
\examples{
\dontrun{
dt_depmap <- .prep_dt_depmap() 
}

}
\seealso{
\code{\link[kaleidoscope]{load_depmap_merged}}
}
\keyword{internal}
