% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{.prep_dt_response_metric_sa}
\alias{.prep_dt_response_metric_sa}
\title{Prep table with metric values for single-agent experiment}
\usage{
.prep_dt_response_metric_sa(
  dt_metrics,
  d_name,
  normalization_type = "RV",
  metric = "xc50",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{d_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")}

\item{fit_source}{string source name for metrics}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{prep_dt_assoc}}
}
\description{
Prep table with metric values for single-agent experiment
}
\examples{
mae <- dsassembly::getDataset(id = "DS000015225")
se <- mae[["single-agent"]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")
d_name <- "G03488889"
dt_response <- .prep_dt_response_metric_sa(dt_metrics, d_name)
dt_response <- .prep_dt_response_metric_sa(dt_metrics, d_name, 
                                           metric = c("xc50", "x_mean", "x_max"))

}
\keyword{prism_plots}
