% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{estimate_plot_size}
\alias{estimate_plot_size}
\title{Estimate the optimal plot size (either ggplot or pheatmap) for saving plots}
\usage{
estimate_plot_size(plt, base_width = 10, base_height = 6, scale_factor = 0.5)
}
\arguments{
\item{plt}{a ggplot or pheatmap object}

\item{base_width}{an integer with default base_width}

\item{base_height}{an integer with default base_height}

\item{scale_factor}{an integer with default scale_factor}
}
\value{
named vector with optimal width and height used in \code{\link[ggplot2]{ggsave}} function
}
\description{
Estimate the optimal plot size (either ggplot or pheatmap) for saving plots
}
\examples{
p <- ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) + ggplot2::geom_point()
estimate_plot_size(p)

}
\author{
Bartosz Czech \email{bartosz.czech@contractors.roche.com}
}
\keyword{internal}
