% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{estimate_render_size}
\alias{estimate_render_size}
\title{Estimate the optimal plot size (either ggplot or pheatmap) for rendering plots}
\usage{
estimate_render_size(plt_list)
}
\arguments{
\item{plt_list}{named list with generated plots to be shown in tabs}
}
\value{
list with height and width for plots fom given list \code{plt_list}
}
\description{
Estimate the optimal plot size (either ggplot or pheatmap) for rendering plots
}
\keyword{internal}
