% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{filter_data_by_barcode}
\alias{filter_data_by_barcode}
\title{Helper function to filter data by barcode and prepare it for plotting}
\usage{
filter_data_by_barcode(data, barcode, barcode_idf)
}
\arguments{
\item{data}{The original data table}

\item{barcode}{The barcode to filter the data by}

\item{barcode_idf}{The identifier for the barcode column}
}
\value{
A subset of the data filtered by the specified barcode
}
\description{
Helper function to filter data by barcode and prepare it for plotting
}
