% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-heatmaply.R
\name{force_heatmaply_limits}
\alias{force_heatmaply_limits}
\title{Force setting limits for colorbar in heatmaply}
\usage{
force_heatmaply_limits(heatmaply, limits)
}
\arguments{
\item{heatmaply}{object of class `heatmapr`}

\item{limits}{a two dimensional numeric vector specifying the data range for the scale}
}
\value{
object of `heatmapr` class with specified data range
}
\description{
Force setting limits for colorbar in heatmaply
}
\examples{
heatmaply <- heatmaply::heatmaply(mtcars)
heatmaply$x$data_index <- 4
force_heatmaply_limits(heatmaply, c(100, 400))


}
\keyword{heatmaply}
