% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-plotly.R
\name{gDR_plotly_config}
\alias{gDR_plotly_config}
\title{Define tools in bar mode for plotly plot}
\usage{
gDR_plotly_config(plt, editable = FALSE, ...)
}
\arguments{
\item{plt}{plotly object}

\item{editable}{logical determines whether the graph is editable or not}

\item{...}{other parameters that may be passed to \code{\link[plotly]{config}}}
}
\value{
plotly plot with gDR config
}
\description{
see: https://github.com/plotly/plotly.js/blob/master/src/plot_api/plot_config.js
}
\examples{
plt <- plotly::plot_ly(mtcars, x = ~mpg, type = "histogram")
plt
gDR_plotly_config(plt) 

}
\keyword{plotly_config}
