% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{generate_color_mappings}
\alias{generate_color_mappings}
\title{Helper function to generate color mappings for Gnumber and Gnumber_2}
\usage{
generate_color_mappings(
  dt_plate_subset,
  untrt_tag = gDRutils::get_env_identifiers("untreated_tag")
)
}
\arguments{
\item{dt_plate_subset}{The subset of data}

\item{untrt_tag}{Untreated tag identifier}
}
\value{
A list containing color mappings for Gnumber and Gnumber_2
}
\description{
Helper function to generate color mappings for Gnumber and Gnumber_2
}
\keyword{internal}
