% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricDistribution.R
\name{get_axis_min}
\alias{get_axis_min}
\title{get min value for the axis based on the data to be visualized}
\usage{
get_axis_min(data, nearest_shift = 0.1)
}
\arguments{
\item{data}{numeric vector}

\item{nearest_shift}{single number nearest value to which round (floor)}
}
\value{
single number
}
\description{
get min value for the axis based on the data to be visualized
}
\keyword{internal}
