% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-colors.R
\name{get_col_luminance}
\alias{get_col_luminance}
\title{Calculate the luminance of a colour}
\usage{
get_col_luminance(col_name)
}
\arguments{
\item{col_name}{string representing a valid colour}
}
\value{
single element numeric vector
}
\description{
Calculate the luminance of a colour
}
\examples{
get_col_luminance("blue")
get_col_luminance("red")
get_col_luminance("#000000")
get_col_luminance("#906090")
get_col_luminance("#906090F2")

}
\keyword{utils_color}
