% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{get_combo_base_data}
\alias{get_combo_base_data}
\title{get base combo data of given filters}
\usage{
get_combo_base_data(
  data,
  drug1_name,
  drug2_name,
  cell_line,
  c_assay = names(gDRutils::get_combo_base_assay_names())[[1]],
  normalization_type = "RV"
)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{drug1_name}{string a name of the first drug to be selected}

\item{drug2_name}{string a name of the second drug to be selected}

\item{cell_line}{string a name of cell line to be selected}

\item{c_assay}{character string specifying the combo base assay to be selected,
any from \code{names(gDRutils::get_combo_excess_field_names())} will do}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\description{
get base combo data of given filters
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
combo_data_l <- gDRutils::convert_combo_data_to_dt(mae[[gDRutils::get_supported_experiments("combo")]])

get_combo_base_data(
 data = combo_data_l,
 drug1_name = "drug_004",
 drug2_name = "drug_021",
 cell_line = "cellline_GB",
 c_assay = "bliss_excess"
)
}
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
\keyword{internal}
