% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{get_combo_score_matrix}
\alias{get_combo_score_matrix}
\title{get score combo data for given filters}
\usage{
get_combo_score_matrix(
  data,
  c_assay = names(gDRutils::get_combo_score_field_names())[[1]],
  normalization_type = "RV"
)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{c_assay}{character string specifying the combo score assay to be selected,
any from \code{names(gDRutils::get_combo_score_field_names())} will do}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
matrix
}
\description{
get score combo data for given filters
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
combo_data_l <- gDRutils::convert_combo_data_to_dt(se)
get_combo_score_matrix(combo_data_l, "bliss_score")
}

}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
\keyword{internal}
