% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{get_drug_axes}
\alias{get_drug_axes}
\title{get drug axes}
\usage{
get_drug_axes(dt, value_col = "smooth")
}
\arguments{
\item{dt}{data.table}

\item{value_col}{string default value is `smooth`}
}
\description{
define the position to nicely plot the combination matrix such that the marginal (single agent)
  are slightly offset from the combination matrix and that the tiles are properly spaced.
  This is also required to get the position to plot the isobolograms
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
\keyword{internal}
