% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{get_hm_title}
\alias{get_hm_title}
\title{Get Legend Title}
\usage{
get_hm_title(metric = "xc50", normalization_type = "GR", dataset_name = NULL)
}
\arguments{
\item{metric}{string name of the metric
one of: "xc50"("GR50" or "IC50" - respectively depending on \code{normalization_type}),
"x_max" ("GR Max" or "E Max") or x_mean" ("GR Mean" or "RV Mean")}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{dataset_name}{string name of dataset}
}
\value{
character title for heatmap
}
\description{
Get Legend Title
}
\examples{
get_hm_title(dataset_name = "Dateset DX123",
             metric = "x_mean",
             normalization_type = "GR")

get_hm_title(metric = "xc50",
             normalization_type = "GR")
             
}
\keyword{pheat_ann}
