% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{get_isobologram_data}
\alias{get_isobologram_data}
\title{get isobologram combo data for given filters}
\usage{
get_isobologram_data(
  data,
  drug1_name,
  drug2_name,
  cell_line,
  drug1_axis,
  drug2_axis,
  c_assay = names(gDRutils::get_combo_assay_names(group = "combo_iso"))[[1]],
  normalization_type = "RV",
  iso_levels = c("0.5", "0.75")
)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{drug1_name}{string a name of the first drug to be selected}

\item{drug2_name}{string a name of the second drug to be selected}

\item{cell_line}{string a name of cell line to be selected}

\item{drug1_axis}{data.table with drug1 data from \code{get_combo_base_data}}

\item{drug2_axis}{data.table with drug2 data from \code{get_combo_base_data}}

\item{c_assay}{character string specifying the combo iso assay to be selected,
any from \code{gDRutils::get_combo_assay_names(group = "combo_iso")} will do}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{charvec isobologram levels to be selected}
}
\description{
get isobologram combo data for given filters
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
\keyword{internal}
