% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{get_legend_title}
\alias{get_legend_title}
\title{Get Legend Title}
\usage{
get_legend_title(
  var,
  has_codrug_data = FALSE,
  default_var = gDRutils::get_prettified_identifiers("concentration2", simplify = TRUE)
)
}
\arguments{
\item{var}{character name of the variable to be shown in legend}

\item{has_codrug_data}{logical whether the data is combo type}

\item{default_var}{character name of the variable to be shown in legend when \code{var} is not selected}
}
\value{
single element list with character in HTML format
}
\description{
prepare legend title formatted as HTML code
}
\examples{
get_legend_title("Drug")
get_legend_title("Drug 2", has_codrug_data = TRUE, default_var = "Conc_2")

}
\keyword{utils}
