% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-const.R
\name{get_metrics_to_transform}
\alias{get_metrics_to_transform}
\title{Get names of metrics that should be presented on a log scale}
\usage{
get_metrics_to_transform()
}
\value{
character vector of standard metric names
}
\description{
Get names of metrics that should be presented on a log scale
}
\examples{
get_metrics_to_transform()

}
\keyword{utils}
