% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-plotly.R
\name{get_plotly_edits}
\alias{get_plotly_edits}
\title{Get list of items that are editable in plotly plots}
\usage{
get_plotly_edits()
}
\value{
list of character vectors with editable items in plotly plots
}
\description{
Get list of items that are editable in plotly plots
}
\examples{
get_plotly_edits()

}
\keyword{plotly_config}
