% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-fun-colors.R
\name{get_qual_colors}
\alias{get_qual_colors}
\title{Create list of qualitative colors}
\usage{
get_qual_colors(n = NULL)
}
\arguments{
\item{n}{number of required colors}
}
\value{
vector with hex colors from qualitative palettes
}
\description{
Create list of qualitative colors
}
\examples{
get_qual_colors()
get_qual_colors(0)
get_qual_colors(5)
get_qual_colors(35)

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{utils_color}
