% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{getColLuminance}
\alias{getColLuminance}
\title{Calculate the luminance of a colour}
\usage{
getColLuminance(col_name)
}
\arguments{
\item{col_name}{string representing a valid colour}
}
\value{
single element numeric vector
}
\description{
Calculate the luminance of a colour
}
\examples{
getColLuminance("blue")
getColLuminance("red")
getColLuminance("#000000")
getColLuminance("#906090")
getColLuminance("#906090F2")

}
\keyword{utils}
