% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{getLongest}
\alias{getLongest}
\title{Find longest factor}
\usage{
getLongest(x, default = 1L)
}
\arguments{
\item{x}{named list of atomic vectors or a data table (no list columns)}

\item{default}{string or integer-like numeric; name or index of \code{x} that will 
be returned in case of ties}
}
\value{
A character string.
}
\description{
Given a list of vectors, return name of the one with the most unique values (levels for factors).
}
\examples{
factorList <- list("letters" = letters, "numbers" = seq_len(10))
getLongest(factorList)

}
\keyword{internal}
