% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{heatmap_combo_metrics_panel}
\alias{heatmap_combo_metrics_panel}
\title{Plot panel of heatmaps of fitted values for combination metrics data}
\usage{
heatmap_combo_metrics_panel(
  dt_excess,
  dt_isobolograms = NULL,
  drug1_name,
  drug2_name,
  cl_name,
  normalization_type = "GR",
  iso_levels = c("0.25", "0.5", "0.75"),
  colors_vec_smooth = NULL,
  colors_vec_excess = NULL,
  no_breaks = 50,
  as_panel = TRUE,
  swap_axes = FALSE
)
}
\arguments{
\item{dt_excess}{data.table representing data from the \code{excess} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "excess")}
and combo \code{SummarizedExperiment}}

\item{dt_isobolograms}{data.table representing data from the \code{isobolograms} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}
and combo \code{SummarizedExperiment}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_name}{string with cell line to be plotted (identifiers \code{CellLineName})}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{character vector with isobologram levels to be selected}

\item{colors_vec_smooth}{character vector of colors (valid names or hex codes) used in the heatmap
for smooth values; the default is the dark purple-light grey palette}

\item{colors_vec_excess}{character vector of colors (valid name or hex codes) used in the heatmap
for excess values; the default is the blue-light grey-red color scale}

\item{no_breaks}{numeric number of breaks on scale}

\item{as_panel}{logical flag whether return list of plot or panel}

\item{swap_axes}{logical flag whether to swap the axes with drugs of the heatmap}
}
\value{
list or panel with heatmaps with value for excess assays for selected drugs and cell line with
   selected isoline and comparison of iso levels
}
\description{
Plot panel of heatmaps of fitted values for combination metrics data
}
\examples{
cl_name <- "cellline_BC"
drug1_name <- "drug_001"
drug2_name <- "drug_026"

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_excess <- gDRutils::convert_se_assay_to_dt(se, "excess")
dt_isobolograms <- gDRutils::convert_se_assay_to_dt(se, "isobolograms")

heatmap_combo_metrics_panel(dt_excess,
                            dt_isobolograms,
                            drug1_name, drug2_name,
                            cl_name,
                            normalization_type = "GR")
                      
cl_name <- "cellline_JE"
drug1_name <- "drug_011"
drug2_name <- "drug_026"

heatmap_combo_metrics_panel(dt_excess,
                            dt_isobolograms,
                            drug1_name, drug2_name,
                            cl_name,
                            normalization_type = "RV",
                            iso_levels = "0.5",
                            as_panel = FALSE)

heatmap_combo_metrics_panel(dt_excess,
                            dt_isobolograms,
                            drug1_name, drug2_name,
                            cl_name,
                            normalization_type = "RV",
                            iso_levels = NULL,
                            as_panel = TRUE,
                            swap_axes = FALSE)

heatmap_combo_metrics_panel(dt_excess,
                            dt_isobolograms,
                            drug1_name, drug2_name,
                            cl_name,
                            normalization_type = "RV",
                            iso_levels = NULL,
                            as_panel = TRUE,
                            swap_axes = TRUE)

}
\keyword{combo_plots}
