% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{heatmap_combo_with_isoref_panel}
\alias{heatmap_combo_with_isoref_panel}
\title{Plot panel of heatmaps with fitted and reference data for isobolograms}
\usage{
heatmap_combo_with_isoref_panel(
  dt_excess,
  dt_isobolograms,
  drug1_name,
  drug2_name,
  cl_names,
  normalization_type = "GR",
  metric = "smooth",
  iso_levels = "0.5",
  colors_vec = NULL,
  no_breaks = 50,
  swap_axes = FALSE
)
}
\arguments{
\item{dt_excess}{data.table representing data from the \code{excess} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "excess")}
and combo \code{SummarizedExperiment}}

\item{dt_isobolograms}{data.table representing data from the \code{isobolograms} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}
and combo \code{SummarizedExperiment}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_names}{character vector with cell line names to be plotted (Cell Line Name);
if \code{NULL} - all available cell lines will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the combo metric;
one of: "smooth" ("Smooth GR" or "Smooth RV" - respectively depending on \code{normalization_type})
"hsa_excess" ("Bliss Excess GR" or "Bliss Excess RV") or "bliss_excess" 
("Bliss Excess GR" or "Bliss Excess RV")}

\item{iso_levels}{character vector with  isobologram levels to be selected;
when \code{NULL} - no isolines will be displayed}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap; 
the default depends on \code{metric}: for "smooth" - the dark purple-light grey palette
and for "hsa_excess" and "bliss_excess" - the blue-light grey-red color scale}

\item{no_breaks}{numeric number of breaks on scale}

\item{swap_axes}{logical flag whether to swap the axes with drugs of the heatmap}
}
\value{
\code{ggplot} object containing panel with heatmaps for fitted values and reference data 
   for isobolograms for selected drug and co-drug and all selected cell line
}
\description{
Plot panel of heatmaps with fitted and reference data for isobolograms
}
\keyword{combo_plots}
