% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{heatmap_combo_with_isoref_qc_panel}
\alias{heatmap_combo_with_isoref_qc_panel}
\title{Plot panel with thefitted values heatmaps and references data for isolobograms 
to control quality of the data}
\usage{
heatmap_combo_with_isoref_qc_panel(
  dt_excess,
  dt_isobolograms,
  drug1_name,
  drug2_name,
  cl_names,
  normalization_type = "GR",
  iso_levels = "0.5",
  colors_vec = NULL
)
}
\arguments{
\item{dt_excess}{data.table representation of the data in \code{excess} assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "excess")}}

\item{dt_isobolograms}{data.table representation of the data in \code{isobolograms} assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_names}{character vector with cell line names names to be plotted (Cell Line NAme); 
if NULL - all available cell liene will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{character vector with  isobologram levels to be selected}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap}
}
\value{
panel with heatmaps for fitted values and references data for isolobograms 
   for selected drug and co-drug by cell line names
}
\description{
Plot panel with thefitted values heatmaps and references data for isolobograms 
to control quality of the data
}
\examples{
drug1_name <- "drug_001"
drug2_name <- "drug_026"

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_excess <- gDRutils::convert_se_assay_to_dt(se, "excess")
dt_isobolograms <- gDRutils::convert_se_assay_to_dt(se, "isobolograms")

cl_names <- unique(dt_excess[["CellLineName"]])[1:4]

heatmap_combo_with_isoref_qc_panel(dt_excess,
                                   dt_isobolograms,
                                   drug1_name, drug2_name,
                                   cl_names,
                                   normalization_type = "GR")

heatmap_combo_with_isoref_qc_panel(dt_excess,
                                   dt_isobolograms,
                                   drug1_name, drug2_name,
                                   cl_names,
                                   iso_levels = c("0.25", "0.5"))
                                   
heatmap_combo_with_isoref_qc_panel(dt_excess,
                                   dt_isobolograms,
                                   drug1_name, drug2_name,
                                   cl_names,
                                   normalization_type = "RV",
                                   iso_levels = NULL,
                                   colors_vec = c("darkcyan", "snow", "darkorange")) 

}
\keyword{QC_plot}
