% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-normalization_qc.R
\name{heatmap_control_mapping_qc}
\alias{heatmap_control_mapping_qc}
\title{Plot drug response curves for single-agent data to control quality of the data}
\usage{
heatmap_control_mapping_qc(dt_treat, dt_controls)
}
\arguments{
\item{dt_treat}{data.table representing data from the \code{RawTreated} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "RawTreated")}}

\item{dt_controls}{data.table representing data from the \code{Controls} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Controls")}}
}
\value{
hetamap of mapping controls to treated
}
\description{
Plot drug response curves for single-agent data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_treat <- gDRutils::convert_se_assay_to_dt(se, "RawTreated")
dt_controls <- gDRutils::convert_se_assay_to_dt(se, "Controls")

heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls)

heatmap_control_mapping_qc(dt_treat = dt_treat[1:1350, ],
                           dt_controls = dt_controls)

dt_treat_NA <- dt_treat[-c(1:135, 270:405),]
heatmap_control_mapping_qc(dt_treat = dt_treat_NA,
                           dt_controls = dt_controls)

dt_controls_NA <- dt_controls[-c(1:305, 611:763, 1221:1750),]
heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls_NA)

heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls[1:3660, ])

}
\keyword{QC_plot}
