% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-fun-colors.R
\name{is_color_dark}
\alias{is_color_dark}
\title{Determine whether or not a color is dark}
\usage{
is_color_dark(col_name)
}
\arguments{
\item{col_name}{string representing a valid color}
}
\value{
logical flag
}
\description{
Determine whether or not a color is dark
}
\examples{
is_color_dark("blue")
is_color_dark("red")
is_color_dark("#000000")

}
\keyword{utils_color}
