% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-colors.R
\name{is_valid_color}
\alias{is_valid_color}
\title{Determine whether or not a color name is valid}
\usage{
is_valid_color(col_name)
}
\arguments{
\item{col_name}{string representing a valid color}
}
\value{
logical flag
}
\description{
A name of color is valid when either is a color name listed by \code{\link[grDevices]{colors}}
or a hexadecimal string of the form \code{#rrggbb}
}
\examples{
is_valid_color("darkblue")
is_valid_color("#FF8C00")
is_valid_color("#FF8C00DC")
is_valid_color("RED")

}
\keyword{utils_color}
