% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-utils.R
\name{isColDark}
\alias{isColDark}
\title{Determine whether or not a colour is dark}
\usage{
isColDark(col_name)
}
\arguments{
\item{col_name}{string representing a valid colour}
}
\value{
logical flag
}
\description{
Determine whether or not a colour is dark
}
\examples{
isColDark("blue")
isColDark("red")
isColDark("#000000")

}
\keyword{utils}
