% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{isValidColor}
\alias{isValidColor}
\title{Determine whether or not a color name is valid}
\usage{
isValidColor(col_name)
}
\arguments{
\item{col_name}{string representing a valid colour}
}
\value{
logical flag
}
\description{
A name of color is valid when either is a color name listed by \code{\link[grDevices]{colors}}
or a hexadecimal string of the form \code{#rrggbb}
}
\examples{
isValidColor("darkblue")
isValidColor("#FF8C00")
isValidColor("#FF8C00DC")
isValidColor("RED")

}
\keyword{utils}
