% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{logSeq}
\alias{logSeq}
\title{create a log-sequence}
\usage{
logSeq(start, end, length)
}
\arguments{
\item{start, end}{numeric, lower and upper margins of the sequence}

\item{length}{integer, resulting sequence length}
}
\value{
A numeric vector, see \code{Details}.
}
\description{
Create a sequence of numbers growing in log-domain.
}
\details{
The result is a numeric vector of length \code{length}.
Differences between items are constant in logarithmic domain
and therefore geometrically increase in linear domain.
}
\seealso{
\code{\link{prepareCurves}}
}
\keyword{internal}
