% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{map_annotations}
\alias{map_annotations}
\title{map identifiers to their corresponding annotation metadata}
\usage{
map_annotations(identifiers, annotation, from)
}
\arguments{
\item{identifiers}{character vector of identifiers to map to annotations, i.e. cell line or drug identifiers}

\item{annotation}{data.table containing column \code{from} and other metadata on the identifiers}

\item{from}{character vector column names in \code{annotation} corresponding to columns in \code{identifiers}.
Order of columns in \code{from} should correspond to order in \code{identifiers}.}
}
\value{
character vector the same length as \code{identifiers} containing the corresponding annotations
for the \code{identifiers}
}
\description{
map identifiers to their corresponding annotation metadata
}
\keyword{internal}
