% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{map_coords}
\alias{map_coords}
\title{Map x,y coordinates from one given data.table to the x,y coords of reference data.table}
\usage{
map_coords(
  ref,
  data,
  r_cols = c("Pos_x", "Pos_y"),
  d_cols = c("Pos_x", "Pos_y")
)
}
\arguments{
\item{ref}{data.table with refrence data}

\item{data}{data.table with data of interest}

\item{r_cols}{two-element charvec with column names for x,y position in reference data.table}

\item{d_cols}{two-element charvec with column names for x,y position in data data.table}
}
\value{
row indices of data.table with x,y coords mapped to x,y coords of data data.table
}
\description{
Map x,y coordinates from one given data.table to the x,y coords of reference data.table
}
\keyword{internal}
