% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{neutralize_spaces}
\alias{neutralize_spaces}
\title{Replace spaces with another character}
\usage{
neutralize_spaces(x, replacement = "_")
}
\arguments{
\item{x}{String where matches are sought}

\item{replacement}{String replacement for spaces}
}
\value{
String with spaces replaced by the specified character
}
\description{
Replace spaces with another character
}
\examples{
neutralize_spaces("GDC-123|Abc x G01234")
neutralize_spaces("MNO-321P 789R YY#1 ")
neutralize_spaces("drug_001 x drug_002", ".")

}
\keyword{internal}
