% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-colors.R
\name{paletteBrew}
\alias{paletteBrew}
\title{Produce a color palette}
\usage{
paletteBrew(n, name, shuffle = FALSE)
}
\arguments{
\item{n}{numeric representing number of colors}

\item{name}{string representing name of a qualitative RColorBrewer palette}

\item{shuffle}{logical representing whether or not to shuffle the colors in the provided palette}
}
\value{
A character vector of hex codes of length \code{n}.

The chosen palette is passed through \code{colorRampPalette}.
Unless \code{n} is higher than the maximum length of the given palette,
the return value depends on the palette itself and the rules
set up in \code{RColorBrewer::brewer.pal}.
Otherwise it is extended and includes intermediate colors.
Visual properties of the palette are likely to be suffer.
}
\description{
Build a color palette of any length from RColorBrewer templates.
}
\examples{
paletteBrew(8, "Accent")
paletteBrew(20, "Accent")
paletteBrew(10, "Set1")

}
\seealso{
\code{RColorBrewer}, \code{colorRampPalette}
}
\keyword{utils_color}
