% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{paletteDisplay}
\alias{paletteDisplay}
\title{Display color palette}
\usage{
paletteDisplay(colors)
}
\arguments{
\item{colors}{character vector of color names or hex codes}
}
\value{
`NULL`
}
\description{
Produce an image of a set of colors for easy viewing.
}
\examples{
paletteDisplay("red")
paletteDisplay("#DD0000")

}
\keyword{internal}
