% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_qc}
\alias{pheatmap_qc}
\title{Plot pretty heatmap for single-agent or combo data to control quality of the data}
\usage{
pheatmap_qc(
  dt_average,
  normalization_type = "GR",
  metric = "x",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = c("black", "grey99"),
  no_breaks = 50,
  cluster_rows = TRUE,
  lbl_by_CellLineName = FALSE,
  lbl_by_DrugName = FALSE
)
}
\arguments{
\item{dt_average}{\code{data.table} representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "x" (value of "GR" or "RV" itself - respectively depending on \code{normalization_type}),
or "x_std" (standard deviation)}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap
note that for \code{metric} "x" the first color will be assigned to the min value, and 
the last one - to the max; for "x_std" - that will be reversed}

\item{no_breaks}{numeric number of breaks on scale}

\item{cluster_rows}{logical flag whether rows should be clustered;
the dendrogram will not be shown for the matrix with any NA value}

\item{lbl_by_CellLineName}{logical flag whether heatmap should be described by CellLineNames instead of clid}

\item{lbl_by_DrugName}{logical flag whether heatmap should be described by DrugName instead of Gnumber}
}
\value{
\code{pheatmap object} containing heatmap for selected metric with annotation - if given
}
\description{
Plot pretty heatmap for single-agent or combo data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]][2:5, ]
dt_average <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Averaged")

hm_1 <- pheatmap_qc(dt_average = dt_average)

hm_2 <- pheatmap_qc(dt_average = dt_average,
                    normalization_type = "RV",
                    colors_vec = c("darkblue", "grey90"),
                    lbl_by_CellLineName = TRUE,
                    lbl_by_DrugName = TRUE)

ggpubr::as_ggplot(hm_1[["gtable"]])
ggpubr::as_ggplot(hm_2[["gtable"]])

se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Averaged")
hm_3 <- pheatmap_qc(dt_average = dt_average,
                    cluster_rows = FALSE)
hm_4 <- pheatmap_qc(dt_average = dt_average,
                    metric = "x_std",
                    cluster_rows = FALSE)

ggpubr::as_ggplot(hm_3[["gtable"]])
ggpubr::as_ggplot(hm_4[["gtable"]])

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\keyword{QC_plot}
