% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_qcs}
\alias{pheatmap_qcs}
\title{Plot pretty heatmap for single-agent or combo data to check quality of data}
\usage{
pheatmap_qcs(
  tab_response,
  metric_growth = "GR",
  metric = "x",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = c("black", "grey99"),
  no_breaks = 50,
  cluster_rows = TRUE,
  lbl_by_CellLineName = FALSE,
  lbl_by_DrugName = FALSE
)
}
\arguments{
\item{tab_response}{\code{data.table} containing drug response metrics
output from \code{\link[gDRutils]{convert_se_assay_to_dt}} for assay "Averaged" 
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{metric_growth}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of metric;
one of: "x" (value of "GR" or "RV" itself - respectively depending on \code{metric_growth}), 
or "x_std" (standard deviation)}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap 
(first colour for min value, last colour - for max value)}

\item{no_breaks}{numeric number of breaks on scale}

\item{cluster_rows}{logical flag whether ows should be clustered}

\item{lbl_by_CellLineName}{logical flag whether heatmap should be describe by CellLineNames instead of clid}

\item{lbl_by_DrugName}{logical flag whether heatmap should be describe by DrugName instead of Gnumber}
}
\value{
heatmap for selected metric with annotation - if given
}
\description{
Plot pretty heatmap for single-agent or combo data to check quality of data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]][2:5, ]
response_metrics <- gDRutils::convert_se_assay_to_dt(se = se, assay_name = "Averaged")

pheatmap_qcs(tab_response = response_metrics)
pheatmap_qcs(tab_response = response_metrics, 
             metric_growth = "RV",
             colors_vec = c("darkblue", "grey90"),
             lbl_by_CellLineName = TRUE,
             lbl_by_DrugName = TRUE)
             
             
se <- mae[[gDRutils::get_supported_experiments("combo")]]
response_metrics <- gDRutils::convert_se_assay_to_dt(se = se, assay_name = "Averaged")
pheatmap_qcs(tab_response = response_metrics,
             cluster_rows = FALSE)
             
pheatmap_qcs(tab_response = response_metrics,
             metric = "x_std",
             cluster_rows = FALSE)

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\keyword{QCS_plot}
