% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_with_anno_cd}
\alias{pheatmap_with_anno_cd}
\title{Plot pretty heatmap with annotations for co-dilution data}
\usage{
pheatmap_with_anno_cd(
  dt_metrics,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = NULL,
  no_breaks = 50,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  distfun = compute_distances,
  annotation_row = NULL,
  annotation_col = NULL,
  annotation_colors = NULL
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by  \code{\link[gDRutils:convert_se_assay_to_dt]{gDRutils::convert_se_assay_to_dt}}
and co-dilution \code{SummarizedExperiment}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}),
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap;
note that the first color will be assigned to the min value, and the last one - to the max}

\item{no_breaks}{numeric number of breaks on scale used for mapping values to colors}

\item{cluster_rows}{logical flag whether indicating rows should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{cluster_cols}{logical flag indicating whether columns should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns;
used for the dendrogram when \code{cluster_rows} or \code{cluster_cols} is set to TRUE
the default is \code{\link{compute_distances}} using Spearman method.}

\item{annotation_row}{\code{data.table} that specifies the annotations shown on the left side
of the heatmap.
Each row defines the features for a specific row. The rows in the data and in the annotation
are matched using corresponding names from the required  \code{DrugName},
\code{DrugName_2} and \code{Concentration_2} columns.
Note that color schemes takes into account if the variable is continuous or discrete.}

\item{annotation_col}{\code{data.table} that specifies the annotations shown above the heatmap.
Each row defines the features for a specific column. The columns in the data and in the annotation
are matched using corresponding names from the required  \code{CellLineName} column.
Note that color schemes takes into account if the variable is continuous or discrete.}

\item{annotation_colors}{named list for specifying \code{annotation_col} and \code{annotation_row}
track colors manually; note list is named with annotation name (column names of \code{annotation_row} -
without \code{DrugName} and column names of \code{annotation_col} -
without \code{CellLineName}), each list item is named vector with valid color name for
each value described in \code{annotation_row} and in \code{annotation_col}, respectively.
Not described elements will be colored by default.}
}
\value{
A named list with elements:
\itemize{
  \item \code{data} a list containing the information visualized in the heatmap:
    \itemize{
      \item \code{matrix} data shown in the heatmap for the selected metric.
      \item \code{annotation_row} a table with row annotations (for \code{DrugName}), if provided.
      \item \code{annotation_col} a table with column annotations (for \code{CellLineName}), if provided.
    }
  \item \code{pheatmap} object containing the heatmap itself.
}
}
\description{
Plot pretty heatmap with annotations for co-dilution data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_codilution_small")
se <- mae[[gDRutils::get_supported_experiments("cd")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")

output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics)
hm_1 <- output[["heatmap"]]
patchwork::wrap_elements(hm_1[["gtable"]])

annotation_manual_col <-
  unique(dt_metrics[, c("CellLineName", "Tissue"), with = FALSE])
annotation_manual_row <-
  unique(dt_metrics[, c("DrugName", "DrugName_2", "Concentration_2",
                        "drug_moa", "drug_moa_2"),
                    with = FALSE])
annotation_map <-
  get_ann_color_map(unique(dt_metrics[, c("Tissue", "drug_moa", "drug_moa_2"), with = FALSE]))

output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics,
                                normalization_type = "RV",
                                metric = "x_mean",
                                colors_vec = c("darkblue", "grey90"),
                                annotation_row = annotation_manual_row,
                                annotation_col = annotation_manual_col,
                                annotation_colors = annotation_map)
hm_2 <- output[["heatmap"]]
patchwork::wrap_elements(hm_2[["gtable"]])

annotation_manual <- data.table::data.table(
  CellLineName =
    c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
  mut_A = c(1, 1, 1, 0, 0),
  mut_B = c("yes", "yes", "no", "no", "no")
)
annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90")
)

output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics,
                                annotation_col = annotation_manual,
                                annotation_colors = annotation_map,
                                hm_title = get_hm_title(
                                  normalization_type = "GR",
                                  metric = "hsa_score",
                                  dataset_name = "Co-dilution data"))
hm_3 <- output[["heatmap"]]
patchwork::wrap_elements(hm_3[["gtable"]])

}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap}}
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{pheat_ann}
