% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_with_anno_combo_metrics}
\alias{pheatmap_with_anno_combo_metrics}
\title{Plot heatmap with annotations for combo metrics data}
\usage{
pheatmap_with_anno_combo_metrics(
  dt_metrics,
  dt_metrics_capped = NULL,
  normalization_type = "GR",
  metric = "x_mean",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = NULL,
  no_breaks = 50,
  cluster_rows = FALSE,
  cluster_cols = TRUE,
  distfun = compute_distances,
  annotation_col = NULL,
  annotation_colors = NULL,
  max_hm_lbl_length = gDRutils::get_settings_from_json("MAX_HM_LBL_LENGTH",
    system.file(package = "gDRplots", "settings.json"))
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{metrics} assay,
outputted by \code{\link[gDRutils:convert_se_assay_to_dt]{gDRutils::convert_se_assay_to_dt}}.
Must contain \code{DrugName}, \code{DrugName_2}, \code{cotrt_value}, and \code{CellLineName}.}

\item{dt_metrics_capped}{\code{data.table} representing data from the \code{"Metrics"} assay,
the same as \code{dt_metrics} but with capped values with 
\code{\link[gDRutils:cap_assay_infinities]{gDRutils::cap_assay_infinities}}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean"); 
but the values from any numeric column can be displayed.}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap;
note that the first color will be assigned to the min value, and the last one - to the max}

\item{no_breaks}{numeric number of breaks on scale used for mapping values to colors}

\item{cluster_rows}{logical flag whether rows should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{cluster_cols}{logical flag whether columns should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns;
used for the dendrogram when \code{cluster_rows} or \code{cluster_cols} is set to TRUE
the default is \code{\link{compute_distances}} using Spearman method.}

\item{annotation_col}{\code{data.table} that specifies the annotations shown above the heatmap.
Each row defines the features for a specific column. The columns in the data and in the annotation
are matched using corresponding names from the required  \code{CellLineName} column.
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_colors}{named list for specifying \code{annotation_col} and \code{annotation_row} 
track colors manually; note list is named with annotation name (column names of \code{annotation_row} - 
without \code{DrugName} and column names of \code{annotation_col} - 
without \code{CellLineName}), each list item is named vector with valid color name for 
each value described in \code{annotation_row} and in \code{annotation_col} - respectively.
Not described elements will be colored in default.}

\item{max_hm_lbl_length}{numeric value for the maximum number of characters in the label;
if set to Inf, no trimming will be performed; for better readability, it is recommended to use 
the default number.}

\item{annotation_row}{\code{NULL}. This parameter is excluded from this function as row annotations 
are auto-generated based on the combination components.}
}
\value{
A named list with elements:
\itemize{
  \item \code{data} a list containing the information visualized in the heatmap:
    \itemize{
      \item \code{matrix} data shown in the heatmap for the selected metric.
      \item \code{annotation_row} a table with row annotations (auto-generated).
      \item \code{annotation_col} a table with column annotations, if provided.
    }
  \item \code{heatmap} the \code{pheatmap} object.
}
}
\description{
Plots a heatmap for metrics data (e.g., "xc50", "x_inf") covering both Single Agents and Combinations.
Rows are sorted alphabetically by the primary drug, prioritizing Single Agent conditions first,
followed by combinations sorted by co-treatment name and concentration.
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")

output <- pheatmap_with_anno_combo_metrics(dt_metrics = dt_metrics,
                                           normalization_type = "RV",
                                           metric = "x_mean")
hm_1 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_1[["gtable"]])

annotation_manual <- data.table::data.table(
  CellLineName =
    c("cellline_AA", "cellline_GB", "cellline_HB"),
  mut_A = c(0, 1, 1),
  mut_B = c("yes", "yes", "no")
)
annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90")
)

output <- pheatmap_with_anno_combo_metrics(dt_metrics = dt_metrics,
                                           metric = "ec50",
                                           normalization_type = "GR",
                                           annotation_col = annotation_manual,
                                           annotation_colors = annotation_map,
                                           hm_title = get_hm_title(
                                             normalization_type = "GR",
                                             metric = "EC50 for combo data ",
                                             dataset_name = "Combo data"))



mae <- gDRutils::get_synthetic_data("combo_triple")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics",
                                               merge_additional_variables = TRUE)

output <- pheatmap_with_anno_combo_metrics(dt_metrics = dt_metrics,
                                           normalization_type = "RV",
                                           metric = "x_mean")
hm_4 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_4[["gtable"]])

}
\keyword{pheat_ann}
