% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{plolty_drug_combo_ratio}
\alias{plolty_drug_combo_ratio}
\title{Draw interactive Drug Combo plot with drug ratios}
\usage{
plolty_drug_combo_ratio(
  data,
  drug1_name,
  drug2_name,
  cell_line,
  c_assay = names(gDRutils::get_combo_excess_field_names())[[1]],
  normalization_type = "RV",
  iso_levels = c("0.5", "0.75"),
  height = 325,
  width = 425
)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{drug1_name}{string a name of the first drug to be selected}

\item{drug2_name}{string a name of the second drug to be selected}

\item{cell_line}{string a name of cell line to be selected}

\item{c_assay}{character string specifying the combo base assay to be selected,
any from \code{names(gDRutils::get_combo_excess_field_names())} will do}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{charvec isobologram levels to be selected}

\item{height}{integer; heatmap height}

\item{width}{integer; heatmap width}
}
\value{
plotly object - line chart with lines for different ration of drugs on isolevels
}
\description{
Draw interactive Drug Combo plot with drug ratios
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
combo_data_l <- 
  gDRutils::convert_combo_data_to_dt(mae[[gDRutils::get_supported_experiments("combo")]])

plolty_drug_combo_ratio(
  data = combo_data_l,
  drug1_name = "drug_005",
  drug2_name = "drug_021",
  cell_line = "cellline_HB",
  normalization_type = "GR",
  c_assay = "hsa_excess"
)

plolty_drug_combo_ratio(
  data = combo_data_l,
  drug1_name = "drug_005",
  drug2_name = "drug_021",
  cell_line = "cellline_HB",
  normalization_type = "RV",
  c_assay = "smooth"
)

plolty_drug_combo_ratio(
  data = combo_data_l,
  drug1_name = "drug_005",
  drug2_name = "drug_021",
  cell_line = "cellline_HB",
  normalization_type = "GR",
  c_assay = "bliss_excess",
  iso_levels = c("0.25", "0.5", "0.75")
)
}

}
\keyword{plugin_plot}
