% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_metric.R
\name{plot_boxplot_metric_sa}
\alias{plot_boxplot_metric_sa}
\title{Plot box plots for metric for single-agent data grouped by selected variable}
\usage{
plot_boxplot_metric_sa(
  dt_metrics,
  group_var,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  grouped_flag = FALSE,
  colored_pts_flag = FALSE,
  colors_vec = NULL,
  with_inf = FALSE
)
}
\arguments{
\item{dt_metrics}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{group_var}{string name of group variable; one of: \code{"CellLineName"} or \code{"DrugName"}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}),
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean"),
but the values from any numeric column can be displayed}

\item{fit_source}{string source name for metrics}

\item{grouped_flag}{a logical flag whether the boxplots should be grouped and
colored by \code{Tissue} for \code{group_var} set as \code{"CellLineName"}
and \code{drug_moa} - for \code{"DrugName"}}

\item{colored_pts_flag}{a logical flag whether the points should be colored by grouped variable -
for \code{group_var} equal \code{"CellLineName"} points will be colored by \code{"DrugName"}
and similarly vice versa}

\item{colors_vec}{character vector with colors (name or hex value) to color boxplots;
for \code{grouped_flag} set as \code{FALSE} only first from vector will be used}

\item{with_inf}{a logical flag indicating whether infinite values should be shown on boxplots}
}
\value{
\code{ggplot} object containing boxplots for selected single-agent metric
   grouped by \code{group_var}
}
\description{
Plot box plots for metric for single-agent data grouped by selected variable
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")

plot_boxplot_metric_sa(dt_metrics,
                       group_var = "CellLineName")

plot_boxplot_metric_sa(dt_metrics,
                       group_var = "DrugName")

plot_boxplot_metric_sa(dt_metrics,
                       group_var = "DrugName",
                       metric = "x_AOC_range",
                       colors_vec = "grey",
                       colored_pts_flag = TRUE)

plot_boxplot_metric_sa(dt_metrics,
                       group_var = "CellLineName",
                       normalization_type = "RV",
                       metric = "x_max",
                       colors_vec = c("gold", "darkorange", "darkcyan", "darkblue"),
                       grouped_flag = TRUE)

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{single-agent_plots}
