% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chemical_genomics.R
\name{plot_cgs_ranking}
\alias{plot_cgs_ranking}
\title{Plot Chemical Genomics Screen GSEA Results}
\usage{
plot_cgs_ranking(results, cell_line, metric)
}
\arguments{
\item{results}{A list object returned from `analyze_cgs`.}

\item{cell_line}{A string specifying the cell line included in the \code{results} to prepare a visualization.}

\item{metric}{A string specifying the metric included in the \code{results} to prepare a visualization.}
}
\value{
A ggplot2 object with cgs results
}
\description{
Generates a ggplot2 visualization of chemical genomics screening data, highlighting GSEA results.
}
\examples{
dt_metrics <- qs::qread(system.file("testdata/cgs_data.qs", package = "gDRplots"))
results <- analyze_cgs(dt_metrics, metrics = c("xc50"), cell_line = "CellLineName_1")
plot_cgs_ranking(results, cell_line = "CellLineName_1", metric = "xc50")
}
