% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{plot_combination_index}
\alias{plot_combination_index}
\title{Plot line plot of combination index}
\usage{
plot_combination_index(
  dt_excess,
  dt_isobolograms,
  drug1_name,
  drug2_name,
  cl_name,
  normalization_type = "GR",
  iso_levels = c("0.25", "0.5", "0.75"),
  colors_vec_iso = NULL
)
}
\arguments{
\item{dt_excess}{data.table representing data from the \code{excess} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "excess")}
and combo \code{SummarizedExperiment}}

\item{dt_isobolograms}{data.table representing data from the \code{isobolograms} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}
and combo \code{SummarizedExperiment}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_name}{string with cell line to be plotted (identifiers \code{CellLineName})}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{character vector with isobologram levels to be selected}

\item{colors_vec_iso}{character vector of colors (valid name or hex) used for the isolines; 
the default is the dark red-orange palette}
}
\value{
combination index plot at different ratios of the two drugs
}
\description{
Plot line plot of combination index
}
\examples{
cl_name <- "cellline_BC"
drug1_name <- "drug_001"
drug2_name <- "drug_026"

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_excess <- gDRutils::convert_se_assay_to_dt(se, "excess")
dt_isobolograms <- gDRutils::convert_se_assay_to_dt(se, "isobolograms")

plot_combination_index(dt_excess,
                       dt_isobolograms,
                       drug1_name, drug2_name,
                       cl_name,
                       normalization_type = "GR")
                       
plot_combination_index(dt_excess,
                       dt_isobolograms,
                       drug1_name, drug2_name,
                       cl_name,
                       normalization_type = "RV",                       
                       colors_vec_iso = c("darkblue", "darkcyan"))                       

cl_name <- "cellline_JE"
drug1_name <- "drug_011"
drug2_name <- "drug_026"

plot_combination_index(dt_excess,
                       dt_isobolograms,
                       drug1_name, drug2_name,
                       cl_name,
                       normalization_type = "RV",
                       iso_levels = "0.5")

}
\keyword{combo_plots}
