% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-response_curve.R
\name{plot_dose_response_combo}
\alias{plot_dose_response_combo}
\title{Plot drug response curves for combo data}
\usage{
plot_dose_response_combo(
  dt_average,
  drug1_name,
  drug2_name,
  cl_name,
  normalization_type = "GR",
  colormap = NULL,
  split_by_conc = FALSE
)
}
\arguments{
\item{dt_average}{data.table representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_name}{string with cell line name to be plotted (identifiers \code{CellLineName})}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colormap}{character vector with colors for \code{group_names} - name or hex value}

\item{split_by_conc}{logical flag whether curves for \code{Concentration_2} plotted on one plot
or separate}
}
\value{
plot with dose-response curves for combo data
}
\description{
Plot drug response curves for combo data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")

cl_name <- "cellline_BC"
drug1_name <- "drug_011"
drug2_name <- "drug_021"

plot_dose_response_combo(dt_average = dt_average,
                         drug1_name = drug1_name,
                         drug2_name = drug2_name,
                         cl_name = cl_name)
                         
plot_dose_response_combo(dt_average = dt_average,
                         drug1_name = drug1_name,
                         drug2_name = drug2_name,
                         cl_name = cl_name,
                         split_by_conc = TRUE)
                         
plot_dose_response_combo(dt_average = dt_average,
                         drug1_name = drug1_name,
                         drug2_name = drug2_name,
                         cl_name = cl_name,
                         normalization_type = "RV",
                         colormap = c("orange", "darkred"))                     
                      
}
\keyword{combo_plots}
