% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-response_curve.R
\name{plot_dose_response_combo_panel}
\alias{plot_dose_response_combo_panel}
\title{Plot panel with drug response curves for single-agent data to control quality of the data}
\usage{
plot_dose_response_combo_panel(
  dt_average,
  cl_name,
  d_names = NULL,
  normalization_type = "GR",
  colors_vec = NULL
)
}
\arguments{
\item{dt_average}{data.table representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and combo \code{SummarizedExperiment}}

\item{cl_name}{string with cell line name to be plotted (identifiers \code{CellLineName})}

\item{d_names}{character vector with drug names to be plotted (Drug Name);
if NULL - all available drugs will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colors_vec}{character vector with colors for \code{group_names} - name or hex value
note that the first color will be assigned to the min value of \code{Concentration_2}, 
and the last one - to the max of \code{Concentration_2}; the default is the orange-black palette}
}
\value{
\code{ggplot} object containing panel with plot with dose-response curves 
   for selected cell line by drugs
}
\description{
Plot panel with drug response curves for single-agent data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")

cl_name <- "cellline_IB"

plot_dose_response_combo_panel(dt_average = dt_average,
                               cl_name = cl_name)

d_names <- c("drug_001", "drug_002")
plot_dose_response_combo_panel(dt_average = dt_average,
                               cl_name = cl_name,
                               d_names = d_names)

}
\keyword{combo_plots}
