% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-response_curve.R
\name{plot_dose_response_combo_qc_panel}
\alias{plot_dose_response_combo_qc_panel}
\title{Plot panel with drug response curves for single-agent data to check quality of data}
\usage{
plot_dose_response_combo_qc_panel(
  dt_average,
  cl_name,
  d_names = NULL,
  metric_growth = "GR",
  as_panel = TRUE
)
}
\arguments{
\item{dt_average}{data.table representation of the data in \code{Averaged} assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}}

\item{cl_name}{string with cell line name to be plotted (identifiers \code{CellLineName})}

\item{d_names}{character vector with drug names to be plotted (Drug Name); 
if NULL - all available drugs will be plotted}

\item{metric_growth}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{as_panel}{logical flag whether return list of plot or panel}
}
\value{
panle with plot with dose-response curves for selected cell line by drugs
}
\description{
Plot panel with drug response curves for single-agent data to check quality of data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")

cl_name <- "cellline_IB"

plot_dose_response_combo_qc_panel(dt_average = dt_average,
                                  cl_name = cl_name,
                                  as_panel = FALSE)

d_names <- c("drug_001", "drug_002")
plot_dose_response_combo_qc_panel(dt_average = dt_average,
                                  cl_name = cl_name,
                                  d_names = d_names)

}
\keyword{QC_plot}
