% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sa-response_curve.R
\name{plot_dose_response_sa_by_CLs}
\alias{plot_dose_response_sa_by_CLs}
\title{Plot drug response curves for single-agent data for selected call lines and drugs}
\usage{
plot_dose_response_sa_by_CLs(
  dt_metrics,
  dt_average,
  cellline_name_vec = NULL,
  drug_name_vec = NULL,
  normalization_type = "GR",
  colormap = NULL,
  plot_averaged_flag = TRUE,
  plot_fit_flag = TRUE
)
}
\arguments{
\item{dt_metrics}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}}

\item{dt_average}{data.table representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}}

\item{cellline_name_vec}{character vector with cell line to be plotted (Cell Line Name)}

\item{drug_name_vec}{character vector with cell line to be plotted (Drug Name)}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colormap}{character vector with colors for \code{group_names} - name or hex value}

\item{plot_averaged_flag}{logical flag whether plot points with average values}

\item{plot_fit_flag}{logical flag whether plot points with fitted values}
}
\value{
list of plots with dose-response curves
}
\description{
Plot drug response curves for single-agent data for selected call lines and drugs
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
cellline_name_vec <- unique(dt_metrics[["CellLineName"]])[2:5]
drug_name_vec <- unique(dt_metrics[["DrugName"]])[5:7]

plot_dose_response_sa_by_CLs(dt_metrics = dt_metrics, 
                             dt_average = dt_average, 
                             cellline_name_vec = cellline_name_vec, 
                             drug_name_vec = drug_name_vec, 
                             normalization_type = "RV", 
                             colormap = c("#B9D3EE", "#FF6347", "#C2F970"))

}
\keyword{single-agent_plots}
