% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sa-response_curve.R
\name{plot_dose_response_sa_qc}
\alias{plot_dose_response_sa_qc}
\title{Plot drug response curves for single-agent data to control quality of the data}
\usage{
plot_dose_response_sa_qc(
  dt_metrics,
  dt_average,
  cl_name,
  d_name,
  normalization_type = "GR",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_metrics}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{dt_average}{data.table representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and single-agent \code{SummarizedExperiment}}

\item{cl_name}{string cell line name to be plotted (Cell Line Name)}

\item{d_name}{string vector with drug name to be plotted (Drug Name)}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{fit_source}{string source name for metrics}
}
\value{
plot with dose-response curves
}
\description{
Plot drug response curves for single-agent data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
cl_name <- dt_metrics[["CellLineName"]][1]
d_name <- dt_metrics[["DrugName"]][1]

plot_dose_response_sa_qc(dt_metrics = dt_metrics,
                         dt_average = dt_average,
                         cl_name = cl_name,
                         d_name = d_name)

}
\keyword{QC_plot}
