% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-normalization_qc.R
\name{plot_fitting_acc}
\alias{plot_fitting_acc}
\title{Visualization for the quality control of the fitting for single-agent data}
\usage{
plot_fitting_acc(dt_assay, cl_name, normalization_type = "GR")
}
\arguments{
\item{dt_assay}{data.table representation of the data in assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}}

\item{cl_name}{string cell line name to be plotted (Cell Line Name)}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
panel with lollipop plots with r2 and rss values for each drug
}
\description{
Visualization for the quality control of the fitting for single-agent data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
cl_name <- dt_metrics[["CellLineName"]][1]

plot_fitting_acc(dt_assay = dt_metrics,
                 cl_name = cl_name,
                 normalization_type = "RV")
}
\keyword{QC_plot}
