% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{plot_plate_data}
\alias{plot_plate_data}
\title{Plot plate data}
\usage{
plot_plate_data(data)
}
\arguments{
\item{data}{The data table containing plate information}
}
\value{
A list of ggplot objects for each barcode
}
\description{
Plot plate data
}
