% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{plot_plate_single_data}
\alias{plot_plate_single_data}
\title{Plot data from a specific column}
\usage{
plot_plate_single_data(data, column_name)
}
\arguments{
\item{data}{The data table containing plate information}

\item{column_name}{The name of the column to plot}
}
\value{
A list of ggplot objects for each barcode and column
}
\description{
Plot data from a specific column
}
\examples{
test_data <- data.table::data.table(
  WellColumn = rep(1:12, each = 8),
  WellRow = rep(LETTERS[1:8], times = 12),
  Gnumber = c(rep("untreated", 48), sample(1:5, size = 48, replace = TRUE)),
  Gnumber_2 = c(rep("untreated", 48), sample(1:5, size = 48, replace = TRUE)),
  Concentration = runif(96, min = 0, max = 100),
  ReadoutValue = runif(96, min = 0, max = 100),
  Barcode = rep(c("A", "B"), 48)
  )
plot_plate_single_data(test_data, "Gnumber")[[1]]
}
\keyword{QC_plot}
