% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_plots.R
\name{plot_SA_1CL}
\alias{plot_SA_1CL}
\title{Plot drug response curves for single-agent data for one selected cell line}
\usage{
plot_SA_1CL(
  SE,
  normalization_type = "GR",
  colormap = NULL,
  plot_averaged_flag = TRUE,
  plot_fit_flag = TRUE
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object holding raw and/or processed dose-response 
data in its assays for one cell line}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colormap}{character vector with colors for \code{group_names} - name or hex value}

\item{plot_averaged_flag}{logical flag whether plot points with average values}

\item{plot_fit_flag}{logical flag whether plot points with fitted values}
}
\value{
plot with dose-response curves
}
\description{
Plot drug response curves for single-agent data for one selected cell line
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
SE <- mae[[1]]

plot_SA_1CL(SE = SE[,colnames(SE)[1]], colormap = c("cadetblue", "orange", "darkblue"))

}
\keyword{internal}
