% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_plots.R
\name{plot_SA_byCLs}
\alias{plot_SA_byCLs}
\title{Plot drug response curves for single-agent data for selected call lines and drugs}
\usage{
plot_SA_byCLs(
  SE,
  cellline_name = NULL,
  drug_name = NULL,
  normalization_type = "GR",
  colormap = NULL,
  plot_averaged_flag = TRUE,
  plot_fit_flag = TRUE
)
}
\arguments{
\item{SE}{single-agent \code{SummarizedExperiment} object holding raw and/or processed 
dose-response data in its assays for one cell line}

\item{cellline_name}{character vector with cell line to be plotted (colnames of SE)}

\item{drug_name}{character vector with cell line to be plotted (rownames of SE)}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colormap}{character vector with colors for \code{group_names} - name or hex value}

\item{plot_averaged_flag}{logical flag whether plot points with average values}

\item{plot_fit_flag}{logical flag whether plot points with fitted values}
}
\value{
list of plots with dose-response curves
}
\description{
Plot drug response curves for single-agent data for selected call lines and drugs
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("small")
SE <- mae[[1]]
cellline_name <- colnames(SE)[2:5]
drug_name <- rownames(SE)[5:7]

plot_SA_byCLs(SE = SE, 
              cellline_name = cellline_name, 
              drug_name = drug_name, 
              normalization_type = "RV", 
              colormap = c("#B9D3EE", "#FF6347", "#C2F970"))
}

}
\keyword{internal}
