% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_scatter_with_corr}
\alias{plot_scatter_with_corr}
\title{Plot scatter with correlation}
\usage{
plot_scatter_with_corr(
  dt_response,
  dt_depmap,
  selected_feat,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features).
outputted by \code{\link{prep_dt_depmap_feat}}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap
(will be used as a plot title)}
}
\value{
\code{ggplot} object containing a scatter plot with correlation
}
\description{
Plot scatter with correlation
}
\examples{
\dontshow{if (nzchar(system.file("depmap_data", package = "gDRtestData"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mae_prism <- gDRutils::get_synthetic_data("prism")
se_prism <- mae_prism[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se_prism, "Metrics")
dt_response <- prep_dt_response_metric_sa(
  dt_metrics = dt_metrics,
  d_name = unique(dt_metrics[["DrugName"]])[1],
  normalization_type = "RV", metric = "x_mean")
feat_data_path <- system.file("depmap_data", package = "gDRtestData")
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
obj_depmap_feat <- prep_dt_depmap_feat(feat_data_path, meta_data_path,
                                       feature_set = "OmicsCNGene")
plot_scatter_with_corr(
  dt_response = dt_response,
  dt_depmap = obj_depmap_feat[["dt_depmap"]],
  selected_feat = names(obj_depmap_feat[["dt_depmap"]])[3])
\dontshow{\}) # examplesIf}
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{prism_plots}
