% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-normalization_qc.R
\name{plot_var_distribution_qc}
\alias{plot_var_distribution_qc}
\title{Plot violin for normalized single-agent data to check quality of data}
\usage{
plot_var_distribution_qc(dt_assay, cl_name, metric = "x", metric_growth = "GR")
}
\arguments{
\item{dt_assay}{data.table representation of the data in assay
output from \code{gDRutils::convert_se_assay_to_dt(se, assya_name)}}

\item{cl_name}{string cell line name to be plotted (Cell Line Name)}

\item{metric_growth}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
plot with violin for each drug
}
\description{
Plot violin for normalized single-agent data to check quality of data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_norm <- gDRutils::convert_se_assay_to_dt(se, "Normalized")
cl_name <- dt_norm[["CellLineName"]][1]

plot_var_distribution_qc(dt_assay = dt_norm,
                         cl_name = cl_name)

dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
plot_var_distribution_qc(dt_assay = dt_average,
                         cl_name = cl_name,
                         metric_growth = "RV")

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
plot_var_distribution_qc(dt_assay = dt_metrics,
                         cl_name = cl_name,
                         metric = "r2")
 
}
\keyword{QC_plot}
