% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_volcano_assoc}
\alias{plot_volcano_assoc}
\title{Volcano plot with association}
\usage{
plot_volcano_assoc(
  dt_assoc,
  selected_feat_meta_col,
  selected_metric,
  condition_info = NULL,
  alpha = 0.05,
  named_p_top = 10,
  max_N = NULL
)
}
\arguments{
\item{dt_assoc}{\code{data.table} with the calculated linear association between DepMap and metrics
outputted by \code{calc_assoc}}

\item{selected_feat_meta_col}{string describing the name of the associated feature/metadata from DepMap}

\item{selected_metric}{string describing the name of the selected metric used the association calculation}

\item{condition_info}{string describing experiment condition 
(preferred: \code{"DrugName"}_\code{"Gnumber"}_\code{"drug_moa"}_\code{"Duration"})}

\item{alpha}{numeric cutoff to identify statistically significant correlations}

\item{named_p_top}{numeric value for p-top statistically significant correlations 
to be labeled on the plot}

\item{max_N}{numeric value for limit the maximum number of non-statistically 
significant points to plot; for default \code{NULL} all points will be plotted.}
}
\value{
\code{ggplot} object containing a volcano plot with association
}
\description{
Volcano plot with association
}
\examples{
Y <- matrix(seq(0.5, 2, length.out = 50), nrow = 50,
            dimnames = list(sprintf("row_\%s", 1:50), "met_1"))
X <- matrix(
  withr::with_seed(42, sample(c(NA, seq(0.35, 23.5, 1.25)), 50*20, replace = TRUE)), nrow = 50,
  dimnames = list(sprintf("row_\%s", 1:50), sprintf("feat_\%s", 1:20)))
tab_assoc <- calc_assoc(X, Y)
plot_volcano_assoc(tab_assoc,
                   selected_feat_meta_col = "feat_XY",
                   selected_metric = "met_RV")
                   
}
\keyword{prism_plots}
