% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_volcano_corr_panel}
\alias{plot_volcano_corr_panel}
\title{Plot panel with volcano plot and scatter plots (top 4) features from DepMap}
\usage{
plot_volcano_corr_panel(dt_response, dt_depmap, selected_metric, selected_feat)
}
\arguments{
\item{dt_response}{\code{data.table} with the experimental response data (rows are samples) 
for one metric outputted by one of functions: \code{\link[gDRplots]{prep_dt_response_metric_sa}},
\code{\link[gDRplots]{prep_dt_response_dose_sa}}, \code{\link[gDRplots]{prep_dt_response_scores}}
or \code{\link[gDRplots]{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap.
(rows are samples, columns are meta);  
outputted by \code{\link[gDRplots]{prep_dt_depmap_feat}}}

\item{selected_metric}{string name of the metric in \code{dt_response}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}
}
\value{
a panel with volcano plot and scatter plots with correlation for feature set
}
\description{
Plot panel with volcano plot and scatter plots (top 4) features from DepMap
}
\keyword{prism_plots}
