% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{plotly_metric_clustering}
\alias{plotly_metric_clustering}
\title{Draw interactive Metric Clustering heatmap}
\usage{
plotly_metric_clustering(
  data,
  variable,
  annotations,
  transpose = FALSE,
  colors,
  dendrogram = "both"
)
}
\arguments{
\item{data}{a numeric matrix with dimnames}

\item{variable}{character string, naming metric to analyze}

\item{annotations}{list of of length 2,
containing row and column annotations as \code{data.table}s}

\item{transpose}{logical; whether or not to transpose the matrix}

\item{colors}{named list with colors for heatmap itself, rows and columns clusterings;
.   names: \code{heatmap}, \code{row} and \code{col}}

\item{dendrogram}{character string indicating whether to compute 'none',
'row', 'column' or 'both' dendrograms. Defaults to 'both'.}
}
\value{
An interactive plot created by \code{iheatmapr}.
}
\description{
Plot metric heat map with clustering
}
\details{
Generate a heat map of a selected metric, clustering cell lines and drugs.
Distances are computed with \code{\link[gDRplots]{computeDistances}}.
}
\seealso{
\code{prepare_data_metric_clustering} \code{MetricClustering}, \code{\link[gDRplots]{computeDistances}}
}
\keyword{plugin_plot}
