% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{plotlyDC1}
\alias{plotlyDC1}
\title{Draw interactive Drug Combo details heatmap with isobolograms}
\usage{
plotlyDC1(
  data,
  drug1_name,
  drug2_name,
  cell_line,
  c_assay = names(gDRutils::get_combo_excess_field_names())[1],
  normalization_type = "RV",
  iso_levels = c("0.5", "0.75"),
  with_iso_points = FALSE,
  height = 325L,
  width = 425L
)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{drug1_name}{string a name of the first drug to be selected}

\item{drug2_name}{string a name of the second drug to be selected}

\item{cell_line}{string a name of cell line to be selected}

\item{c_assay}{character string specifying the combo base assay to be selected,
any from \code{names(gDRutils::get_combo_excess_field_names())} will do}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{charvec isobologram levels to be selected}

\item{with_iso_points}{logical flag whether draw the point for IC50/GR50 isobologram}

\item{height}{integer; heatmap height}

\item{width}{integer; heatmap width}
}
\value{
plotly object - heatmap for 2-way concentration arrays with isobolograms
}
\description{
Functions for detailed visualization of combo drug treatments, i.e. 2-way concentration arrays.
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
combo_data_l <- gDRutils::convert_combo_data_to_dt(mae[[gDRutils::get_supported_experiments("combo")]])

plotlyDC1(
 data = combo_data_l,
 drug1_name = "drug_005",
 drug2_name = "drug_021",
 cell_line = "cellline_HB",
 normalization_type = "GR",
 c_assay = "hsa_excess"
)

plotlyDC1(
  data = combo_data_l,
  drug1_name = "drug_005",
  drug2_name = "drug_021",
  cell_line = "cellline_HB",
  normalization_type = "RV",
  c_assay = "smooth"
)

plotlyDC1(
  data = combo_data_l,
  drug1_name = "drug_005",
  drug2_name = "drug_021",
  cell_line = "cellline_HB",
  normalization_type = "GR",
  c_assay = "bliss_excess",
  iso_levels = c("0.25", "0.5", "0.75")
)
}

}
\keyword{plugin_plot}
