% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-drugCombo.R
\name{plotlyDCMother}
\alias{plotlyDCMother}
\title{Draw interactive Drug Combo heatmap}
\usage{
plotlyDCMother(data, metric_combo, metric_growth)
}
\arguments{
\item{data}{a list of data.table(s) with combo data}

\item{metric_combo}{character string specifying which combo score metric is used;
one of \code{names(gDRutils::get_combo_score_field_names())}}

\item{metric_growth}{character string specifying which growth metric is used;
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
plotly object - heatmap
}
\description{
Build experiment overview heat map.
}
\details{
Produces a clustered heat map of cell line vs. drug pair.
The app registers clicks on panes and opens a pop-up with
detailed exploration of the drug interaction.
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
combo_data_l <- gDRutils::convert_combo_data_to_dt(mae[[gDRutils::get_supported_experiments("combo")]])
plotlyDCMother(data = combo_data_l, 
               metric_combo = "hsa_score", 
               metric_growth = "RV")

plotlyDCMother(data = combo_data_l, 
               metric_combo = "CIScore_50", 
               metric_growth = "GR")
}

}
\keyword{plugin_plot}
