% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricContrast.R
\name{plotlyMC}
\alias{plotlyMC}
\title{Draw interactive Metric Contrast plot}
\usage{
plotlyMC(
  data,
  var_x,
  var_y,
  var_col,
  var_txt,
  metric = "growth metric",
  identity = FALSE,
  correlation = FALSE,
  axis_limits = c("fixed", "upper", "lower", "free"),
  source = "metricContrast",
  with_labelR = FALSE
)
}
\arguments{
\item{data}{a data table containing metrics for various conditions
(cell lines treated with drugs)}

\item{var_x, var_y}{character strings assigning variables to axes}

\item{var_col}{character string assigning variable to coloring}

\item{var_txt}{character string assigning variable to tooltip}

\item{metric}{character string representing name of growth metric that is plotted}

\item{identity}{logical flag specifying whether to add an identity line}

\item{correlation}{logical flag specifying whether to print statistics
of a Spearman correlation test; a linear fit is also added for
the metrics that are plotted on linear axes}

\item{axis_limits}{character string; fix or release limits of Y axis, see \code{Details}}

\item{source}{character string representing a match of this string's value 
with the source argument in \code{event_data()}}

\item{with_labelR}{logical flag whether to enabale \code{plotlyLabelR} functionality}
}
\value{
a plotly object
}
\description{
Build a an intereactive plot for comparing a metric between two groups.
}
\details{
Draws a scatter plot, where one metric is compared between different conditions.
Depending on the variables assigned to the axes (one drug or one cell line each),
the points correspond to the values of the selected metric for those cell lines / drugs.
If cell lines are put on axes, drugs will be color coded and vice versa.

Special rules apply for GR50, IC50 and EC50.
These variables are plotted in log-scale.
The argument \code{axis_limits} determines whether the Y axis limits will be
fixed at 1e-4 to 100 (fixed), derived from data (free), or whether one
limit will be fixed and the other derived from the data (upper and lower).

Note that in the app axes are editable and the user can alter the range by dragging the axis.

On request from shiny app an identity line (\code{y = x}) will be drawn.
Likewise, a Spearman correlation test can be run on the data and its results can be
printed on the plot. Unless \code{metric} is one of the *50s, a linear fit will also be plotted.

The plot is generated with \code{\link[plotly]{plot_ly}} and has some interactivity.
}
\seealso{
\code{prepareDataMC} \code{MetricContrast}
}
\keyword{plugin_plot}
