% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricDistribution.R
\name{plotlyMD}
\alias{plotlyMD}
\title{Draw interactive Metric Distribution plot}
\usage{
plotlyMD(
  data,
  var_x,
  var_y,
  var_grp,
  var_col,
  type,
  points = "off",
  axis_limits = c("fixed", "upper", "lower", "free"),
  show_tick_labels
)
}
\arguments{
\item{data}{a data table containing metrics for various conditions
(cell lines treated with drugs)}

\item{var_x, var_y, var_col, var_grp}{variables of \code{data} to assign to plot elements}

\item{type}{character string defining type of plot: box or violin}

\item{points}{character string specifying whether to add points on top of boxes/violins}

\item{axis_limits}{character string; fix or release limits of Y axis, see \code{Details}}

\item{show_tick_labels}{boolean; whether or not to show X axis labels}
}
\value{
a plotly object
}
\description{
Build an interactive plot of metrics distribution.
}
\details{
Draws a box- or violin plot, depending on user input, with optionally added points.
Outliers are omitted from the boxplot.
The color scheme, determined by \code{var_col}, applies to both box/violin and points.

Special rules apply for GR50, IC50 and EC50.
These variables are plotted in log-scale.
The argument \code{axis_limits} determines whether the Y axis limits will be
fixed at 1e-4 to 100 (fixed), derived from data (free), or whether one
limit will be fixed and the other derived from the data (upper and lower).

Note that in the app axes are editable and the user can alter the range by dragging the axis.
}
\seealso{
\code{MetricDistribution}
}
\keyword{plugin_plot}
