% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{plotlyMH}
\alias{plotlyMH}
\title{Draw interactive Metric Clustering plot}
\usage{
plotlyMH(
  data,
  variable,
  annotations,
  transpose = FALSE,
  colors,
  dendrogram = "both"
)
}
\arguments{
\item{data}{a numeric matrix with dimnames}

\item{variable}{character string, naming metric to analyze}

\item{annotations}{list of of length 2,
containing row and column annotations as \code{data.table}s}

\item{transpose}{logical; whether or not to transpose the matrix}

\item{dendrogram}{character string indicating whether to compute 'none',
'row', 'column' or 'both' dendrograms. Defaults to 'both'.}
}
\value{
An interactive plot created by \code{iheatmapr}.
}
\description{
Plot metric heat map with clustering
}
\details{
Generate a heat map of a selected metric, clustering cell lines and drugs.
Distances are computed with \code{\link[gDRplots]{computeDistances}}.
}
\seealso{
\code{prepareDataMH} \code{MetricClustering}, \code{\link[gDRplots]{computeDistances}}
}
\keyword{plugin_plot}
