% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricRanking.R
\name{plotlyMR}
\alias{plotlyMR}
\title{Draw interactive Metric Ranking plot}
\usage{
plotlyMR(
  data,
  var_x,
  var_y,
  var_col,
  var_grp,
  title,
  show_tick_labels,
  axis_limits = c("fixed", "upper", "lower", "free")
)
}
\arguments{
\item{data}{\code{data.table} prepared by \code{prepareDataMR}}

\item{var_x}{character string; name of variable to put on X axis}

\item{var_y}{character string; name of variable to put on Y axis}

\item{var_col}{character string; name of variable to color bars by}

\item{var_grp}{character string; name of variable to group values by}

\item{title}{name of variable to construct plot title from}

\item{show_tick_labels}{boolean; whether or not to show X axis labels}

\item{axis_limits}{character string; fix or release limits of Y axis, see \code{Details}}
}
\value{
a \code{plotly} object
}
\description{
Create a barplot that ranks drugs or cell lines across a growth metric.
}
\details{
Special rules apply for GR50, IC50 and EC50.
These variables are plotted in log-scale.
The argument \code{axis_limits} determines whether the Y axis limits will be
fixed at 1e-4 to 100 (fixed), derived from data (free), or whether one
limit will be fixed and the other derived from the data (upper and lower).

Note that in the app axes are editable and the user can alter the range by dragging the axis.
}
\keyword{plugin_plot}
