% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{plotlyRCAll}
\alias{plotlyRCAll}
\title{plot reactive response curves}
\usage{
plotlyRCAll(
  curve_data,
  var_y,
  range_x = c(0.001, 50),
  plot_width = 400L,
  plot_height = 300L
)
}
\arguments{
\item{curve_data}{a data.table containing coordinates for plotting predicted dose response curves}

\item{var_y}{which variable to plot: GR value or Relative Viability}

\item{range_x}{numeric vector of length 2, specifying range of X axis}

\item{plot_width, plot_height}{numeric; dimensions of plot}
}
\value{
A plotly object with highlights. The plot contains only response curves.

The app will tracks clicks and create a secondary plot with a more detailed view of the selected curves.
}
\description{
Plot all curves in a given data set on a reactive plot.
}
\examples{
SE <- gDRutils::get_synthetic_data("small")[[1]]
dt <- gDRutils::convert_se_assay_to_custom_dt(SE, "Metrics")
prepared_curves <- prepareCurves(dt)
plotlyRCAll(prepared_curves, "GR value")

}
\keyword{drawResponseOverview}
