% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{plotlyRCSelected}
\alias{plotlyRCSelected}
\title{plot drug dose response}
\usage{
plotlyRCSelected(
  data,
  var_y,
  layers,
  curves,
  range_x = c(0.001, 50),
  extras,
  plot_width = 400L,
  plot_height = 300L
)
}
\arguments{
\item{data}{a data.table with cell fitness data taken from a \code{SummarizedExperiment}}

\item{var_y}{variable to plot on Y axis}

\item{layers}{character vector of layers to plot, see \code{Details}}

\item{curves}{a data.table with drug response predicted from growth metrics}

\item{range_x}{numeric vector of length 2, specifying range of X axis}

\item{extras}{a list of extra features that comprise the fifth layer, see \code{Details}}

\item{plot_width, plot_height}{numeric; dimensions of plot}
}
\value{
a plotly object
}
\description{
Make an interactive plot of cell fitness against a range of drug concentrations.
}
\details{
Any combination of four layers can be drawn, depending on \code{layers}:
all measurement values with "observations",
average GR value for each drug concentration with "average",
error bars (standard deviation) with "error",
and the fitted response curve with "curve" (the default).

The fifth layer, "extras", highlights points corresponding to GR50/IC50 and GR max/E max.
This quasi-layer is composed of points, which are a trace, lines, which are shapes,
and labels, which are also a trace. The labels serve as axis ticks for the vertical and horizontal lines.
The information necessary to draw the layer is supplied by the \code{extras} argument,
see \code{\link{prepareExtras}}.
It is a list of length 2,
the first item being a \code{data.table} with coordinates of points
passed to \code{plotly::add_markers},
and the second being a list of line descriptions
passed the \code{shapes} argument of \code{plotly::layout}.

Curve data is supplied in a separate data frame.
}
\section{Recursive plotting}{

If the data set contains multiple co-drugs, this function is called recursively
so that data for each co-drug is plotted on a separate panel. Panel are arranged vertically.
}

\examples{
se <- gDRutils::get_synthetic_data("small")[[1]][1:3, 4:6]
dt <- gDRutils::convert_se_assay_to_custom_dt(se, "Metrics")
prepared_curves <- prepareCurves(dt)
plotlyRCSelected(
  data = prepared_curves,
  var_y = "GR value",
  layers = c("curve", "average", "error"),
  curves = prepared_curves,
  extras = prepareExtras(dt)
)


}
\seealso{
\code{replaceValues}
}
\keyword{drawResponseOverview}
\keyword{internal}
