% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-responseGrid.R
\name{plotlyRGPanel}
\alias{plotlyRGPanel}
\title{Draw interactive Response Grid panel}
\usage{
plotlyRGPanel(data, var_y, range_x = c(0.001, 50), title = NA)
}
\arguments{
\item{data}{data frame of curve coordinates}

\item{var_y}{name of metric to plot: GR value or Relative Viability, given as character string}

\item{range_x}{numeric vector of length 2 that specifies the limits of the X axis}

\item{title}{character string defining the column, whose content will form the panel's title;
if set to NA (the default), no title is displayed}
}
\value{
A \code{plotly} object.
}
\description{
Build a panel for the drug response grid.
}
\seealso{
\code{\link[gDRplots]{prepareCurves}}
}
\keyword{plugin_plot}
