% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_assoc_summary}
\alias{prep_assoc_summary}
\title{Prepare summary table with statistically significant associations}
\usage{
prep_assoc_summary(
  dir_path,
  ls_file,
  alpha = 0.05,
  n_stat_sig_row = 10,
  read_file_fun = readxl::read_excel,
  as_list = FALSE
)
}
\arguments{
\item{dir_path}{A string path to the directory containing files with associations data.}

\item{ls_file}{A character vector with names of files containing associations data.}

\item{alpha}{A numeric cutoff to identify statistically significant correlations}

\item{n_stat_sig_row}{A numeric value specifying the maximum number of statistically
significant associations (rows) to include from each file.}

\item{read_file_fun}{A function to read the data from file; default is \code{readxl::read_excel}}

\item{as_list}{A logical flag indicating whether the result should be returned
as a list or as a table.}
}
\value{
A \code{DT::datatable} object.
}
\description{
Prepare summary table with statistically significant associations
}
\examples{
\donttest{
prep_assoc_summary(
  dir_path = "path/to/assoc/results",
  ls_file = c("chunk__feat_DrugA_RV_gDR_log10_xc50.xlsx"),
  alpha = 0.05
)
}

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
