% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_double_table_chunk}
\alias{prep_double_table_chunk}
\title{Prepare markdown chunk based on a doubly nested list of tables}
\usage{
prep_double_table_chunk(
  tbl_list,
  chunk_name,
  dwn_list = NULL,
  header_level = 3,
  tabset_options = c("tabset", "tabset-dropdown"),
  sorting_opts = NULL
)
}
\arguments{
\item{tbl_list}{A doubly nested named list of tables. The outer list represents cell lines,
and the inner lists represent metrics. Names are used as headings.}

\item{chunk_name}{A character string specifying the base name for the generated code chunks. Avoid spaces.}

\item{dwn_list}{A named list of links to location (relative paths) where table or plots are saved,
which when clocked, will be downloaded. It must have the same structure as \code{plt_list}.}

\item{header_level}{An integer specifying the markdown header level to use (e.g., 1 for `#`, 2 for `##`, etc.).}

\item{tabset_options}{A character vector of options for the tabset. This is only used
when \code{plt_list} is a nested list.
Possible values are "unnumbered", "tabset", and "tabset-dropdown".}

\item{sorting_opts}{A vector specifying global sorting options for all tables.
Column names can be preceded by "-" to indicate descending order.}
}
\value{
A list of character vectors. Each element corresponds to a cell line. Each character vector
  represents markdown code for the cell line's tabset.
}
\description{
Generates markdown code for displaying tables in a document using `knitr::knit()`.
Handles doubly nested lists, allowing for tabbed sections for cell lines and then metrics.
The inner header level (for metrics) is automatically set to one level greater than the outer header level.
}
\examples{
nested_tables <- list(
  CellLine1 = list(MetricA = mtcars[1:5, ], MetricB = mtcars[6:10, ]),
  CellLine2 = list(MetricC = iris[1:5, ], MetricD = iris[6:10, ])
)
sorting_options <- c("cyl", "-hp")
prep_double_table_chunk(nested_tables, "nested_tables", header_level = 2,
  tabset_options = "tabset", sorting_opts = sorting_options)

}
\author{
Bartosz Czech \email{czech.bartosz@external.gene.com}
}
\keyword{internal}
