% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_assoc}
\alias{prep_dt_assoc}
\title{Prep table with calculated linear associations}
\usage{
prep_dt_assoc(dt_response, dt_depmap, selected_feat_meta_col = NULL)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric}

\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap.
(rows are samples, columns are features or meta);
outputted by one of \code{\link{prep_dt_depmap_feat}} or
\code{\link{prep_dt_depmap_meta}}}

\item{selected_feat_meta_col}{string name of feature/meta column in DepMap}
}
\value{
A named list with elements, that may be input to \code{\link{plot_volcano_assoc}}
\itemize{
  \item \code{dt_assoc} \code{data.table} with calculated association values between
     feature/meta of DepMap and selected metric,
  \item \code{condition_info} string describing experiment condition (drugs),
  \item \code{selected_feat_meta_col} string name of feature/meta.
}
}
\description{
Prep table with calculated linear associations
}
\examples{
\dontshow{if (nzchar(system.file("depmap_data", package = "gDRtestData"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mae_prism <- gDRutils::get_synthetic_data("prism")
se_prism <- mae_prism[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se_prism, "Metrics")
dt_response <- prep_dt_response_metric_sa(
  dt_metrics = dt_metrics,
  d_name = unique(dt_metrics[["DrugName"]])[1],
  normalization_type = "RV", metric = "x_mean")
feat_data_path <- system.file("depmap_data", package = "gDRtestData")
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
obj_depmap_feat <- prep_dt_depmap_feat(feat_data_path, meta_data_path,
                                       feature_set = "OmicsCNGene")
prep_dt_assoc(
  dt_response = dt_response,
  dt_depmap = obj_depmap_feat[["dt_depmap"]],
  selected_feat_meta_col = obj_depmap_feat[["selected_feat_meta_col"]])
\dontshow{\}) # examplesIf}
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{prism_plots}
