% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_response_dose_sa}
\alias{prep_dt_response_dose_sa}
\title{Prep table with metric values by doses for single-agent experiment}
\usage{
prep_dt_response_dose_sa(
  dt_average,
  d_name,
  normalization_type = "RV",
  metric = "x",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_average}{\code{data.table} representing data from the \code{Averaged} assay,
outputted by \code{\link[gDRutils:convert_se_assay_to_dt]{gDRutils::convert_se_assay_to_dt}}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{d_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "x" (value of "GR" or "RV" itself - respectively depending on \code{normalization_type}),
or "x_std" (standard deviation)}

\item{fit_source}{string source name for metrics}
}
\value{
\code{data.table} with selected metric, input to \code{\link{prep_dt_assoc}}
}
\description{
Prep table with metric values by doses for single-agent experiment
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Averaged")
d_name <- "drug_004"
dt_response <- prep_dt_response_dose_sa(dt_average, d_name)

}
\keyword{prism_plots}
