% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_response_metric_diff}
\alias{prep_dt_response_metric_diff}
\title{Prep table with metric values for combination experiment}
\usage{
prep_dt_response_metric_diff(
  dt_metrics,
  d_name,
  d_name2,
  resistant_cl = NULL,
  sensitive_cl = NULL,
  normalization_type = "RV",
  metric = "xc50",
  fit_source = "gDR",
  additional_cols = NULL
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and combo \code{SummarizedExperiment}}

\item{d_name}{string representing the drug name to be plotted (identifier \code{DrugName}).
If set to NULL, the function will return a table for all available DrugName}

\item{d_name2}{string representing the drug name to be plotted (identifier \code{DrugName_2}).
If set to NULL, the function will return a table for all available DrugName_2}

\item{resistant_cl}{string representing the resistant cell line name.
If set to NULL, the function will return a table for all available cell lines.}

\item{sensitive_cl}{string representing the sensitive cell line name.
If set to NULL, the function will return a table for all available cell lines.}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max"), "x_mean" ("GR Mean" or "RV Mean") "ec50" ("GEC50" or "EC50") 
or "x_AOC_range" ("GR AOC within set range" or "RV AOC within set range")}

\item{fit_source}{string source name for metrics}

\item{additional_cols}{character vector with additional cols that should be included in the output}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{prep_dt_assoc}}
}
\description{
Prep table with metric values for combination experiment
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")
d_name <- "drug_004"
d_name2 <- "drug_026"
dt_response <- 
  prep_dt_response_metric_diff(dt_metrics, d_name, d_name2,
  metric = c("xc50", "x_mean", "x_max"))

resistant_cl <- "cellline_GB"
sensitive_cl <- "cellline_HB"

dt_response <- 
  prep_dt_response_metric_diff(dt_metrics, d_name = NULL, d_name2 = NULL,
  resistant_cl, sensitive_cl,
  metric = c("xc50", "x_mean", "x_max"))

}
\keyword{prism_plots}
