% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_response_sa}
\alias{prep_dt_response_sa}
\title{Prep table with metric values for single-agent experiment}
\usage{
prep_dt_response_sa(
  dt_average,
  dt_metrics,
  d_name,
  normalization_type = "RV",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_average}{\code{data.table} representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{d_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{fit_source}{string source name for metrics}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{prep_dt_assoc}}
}
\description{
Prep table with metric values for single-agent experiment
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_average <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Averaged")
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")
d_name <- "drug_004"
dt_response <- prep_dt_response_sa(dt_average, dt_metrics, d_name)

}
\keyword{prism_plots}
