% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{prep_pheatmap_matrix}
\alias{prep_pheatmap_matrix}
\title{Prep matrix with metric value based on the Metrics assay}
\usage{
prep_pheatmap_matrix(
  dt_response,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  experiment_type = gDRutils::get_supported_experiments("sa")
)
}
\arguments{
\item{dt_response}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{\link[gDRutils:convert_se_assay_to_dt]{gDRutils::convert_se_assay_to_dt}}
and single-agent \code{SummarizedExperiment} or
\code{data.table} representing data from the \code{scores} assay,
outputted by \code{\link[gDRutils:convert_se_assay_to_dt]{gDRutils::convert_se_assay_to_dt}}
and combo \code{SummarizedExperiment}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}),
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean");
but the values from any numeric column can be displayed.}

\item{fit_source}{string source name for metrics}

\item{experiment_type}{string with experiment name;
one of: "single-agent", "combination" or "co-dilution"}
}
\value{
matrix with values for selected metric with \code{CellLinName} in the rows
   and  \code{DrugName} (or combination of \code{DrugName} and \code{DrugName_2}) in the columns
}
\description{
Prep matrix with metric value based on the Metrics assay
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")

mat <- prep_pheatmap_matrix(dt_response = dt_metrics,
                            experiment_type = gDRutils::get_supported_experiments("sa"))

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_scores <- gDRutils::convert_se_assay_to_dt(se = se,
                                              assay_name = "scores")

mat <- prep_pheatmap_matrix(dt_response = dt_scores,
                            metric = "hsa_score",
                            normalization_type = "RV",
                            experiment_type = gDRutils::get_supported_experiments("combo"))

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{pheat_ann}
