% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{prepareCurves}
\alias{prepareCurves}
\title{prepare curve plotting data}
\usage{
prepareCurves(metrics, range_x = c(0.001, 50), density = 100)
}
\arguments{
\item{metrics}{a \code{data.table} with drug response metrics taken from a \code{SummarizedExperiment}}

\item{range_x}{range of concentration for which to predict response; numeric vector of length 2}

\item{density}{number of concentrations for which to predict response; single numeric}
}
\value{
A data.table with predicted response for the given concentration range.

Prediction is made for both GR value and Relative Viability,
across a \code{density}-long sequence of concentrations within \code{range_x}.
The concentrations are generated with \code{logSeq}.
}
\description{
Simulate and plot dose response curves for all existing combinations of Cell Line and Drug.
}
\examples{
SE <- gDRutils::get_synthetic_data("small")[[1]]
dt <- gDRutils::convert_se_assay_to_custom_dt(SE, "Metrics")
prepareCurves(dt)

}
\keyword{drawResponseOverview}
