% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricContrast.R
\name{prepareDataMC}
\alias{prepareDataMC}
\title{Prepare data for plotting Metric Contrast}
\usage{
prepareDataMC(data, choices, variable, var_col)
}
\arguments{
\item{data}{a data table containing growth metrics}

\item{choices}{a list such as one returned by the \code{selectGroups} module,
of variables and their values that will be plotted}

\item{variable}{name of column to be plotted}

\item{var_col}{name of column used for coloring dots}
}
\value{
A \code{data.table} where the column containing the two primary variable values
        has separated in two, and the column containing the secondary variable
        values remained.
        If \code{data} contains co-treatment information, it will be preserved.
}
\description{
Select columns and reshape data, bringing it to a format expected by \code{plotly_metric_contrast}.
}
\details{
If \code{variable} is one of the \code{transformed_metrics} (stored in global variable),
it will be log transformed. Infinite values that arise here will be changed to NAs,
which will be dropped in the app server before plotting. The app will display a message
if that happens
}
\seealso{
\code{plotly_metric_contrast} \code{MetricContrast}
}
\keyword{plugin_plot}
