% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metricClustering.R
\name{prepareDataMH}
\alias{prepareDataMH}
\title{Prepare matrix for clustering and heat map construction.}
\usage{
prepareDataMH(data, variable)
}
\arguments{
\item{data}{a data.table containing data and metadata of metrics, conditions, and treatments}

\item{variable}{character string; name of variable to cluster on}
}
\value{
A named numeric matrix with drugs in rows and cell lines in columns.
}
\description{
The function calls \code{dcast} to reshape the data.table.
\code{variable} is averaged at this stage if there are replicates.
}
\seealso{
\code{plotlyMH}, \code{MetricClustering}
}
\keyword{plugin_plot}
