% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-drawResponseOverview.R
\name{prepareExtras}
\alias{prepareExtras}
\title{Prepare coordinates of extra features}
\usage{
prepareExtras(metrics, range_x = c(0.001, 50))
}
\arguments{
\item{metrics}{a \code{data.table} with drug response metrics taken from a \code{SummarizedExperiment}}

\item{range_x}{range of concentration for which to predict response; numeric vector of length 2}
}
\value{
List of length 2.
The first item is a \code{data.table} with coordinates of points
passed to \code{plotly::add_markers}.
The second is a named list of line descriptions
passed the \code{shapes} argument of \code{plotly::layout}.
}
\description{
Prepare coordinates of extra features
}
\examples{
SE <- gDRutils::get_synthetic_data("small")[[1]]
dt <- gDRutils::convert_se_assay_to_custom_dt(SE, "Metrics")
prepareExtras(dt)

}
\keyword{drawResponseOverview}
