% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-internals.R
\name{reformat_untreated_cases}
\alias{reformat_untreated_cases}
\title{cleanup untreated tags}
\usage{
reformat_untreated_cases(label)
}
\arguments{
\item{label}{character/numeric vector with untreated tags to correct}
}
\value{
character/numeric vector corrected untreated tags
}
\description{
cleanup untreated tags
}
\examples{
reformat_untreated_cases("Cell Line Name: d\nDrug Name: a\n(untreated at 0 &mu;M)\nTitle: 1.00")
# Cell Line Name: d\nDrug Name: a\nTitle: 1.00
reformat_untreated_cases(c(
  "Cell Line Name: d\nDrug Name: a\n(untreated at 0 &mu;M)\nTitle: 1.00", 
  "Cell Line Name: d\nDrug Name: a\n(untreated at 4 &mu;M)\nTitle: 1.00"
 ))
 # [1] Cell Line Name: d\nDrug Name: a\nTitle: 1.00
 # [2] Cell Line Name: d\nDrug Name: a\n(untreated at 4 &mu;M)\nTitle: 1.00

}
\seealso{
\code{MetricClustering}, \code{build_label_clustering}
}
\keyword{utils}
