% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-internals.R
\name{replace_values}
\alias{replace_values}
\title{Replace one value with one or more other values in a column of a data table.}
\usage{
replace_values(x, column, replacee, replacement)
}
\arguments{
\item{x}{a \code{data.table}}

\item{column}{character string; name of column to do the replacement in}

\item{replacee}{character string; value to be replaced}

\item{replacement}{character vector of values to replace \code{replacee} with; see \code{Details}}
}
\value{
A modified data table Keep in mind that if \code{replacement} has length of more than one,
        the number of rows will be greater than that of \code{x}.

The part of \code{x} where \code{column} is equal to \code{replacee} is removed.
A copy of that part of the data is then appended where \code{replacee} has been substituted
with a single item of \code{replacement} and this repeated for each item in \code{replacement}.
The resulting data.table will be longer than \code{x} if \code{replacement} is longer than 1.
}
\description{
Replace one value with one or more other values in a column of a data table.
}
\examples{
vals <- data.table::data.table(value = runif(10, 1, 2),
group = rep(c("A", "B"), each = 10))
nts <- data.table::data.table(
  value = runif(2, 0, 0.1),
  group = rep("nt", 2)
)
dt <- rbind(vals, nts)
dtr <- replace_values(x = dt, column = "group", replacee = "nt", replacement = c("A", "B"))
                              

}
\seealso{
\code{MetricRanking}, \code{plotly_metric_ranking}
}
\keyword{utils}
