% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-utils.R
\name{round_to_unique_string}
\alias{round_to_unique_string}
\title{Round numbers to unique string}
\usage{
round_to_unique_string(num_vec, initial_digits = 4)
}
\arguments{
\item{num_vec}{a numeric vector to be rounded and change into character}

\item{initial_digits}{numeric value for number of decimal places to start rounding with}
}
\value{
a character vector of unique numeric strings.
}
\description{
Rounds a numeric vector to the minimum precision needed to ensure uniqueness
}
\examples{
vec <- c(0.00000000, 0.00000256, 0.00001280, 0.00006400, 0.00032000,
         0.00160000, 0.00800000, 0.04000000, 0.20000000, 1.00000000)

round_to_unique_string(vec)

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{utils}
