% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_drugs_1_2}
\alias{swap_drugs_1_2}
\title{Help function to change column suffix}
\usage{
swap_drugs_1_2(dt_)
}
\arguments{
\item{dt_}{data.table with data}
}
\value{
swapped data.table
}
\description{
Help function to change column suffix
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("small")
SE <- mae[[1]]
dt_ <- gDRutils::convert_se_assay_to_dt(SE, assay_name = "Metrics", wide_structure = TRUE)

swapped_dt_ <- swap_drugs_1_2(dt_)
}

}
\keyword{internal}
