% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chemical_genomics.R
\name{analyze_cgs}
\alias{analyze_cgs}
\title{Analyze Chemical Genomics (CGS) Data and Perform GSEA}
\usage{
analyze_cgs(
  dt_metrics,
  metrics,
  cl_name = NULL,
  cellline1 = NULL,
  cellline2 = NULL,
  normalization_type = "RV"
)
}
\arguments{
\item{dt_metrics}{A data.table containing screening data. Requires columns: `drug_moa`, `DrugName`, `CellLineName`,
and columns specified in the `metrics` argument.}

\item{metrics}{A character vector specifying the response metrics to analyze:
"x_mean", "x_AOC_range", "xc50", "ec50", "x_max".}

\item{cl_name}{An optional string specifying a single cell line to analyze. If NULL (default),
all cell lines in the data are analyzed. Should be NULL if \code{cellline1} and \code{cellline2} are provided.}

\item{cellline1}{An optional string representing the first cell line name. 
Should be specified alongside \code{cellline2}.}

\item{cellline2}{An optional string representing the second cell line name. 
Should be specified alongside \code{cellline1}.}

\item{normalization_type}{A string with normalization types to be selected, one of:
"GR" ("GRvalue") or "RV" ("RelativeViability").
Passed to \code{\link{prep_dt_response_metric_diff}}.}
}
\value{
A list of results, where each element corresponds to a cell line or cell line difference.
Each result contains:
\itemize{
  \item \code{fgsea}: GSEA results for the specified metrics,
  \item \code{metrics_diff}: The prepped data.table used for the GSEA analysis,
  \item \code{moa_list}: A list of DrugNames grouped by drug_moa.
}
}
\description{
This function analyzes chemical genomics (CGS) data, filters by compound mechanism of action (MOA),
prepares the data for Gene Set Enrichment Analysis (GSEA), and performs GSEA for specified cell lines and metrics.
}
\examples{
dt_metrics <- qs::qread(system.file("testdata/cgs_data.qs", package = "gDRplots"))
analyze_cgs(dt_metrics, metrics = "xc50", cl_name = "CellLineName_1")
cellline1 <- "CellLineName_1"
cellline2 <- "CellLineName_2"
analyze_cgs(dt_metrics, "xc50", cl_name = NULL, cellline1, cellline2)

}
\author{
Bartosz Czech \email{czech.bartosz@external.gene.com}
}
\keyword{cgs_plots}
