% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-associations.R
\name{calc_assoc}
\alias{calc_assoc}
\title{Calculate linear associations}
\usage{
calc_assoc(X, Y)
}
\arguments{
\item{X}{\code{matrix} Dependent variables data matrix (rows are samples, columns are features).
Must have the same length as \code{Y}.}

\item{Y}{\code{vector} or \code{matrix} Experimental reponse data (rows are samples).
Must have the same length as \code{X}.}
}
\description{
Calculate the linear model associations between dependent variables and reponse variable(s) of interest.
}
\author{
James Hawley
}
\keyword{internal}
