% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{create_download_link}
\alias{create_download_link}
\title{Prepare markdown chunk with download link}
\usage{
create_download_link(img_path, link_txt = "Download Table")
}
\arguments{
\item{img_path}{string with relative path to file with plot that will be downloaded}

\item{link_txt}{string with text describing link}
}
\value{
string with html download code
}
\description{
Generate markdown code with a html link item that, when clicked, downloads a file.
The function output should be wrapped in \code{knitr::knit()}.
}
\seealso{
\code{\link[knitr]{knit}}
}
\keyword{internal}
