% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism-plotlist.R
\name{create_PRISM_plot_list_sa}
\alias{create_PRISM_plot_list_sa}
\title{Create a nested list of plots for PRISM data with single-agent metrics}
\usage{
create_PRISM_plot_list_sa(
  drug_name_vec,
  dt_metrics,
  dt_average = NULL,
  normalization_type_vec = "RV",
  metric = c("xc50", "x_mean", "x_max"),
  fit_source = "gDR",
  meta_data_path,
  feat_data_path,
  feature_sets,
  metadata_columns = NULL,
  clear_taxonomy_info = TRUE,
  with_decoding = FALSE
)
}
\arguments{
\item{drug_name_vec}{character vector with drug names to be plotted (identifiers \code{DrugName})}

\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{dt_average}{\code{data.table} representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and single-agent \code{SummarizedExperiment}}

\item{normalization_type_vec}{character vector with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability") or both}

\item{metric}{character vector with names of metric;
 chosen from: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}),
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")}

\item{fit_source}{string source name for metrics}

\item{meta_data_path}{string path to metadata file describing all cancer models/cell lines
which are referenced by a dataset contained within the DepMap portal.
It is usually a file named \code{Model.csv} or \code{Model.csv.gz}.}

\item{feat_data_path}{string path to the directory containing the molecular feature set file to load from DepMap.}

\item{feature_sets}{character vector containing the names of the molecular feature sets to load from DepMap.
These names should also correspond to the file names containing the feature data
(without the extension, which is assumed to be \code{csv} or \code{csv.gz})}

\item{metadata_columns}{character vector with the metadata columns to load for DepMap cell lines}

\item{clear_taxonomy_info}{logical flag whether to remove taxonomy information for gene names in table
with the molecular feature sets from DepMap.}

\item{with_decoding}{logical whether the feature OmicsArmLevelCNA, OmicsSomaticMutationsMatrixHotspot
and OmicsSomaticMutationsMatrixDamaging should be encoded into a 0-1 scheme}
}
\value{
A named list with elements:
\itemize{
  \item \code{ls_plot} nested list of plots for selected type of experiment
  \item \code{ls_assoc_data} nested list of table with association data
}
}
\description{
Create a nested list of plots for PRISM data with single-agent metrics
}
\examples{
\donttest{
create_PRISM_plot_list_sa(
  drug_name_vec = c("DrugA"),
  dt_metrics = gDRutils::convert_se_assay_to_dt(se_sa, "Metrics"),
  meta_data_path = "path/to/Model.csv",
  feat_data_path = "path/to/feature_dir",
  feature_sets = c("OmicsSignaturesProfile")
)
}

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
