% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-prism-plotlist.R
\name{create_PRISM_plot_list_sa}
\alias{create_PRISM_plot_list_sa}
\title{Create nested plots list for PRISM data}
\usage{
create_PRISM_plot_list_sa(
  drug_name_vec,
  dt_metrics,
  dt_average,
  normalization_type_vec = "RV",
  metric = c("xc50", "x_mean", "x_max"),
  fit_source = "gDR",
  feature_sets,
  prefixes,
  metadata_columns = NULL
)
}
\arguments{
\item{drug_name_vec}{character vector with cell line to be plotted \code{Drug Name}}

\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{dt_average}{\code{data.table} representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{normalization_type_vec}{character vector with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability") or both}

\item{metric}{string name of metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")}

\item{fit_source}{string source name for metrics}

\item{feature_sets}{character vector names of the molecular feature set to load from DepMap.}

\item{prefixes}{character vector prefixes to use for the each feature set in \code{feature_sets};
has to be the same length as \code{feature_sets}}

\item{metadata_columns}{character vector with the metadata columns to load for DepMap cell lines}
}
\value{
list with plots (nested structure)
}
\description{
Create nested plots list for PRISM data
}
\keyword{prism_plots}
