% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism-plotlist.R
\name{create_PRISM_summary_list}
\alias{create_PRISM_summary_list}
\title{Create a list of PRISM association table}
\usage{
create_PRISM_summary_list(assoc_summary_RV, assoc_summary_GR = NULL)
}
\arguments{
\item{assoc_summary_RV}{A \code{data.table} with associations for normalization
type of "Relative Viability", outputted by \code{gDRplots::prep_assoc_summary()}}

\item{assoc_summary_GR}{A \code{data.table} with associations for normalization
type of "GR Value", outputted by \code{gDRplots::prep_assoc_summary()}}
}
\value{
A list of table split by drug name and normalization type
}
\description{
Create a list of PRISM association table
}
\examples{
dir_path <- system.file("testdata", package = "gDRplots")
ls_file <- list.files(dir_path, pattern = "[.]xlsx$")
assoc_tab <- prep_assoc_summary(dir_path = dir_path, ls_file = ls_file)
create_PRISM_summary_list(
  assoc_summary_RV = assoc_tab[grepl("RV", assoc_tab[["src"]]), ],
  assoc_summary_GR = assoc_tab[grepl("GR", assoc_tab[["src"]]), ]
)

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{prism_plots}
