% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism-plotlist.R
\name{create_PRISM_summary_list}
\alias{create_PRISM_summary_list}
\title{Create a list of PRISM association table}
\usage{
create_PRISM_summary_list(assoc_summary_RV, assoc_summary_GR)
}
\arguments{
\item{assoc_summary_RV}{A \code{data.table} with associations for normalization 
type of "Relative Viability", outputted by \code{gDRplots::::prep_assoc_summary()}}

\item{assoc_summary_GR}{A \code{data.table} with associations for normalization 
type of "GR Value", outputted by \code{gDRplots::::prep_assoc_summary()}}
}
\value{
A list of table split by drug name and normalization type
}
\description{
Create a list of PRISM association table
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
