% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{create_zoom_link}
\alias{create_zoom_link}
\title{Prepare markdown chunk with zoom link}
\usage{
create_zoom_link(img_path, link_txt = "Zoom In for Details")
}
\arguments{
\item{img_path}{string with relative path to file with plot to be shown}

\item{link_txt}{string with text describing link}
}
\value{
string with html link code
}
\description{
Generate markdown code for html link item which when clicked opens plot in new browser.
The function output should be wrapped in \code{knitr::knit()}.
}
\examples{
create_zoom_link("plots/scatter.svg")
create_zoom_link(NA)

}
\seealso{
\code{\link[knitr:knit]{knitr::knit}}
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{internal}
