% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-response_curve.R
\name{.get_combo_curves_colors}
\alias{.get_combo_curves_colors}
\title{Get color palette for the dose response curves for combination data}
\usage{
.get_combo_curves_colors(ls_conc_2)
}
\arguments{
\item{ls_conc_2}{factor vector with values for \code{Concentration_2}}
}
\value{
gDR palette for Concentration_2 given in \code{ls_conc_2}
}
\description{
Get color palette for the dose response curves for combination data
}
\examples{
\donttest{
ls_conc <- factor(c("0.001", "0.01", "1"))
.get_combo_curves_colors(ls_conc)
}
}
\keyword{internal}
