% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_combo_panel_type}
\alias{.get_combo_panel_type}
\title{Check type of concentrations set for combination of drug per cell line}
\usage{
.get_combo_panel_type(ls_vec_conc)
}
\arguments{
\item{ls_vec_conc}{a list with vectors with concentration per cell line}
}
\value{
a string decribing type of concentration list - one of:
\itemize{
  \item \code{common} vectors in the input list have common part; heatmaps in the panel
  should be created jointly with function \code{\link{heatmap_combo_with_isoref_panel_common}}
  \item \code{independent} vectors in the input list do not have common part; heatmaps in the panel
  should be created independently with function \code{\link{heatmap_combo_with_isoref_panel_independent}}
}
}
\description{
This function checks if the concentration vectors per cell line have common part or not.
It is required to decide which function to use for plotting the heatmaps panel (the set of heatmaps
with combo metrics plot for combination of selected drug with selected codrug, each heatmap per one cell line):
\code{\link{heatmap_combo_with_isoref_panel_common}} that plot heatmaps with shared axes
or \code{\link{heatmap_combo_with_isoref_panel_independent}} that plot heatmaps independently.
}
\details{
Possible combinations (0 is not taken into account as it is always present):
\itemize{
  \item \code{common}
    \itemize{
      \item all vectors have common part and start and end conc are the same\cr
            0, 0.03, 0.1, 0.3, 1\cr
      \item all vectors have common part and start and end conc are the same, but there are some gap\cr
            0, 0.003, 0.01, 0.03, 0.1, 0.3, 1\cr
            0, 0.003, 0.01, 0.03, ___, 0.3, 1\cr
         }
  \item \code{independent}
    \itemize{
      \item all vectors have common part and start conc are the same, but end conc is different (no gap)\cr
            0, 0.003, 0.01, 0.03, 0.1, 0.3, __\cr
            0, 0.003, 0.01, 0.03, 0.1, 0.3, 1\cr
      \item no common part between vectors\cr
            0, 0.003, 0.01, 0.03, __, __, __\cr
            0, ____,  ____, ____, 0.1, 0.3, 1\cr
      \item all vectors have common part but start and end conc are different (shifted range)\cr
            0, 0.003, 0.01, 0.03, 0.1, 0.3, 1, __, __\cr
            0, ____,  ____, 0.03, 0.1, 0.3, 1, 3, 10\cr
         }
}
}
\examples{
\donttest{
ls_conc <- list(c(0, 0.003, 0.01, 0.03), c(0, 0.003, 0.01, 0.03, 0.1))
gDRplots:::.get_combo_panel_type(ls_conc)
}

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{internal}
