% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_combo_panel_type}
\alias{.get_combo_panel_type}
\title{Check type of concentrations set for combination of drug per cell line}
\usage{
.get_combo_panel_type(ls_vec_conc)
}
\arguments{
\item{ls_vec_conc}{a list with vectors with concentration per cell line}
}
\value{
a string decribing type of concentration list - one of:
\itemize{
  \item \code{common} vectors in the input list have common part; heatmaps in the panel
  should be created jointly with function \code{\link{heatmap_combo_with_isoref_panel_common}}
  \item \code{independent} vectors in the input list do not have common part; heatmaps in the panel
  should be created independently with function \code{\link{heatmap_combo_with_isoref_panel_independent}}
}
}
\description{
Check type of concentrations set for combination of drug per cell line
}
\examples{
\dontrun{
ls_conc <- list(c(0, 0.003, 0.01, 0.03), c(0, 0.003, 0.01, 0.03, 0.1))
.get_combo_panel_type(ls_conc)
}

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
