% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{.get_data_type}
\alias{.get_data_type}
\title{Check data type}
\usage{
.get_data_type(dt_, desc_col = NULL)
}
\arguments{
\item{dt_}{\code{data.table} with dependent variables data in the wide format,
where rows are samples, columns are feature levels}

\item{desc_col}{a character vector with column names describing the data and which
do not contain data itself}
}
\value{
a string describing type of data - "numeric" or "categorical"
}
\description{
Check data type
}
\examples{
\donttest{
tab_cat <- data.table::data.table(
  ID = sprintf("ID_\%s", seq_len(5)),
  brown = c(0, 1, 1, 0, 0),
  blue = c(1, 0, NA, 0, 1),
  green = c(0, 0, 0, 1, 0)
)
.get_data_type(dt_ = tab_cat, desc_col = "ID")

tab_feat <- data.table::data.table(
  ID = sprintf("ID_\%s", seq_len(5)),
  grp = LETTERS[seq_len(5)],
  low = c(0, 1, 1, NA, 0),
  med = c(1, 1, NA, 0, 1),
  high = c(0, 1, 0, 1, 0)
)
.get_data_type(dt_ = tab_feat, desc_col = c("ID", "grp"))
}
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
