% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{.get_data_type}
\alias{.get_data_type}
\title{Check data type}
\usage{
.get_data_type(dt_, desc_col = NULL)
}
\arguments{
\item{dt_}{\code{data.table} with dependent variables data in the wide format,
where rows are samples, columns are feature levels}

\item{desc_col}{a character vector with column names describing the data and which 
do not contain data itself}
}
\value{
a string describing type of data - "numeric" or "categorical"
}
\description{
Check data type
}
\keyword{prism_plots}
