% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_excess_palette}
\alias{.get_excess_palette}
\title{Get color palette for the excess values}
\usage{
.get_excess_palette(no_breaks)
}
\arguments{
\item{no_breaks}{numeric number of breaks on scale}
}
\value{
gDR palette for excess values with given \code{no_breaks}
}
\description{
Get color palette for the excess values
}
\examples{
.get_excess_palette(20)
}
\keyword{internal}
