% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{.get_info_from_name}
\alias{.get_info_from_name}
\title{Help function}
\usage{
.get_info_from_name(file_name, normalization_type = "RV")
}
\arguments{
\item{file_name}{A string with file name}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
A named list with elements:
\itemize{
  \item \code{drug_grid} a string with drug name
  \item \code{feat_meta} a string with omic name (feature or meta)
}
}
\description{
This function retrieves information about the feature name and drug name from a file name,
which has the schema <chunk_name>__<feature_name>_<drug_name>_<gDR_metric_name>
}
\examples{
\dontrun{
f_n <- "name_chunk__FEAT_DRUG_ABC_RV_gDR_log10_xc50.xlsx"
get_info_from_name(f_n)
}

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
