% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_iso_colors}
\alias{.get_iso_colors}
\title{Get color palette for the isobologram levels}
\usage{
.get_iso_colors(iso_levels)
}
\arguments{
\item{iso_levels}{character vector with isobologram levels}
}
\value{
gDR palette for isoline given in \code{iso_levels}
}
\description{
Get color palette for the isobologram levels
}
\examples{
\donttest{
ls_iso_lvl <- c("0.25", "0.5", "0.75")
.get_iso_colors(ls_iso_lvl)
}
}
\keyword{internal}
