% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{.get_n_top_asssoc}
\alias{.get_n_top_asssoc}
\title{Get n-top linear associations}
\usage{
.get_n_top_asssoc(dt_assoc, n_top = 4)
}
\arguments{
\item{dt_assoc}{\code{data.table} with the calculated linear association between DepMap and metrics
outputted by \code{calc_assoc}}

\item{n_top}{number of requested top linear associations}
}
\value{
a vector with name of the n-top linear associations;
    when \code{n_top} will be higher than number of available features - only available will be returned.
}
\description{
Currently, n-top linear associations are selected based on values of \code{q_value}
(column sorted increasing) and values of \code{rho} (column sorted decreasing)
}
\examples{
\donttest{
Y <- matrix(seq(0.5, 2, length.out = 50), nrow = 50,
            dimnames = list(sprintf("row_\%s", 1:50), "met_1"))
X <- matrix(
  withr::with_seed(42, sample(c(NA, seq(0.35, 23.5, 1.25)),
                              size = 50*20, replace = TRUE)),
  nrow = 50,
  dimnames = list(sprintf("row_\%s", 1:50), sprintf("feat_\%s", 1:20)))
tab_assoc <- calc_assoc(X, Y)
gDRplots:::.get_n_top_asssoc(tab_assoc)
}

}
\keyword{prism_plots}
