% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{.get_n_top_asssoc}
\alias{.get_n_top_asssoc}
\title{Get n-top linear associations}
\usage{
.get_n_top_asssoc(dt_assoc, n_top = 4)
}
\arguments{
\item{dt_assoc}{\code{data.table} with the calculated linear association between DepMap and metrics
outputted by \code{kaleidoscope::calc_assoc}}

\item{n_top}{number of requested top linear associations}
}
\value{
a vector with name of the n-top linear associations; 
    when \code{n_top} will be higher than number of available features - only available will be returned.
}
\description{
Currently, n-top linear associations are selected based on values of \code{q_value} 
(column sorted increasing) and values of \code{rho} (column sorted decreasing)
}
\keyword{prism_plots}
