% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.get_pheatmap_cluster_param}
\alias{.get_pheatmap_cluster_param}
\title{Compute value of param cluster_rows or cluster_cols in pheatmap::pheatmap}
\usage{
.get_pheatmap_cluster_param(
  mat_to_cluster,
  distfun = stats::dist,
  additional_condition = TRUE
)
}
\arguments{
\item{mat_to_cluster}{numeric matrix to be clustered; cluster dimension must be named}

\item{distfun}{function used to compute the distance (dissimilarity) between rows;
defaults to \code{\link[stats:dist]{stats::dist}} using euclidean euclidean.}

\item{additional_condition}{additional logical flag whether rows/columns should be clustered}
}
\value{
logical flag determining if rows should be clustered or \code{hclust} object.
}
\description{
The \code{cluster_rows} and \code{cluster_cols} parameters  pheatmap::pheatmap may take values:
- boolean determining if rows/columns should be clustered
- \code{hclust} object,
this function allows to calculate proper value taking into account matrix, additional condition and
selected function used to compute the distance in \code{hclust} object
}
\details{
To compute the correct value when clustering columns - use the transposed source matrix as \code{mat_to_cluster}
}
\examples{
\donttest{
mat <- matrix(1:24, nrow = 4)
rownames(mat) <- sprintf("row_\%s", 1:4)
colnames(mat) <- sprintf("col_\%s", 1:6)
gDRplots:::.get_pheatmap_cluster_param(mat)
gDRplots:::.get_pheatmap_cluster_param(t(mat))
gDRplots:::.get_pheatmap_cluster_param(t(mat), distfun = gDRplots::compute_distances)

mat[2,2] <- NA
mat[2,1] <- Inf
gDRplots:::.get_pheatmap_cluster_param(mat)
gDRplots:::.get_pheatmap_cluster_param(mat, distfun = gDRplots::compute_distances)
gDRplots:::.get_pheatmap_cluster_param(t(mat), distfun = gDRplots::compute_distances)
add_cond <- NCOL(mat) > 10
gDRplots:::.get_pheatmap_cluster_param(mat, distfun = gDRplots::compute_distances,
                                        additional_condition = add_cond)
}

}
\seealso{
\code{\link{compute_distances}}
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
