% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.get_pheatmap_number_color}
\alias{.get_pheatmap_number_color}
\title{Compute color for number font in pheatmap::pheatmap based on given color palette and breaks}
\usage{
.get_pheatmap_number_color(
  mat_with_metric,
  colors_vec,
  breaks,
  dark_color_font = "white",
  light_color_font = "black"
)
}
\arguments{
\item{mat_with_metric}{numeric matrix with metric values; must have named rows and columns}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap;
must to be one item shorter than \code{no_breaks}}

\item{breaks}{numeric vector of breaks on scale used for mapping values to colors}

\item{dark_color_font}{string with valid color name of font for field with dark background}

\item{light_color_font}{string with valid color name of font for field without dark background}
}
\value{
named \code{matrix} with number color
}
\description{
Compute color for number font in pheatmap::pheatmap based on given color palette and breaks
}
\examples{
\donttest{
mat <- matrix(-14:30, ncol = 5,
              dimnames = list(letters[1:9], LETTERS[1:5]))
no_breaks <- 15
breaks <- seq(from = min(mat), to = max(mat), length.out = no_breaks + 1)
ls_colors <- c("limegreen", "darkblue", "orange")
hm_colors <- grDevices::colorRampPalette(ls_colors)(no_breaks)

number_color <- .get_pheatmap_number_color(mat, hm_colors, breaks)

pheatmap::pheatmap(mat,
                   breaks = breaks,
                   color = hm_colors,
                   display_numbers = TRUE,
                   number_color = number_color,
                   cluster_rows = FALSE,
                   cluster_cols = FALSE)
}

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
