% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{.get_smooth_palette}
\alias{.get_smooth_palette}
\title{Get color palette for the smooth values}
\usage{
.get_smooth_palette(no_breaks)
}
\arguments{
\item{no_breaks}{numeric number of breaks on scale}
}
\value{
gDR palette for smooth values with given \code{no_breaks}
}
\description{
Get color palette for the smooth values
}
\examples{
\donttest{
gDRplots:::.get_smooth_palette(25)
}
}
\keyword{internal}
