% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{.prep_dt_OmicsArmLevelCNA}
\alias{.prep_dt_OmicsArmLevelCNA}
\title{Encode OmicsArmLevelCNA as not mutated and mutated}
\usage{
.prep_dt_OmicsArmLevelCNA(dt_depmap)
}
\arguments{
\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap.
(rows are samples, columns are features or meta);  
outputted by one of \code{\link{prep_dt_depmap_feat}} for OmicsArmLevelCNA}
}
\value{
\code{data.table} with OmicsArmLevelCNA decoded as mutated - not mutated
}
\description{
OmicsArmLevelCNA is arm-level copy number alteration inferred using absolute copy number data 
from PureCN, method from the Ben-David et al. (2021) paper (https://www.nature.com/articles/s41586-020-03114-6).
Chromosome arms: \emph{1} indicates arm-level gain, \emph{-1} indicates arm-level loss, 
and \emph{0} indicates copy-neutral.
}
\details{
This function transform each chromosome column (e.g., \code{3p}) into two new binary columns: 
\code{3p_loss} and \code{3p_gain}. \code{3p_loss} is \emph{1} for values of \emph{-1} in the original column 
and \emph{0} otherwise. \code{3p_gain} is \emph{1} for values of \emph{1} in the original column 
and \emph{0} otherwise. The original chromosome column is then removed.
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
