% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{.prep_dt_response_scores}
\alias{.prep_dt_response_scores}
\title{Prep table with metric values for combination experiment}
\usage{
.prep_dt_response_scores(
  dt_scores,
  d_name,
  normalization_type = "RV",
  metric = "hsa_score",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_scores}{\code{data.table} representing data from the \code{scores} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "scores")}
and combo \code{SummarizedExperiment}}

\item{d_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively 
depending on \code{normalization_type}), "bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{prep_dt_assoc}}
}
\description{
Prep table with metric values for combination experiment
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_scores <- gDRutils::convert_se_assay_to_dt(se = se,
                                              assay_name = "scores")
d_name <- "drug_004"
dt_response <- .prep_dt_response_scores(dt_scores, d_name)
dt_response <- .prep_dt_response_scores(dt_scores, d_name,
                                        metric = c("hsa_score", "bliss_score"))

}
\keyword{prism_plots}
