% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.stop_on_aggregation}
\alias{.stop_on_aggregation}
\title{stop wrapper for `data.table::dcast` to handle unexpected aggregation}
\usage{
.stop_on_aggregation()
}
\description{
stop wrapper for `data.table::dcast` to handle unexpected aggregation
}
\keyword{internal}
