% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.stop_on_aggregation}
\alias{.stop_on_aggregation}
\title{stop wrapper for `data.table::dcast` to handle unexpected aggregation}
\usage{
.stop_on_aggregation(fname, formula)
}
\arguments{
\item{fname}{string with the name of the function that failed to the \code{data.table::dcast} aggregation}

\item{formula}{string with the formula used in \code{data.table::dcast}}
}
\value{
\code{NULL}
}
\description{
idea from: https://github.com/Rdatatable/data.table/issues/5386
}
\author{
Arkadiusz Gladki \email{arkadiusz.gladki@contractors.roche.com}
}
\keyword{internal}
