% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{.trim_labels}
\alias{.trim_labels}
\title{Trim labels to the required number of characters}
\usage{
.trim_labels(lbls_vec, max_lbl_length = Inf)
}
\arguments{
\item{lbls_vec}{character vector with labels to be trimmed}

\item{max_lbl_length}{numeric value for the maximum number of characters in the label;
if set to Inf, no trimming will be performed}
}
\value{
character vectors with trimmed labels
}
\description{
Trim labels to the required number of characters
}
\examples{
\donttest{
ls_lbls <- c(
  "short_lbl", "short_lbl", "veryveryverylong",
  "long_duplicates|lbl_1AB", "long_duplicates|lbl_1AB", "long_duplicates|lbl_123"
)

gDRplots:::.trim_labels(lbls_vec = ls_lbls)
gDRplots:::.trim_labels(lbls_vec = ls_lbls, max_lbl_length = 15)
}

}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{internal}
