% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{escape_special_characters}
\alias{escape_special_characters}
\title{Escape colon and hash}
\usage{
escape_special_characters(x)
}
\arguments{
\item{x}{String}
}
\value{
Original string with \code{:}s and \code{#}s escaped
}
\description{
Escape colon and hash
}
\examples{
escape_special_characters("ABC:123")
escape_special_characters("AD_12")
escape_special_characters("AD#12")

}
\keyword{internal}
