% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{fill_ann_color_map}
\alias{fill_ann_color_map}
\title{Fill missing value in color map for annotation}
\usage{
fill_ann_color_map(dt_ann, map_ann)
}
\arguments{
\item{dt_ann}{\code{data.table} with the annotations}

\item{map_ann}{\code{list} with the annotations}
}
\value{
list with color maping for the annotations
}
\description{
Fill missing value in color map for annotation
}
\examples{
\dontrun{
annotation_manual <- data.table::data.table(
  CellLineName = c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
  mut_A = c(0, 0, 1, 2, 3),
  mut_B = c("yes", "yes", "no", NA, NA),
  mut_C = c("AA", "AA", "AB", NA, "B")
)

annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90", "not_checked" = "lightblue"),
  mut_C = c("AA" = "red", "AB" = "orange", "B" = "yellow")
)

fill_ann_color_map(dt_ann = annotation_manual, map_ann = annotation_map)
}

}
\seealso{
\code{\link{pheatmap_with_anno_sa}} \code{\link{pheatmap_with_anno_combo}}
}
\keyword{utils_color}
