% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRplots-package.R
\docType{package}
\name{gDRplots-package}
\alias{gDRplots}
\alias{gDRplots-package}
\title{gDRplots: 'gDRplots'}
\value{
package help page
}
\description{
gDRplots provides static visualization functions and dedicated helpers for the gDR drug response platform. Designed for high-throughput screening (HTS), pharmacogenomics, and chemical genomics research, it enables comprehensive analysis of single-agent, co-dilution, and drug combination cell viability assays. The package supports quality control (plate layouts, concentration-response curves), metric visualization (heatmaps, waterfall charts, boxplots), and combination synergy analysis (Bliss, HSA excess maps, isobologram plots).
}
\note{
To learn more about functions start with `help(package = "gDRplots")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRplots}
  \item \url{https://gdrplatform.github.io/gDRplots/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRplots/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Janina Smola
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Dariusz Scigocki
}

}
\keyword{internal}
