% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRplots-package.R
\docType{package}
\name{gDRplots-package}
\alias{gDRplots}
\alias{gDRplots-package}
\title{gDRplots: 'gDRplots'}
\value{
package help page
}
\description{
gDRplot is a package that stores functions for gDR visualisations and their dedicated helpers.
}
\note{
To learn more about functions start with `help(package = "gDRplots")`
}
\author{
\strong{Maintainer}: gDR team \email{gdr-support-d@gene.com} (\href{https://orcid.org/0009-0006-9310-1135}{ORCID})

Authors:
\itemize{
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Janina Smola
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
}

}
\keyword{internal}
