% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{generate_datatable}
\alias{generate_datatable}
\title{Generate a customized datatable}
\usage{
generate_datatable(
  data,
  options = list(scrollX = TRUE, dom = "t"),
  width = "100\%",
  ...
)
}
\arguments{
\item{data}{A dataset to be displayed in the datatable. Must be of class 
`data.table`, `data.frame`, or `DFrame`.}

\item{options}{A list of options to customize the DataTable. Defaults to 
`list(scrollX = TRUE, dom = "t")`.}

\item{width}{A character string specifying the width of the table. Defaults to "100\%".}

\item{...}{Additional arguments passed to `DT::datatable`.}
}
\value{
A `DT::datatable` object.
}
\description{
This function creates a `DT::datatable` object with default settings for 
horizontal scrolling (`scrollX = TRUE`) and no table controls (`dom = "t"`). 
It supports input data of types `data.table`, `data.frame`, or `DFrame`. 
Additional options and arguments can be passed to customize the table further.
}
\examples{
generate_datatable(iris)

}
\author{
Bartosz Czech \email{bartosz.czech@contractors.roche.com}
}
\keyword{internal}
