% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{get_ann_color_map}
\alias{get_ann_color_map}
\title{Create color map for annotation}
\usage{
get_ann_color_map(dt_ann)
}
\arguments{
\item{dt_ann}{\code{data.table} with the annotations}
}
\value{
list with color mapping for the annotations
}
\description{
Create color map for annotation
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]][2:5, ]
dt_average <- gDRutils::convert_se_assay_to_dt(se = se, assay_name = "Averaged")
dt_ann <- dt_average[,.SD, .SDcols = c("Tissue", "ReferenceDivisionTime")]

get_ann_color_map(dt_ann)
}

}
\seealso{
\code{\link{pheatmap_qc}}
}
\keyword{utils_color}
