% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{get_r_file_path}
\alias{get_r_file_path}
\title{extract path of the executed R file}
\usage{
get_r_file_path(
  test_mode = FALSE,
  test_path = system.file(package = "gDRplots", "DESCRIPTION")
)
}
\arguments{
\item{test_mode}{should the function be run in the test mode?}

\item{test_path}{string with the path to the test file}
}
\value{
string with the path to the executed Rscript file
}
\description{
extract path of the executed R file
}
\keyword{internal}
