% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{get_r_file_path}
\alias{get_r_file_path}
\title{Extract path of the executed R file}
\usage{
get_r_file_path(test_mode = FALSE)
}
\arguments{
\item{test_mode}{logical flag whether the function be run in the test mode}
}
\value{
string with the path to the executed Rscript file
}
\description{
Extract path of the executed R file
}
\keyword{internal}
