% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_plots.R
\name{grob_SA}
\alias{grob_SA}
\title{Plot drug response curves for single-agent data}
\usage{
grob_SA(
  dt_metrics,
  dt_average,
  grouping,
  group_names = NULL,
  normalization_type = "GR",
  colormap = NULL,
  plot_averaged_flag = TRUE,
  plot_fit_flag = TRUE
)
}
\arguments{
\item{dt_metrics}{data.table representation of the data in \code{Metrics} assay
output from \code{gDRutils::convert_se_assay_to_dt(SE, "Metrics")}}

\item{dt_average}{data.table representation of the data in \code{Averaged} assay
output from \code{gDRutils::convert_se_assay_to_dt(SE, "Averaged")}}

\item{grouping}{string name of dim of SE to plot; it has to be opppsoite to this used to filter SE
(that is when rownames(SE) == 1 it has to be "cId", otherwise - "rId")}

\item{group_names}{character vector with names to subset from SE (the same dim as \code{grouping});
if \code{NULL} then all values will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colormap}{character vector with colors for \code{group_names} - name or hex value}

\item{plot_averaged_flag}{logical flag whether plot points with average values}

\item{plot_fit_flag}{logical flag whether plot points with fitted values}
}
\value{
plot with dose-response curves
}
\description{
Plot drug response curves for single-agent data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
SE <- mae[[1]]
iR <-  rownames(SE)[1]
grouping <- "cId"
dt_metrics <- gDRutils::convert_se_assay_to_dt(SE[iR], "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(SE[iR], "Averaged")
group_names <- colnames(SE)[2:5]

grob_SA(dt_metrics = dt_metrics, 
        dt_average = dt_average, 
        grouping = grouping,
        group_names = group_names)

}
\keyword{single-agent_plots}
