% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{heatmap_combo_with_isoref}
\alias{heatmap_combo_with_isoref}
\title{Plot heatmaps of averaged values for combination data}
\usage{
heatmap_combo_with_isoref(
  dt_excess,
  dt_isobolograms,
  drug1_name,
  drug2_name,
  cl_name,
  normalization_type = "GR",
  iso_levels = "0.5",
  colors_vec = NULL,
  no_breaks = 50
)
}
\arguments{
\item{dt_excess}{data.table representing data from the \code{excess} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "excess")}}

\item{dt_isobolograms}{data.table representing data from the \code{isobolograms} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_name}{string with cell line to be plotted (identifiers \code{CellLineName})}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{character vector with  isobologram levels to be selected}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap}

\item{no_breaks}{numeric number of breaks on scale}
}
\value{
list or panel with heatmaps with values for excess assays for selected drugs and cell line with
   selected isoline and comparison of iso levels
}
\description{
Plot heatmaps of averaged values for combination data
}
\examples{
cl_name <- "cellline_BC"
drug1_name <- "drug_001"
drug2_name <- "drug_026"

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_excess <- gDRutils::convert_se_assay_to_dt(se, "excess")
dt_isobolograms <- gDRutils::convert_se_assay_to_dt(se, "isobolograms")

heatmap_combo_with_isoref(dt_excess,
                          dt_isobolograms,
                          drug1_name, drug2_name,
                          cl_name)
                          
heatmap_combo_with_isoref(dt_excess,
                          dt_isobolograms,
                          drug1_name, drug2_name,
                          cl_name,
                          iso_levels = NULL,
                          colors_vec = c("darkcyan", "snow", "darkorange"))                      
                          
heatmap_combo_with_isoref(dt_excess,
                          dt_isobolograms,
                          drug1_name, drug2_name,
                          cl_name,
                          normalization_type = "RV",
                          iso_levels = c("0.25", "0.75"))

}
\keyword{combo_plots}
