% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{heatmap_combo_with_isoref_panel}
\alias{heatmap_combo_with_isoref_panel}
\title{Plot panel of heatmaps with fitted and reference data for isobolograms
to control quality of the data}
\usage{
heatmap_combo_with_isoref_panel(
  dt_excess,
  dt_isobolograms,
  drug1_name,
  drug2_name,
  cl_names,
  normalization_type = "GR",
  iso_levels = "0.5",
  colors_vec = NULL,
  no_breaks = 50
)
}
\arguments{
\item{dt_excess}{data.table representing data from the \code{excess} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "excess")}}

\item{dt_isobolograms}{data.table representing data from the \code{isobolograms} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "isobolograms")}}

\item{drug1_name}{string with drug name to be plotted (identifiers \code{DrugName})}

\item{drug2_name}{string with co-drug name to be plotted (identifiers \code{DrugName_2})}

\item{cl_names}{character vector with cell line names to be plotted (Cell Line Name);
if \code{NULL} - all available cell lines will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{iso_levels}{character vector with  isobologram levels to be selected;
when \code{NULL} - no isolines will be displayed}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap; 
the default is the dark purple-light grey palette}

\item{no_breaks}{numeric number of breaks on scale}
}
\value{
panel with heatmaps for fitted values and reference data for isobolograms
   for selected drug and co-drug by cell line names
}
\description{
Plot panel of heatmaps with fitted and reference data for isobolograms
to control quality of the data
}
\examples{
cl_names <-
  c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC", "cellline_FD")

drug1_name <- "drug_001"
drug2_name <- "drug_026"

mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_excess <- gDRutils::convert_se_assay_to_dt(se, "excess")
dt_isobolograms <- gDRutils::convert_se_assay_to_dt(se, "isobolograms")

heatmap_combo_with_isoref_panel(dt_excess,
                                dt_isobolograms,
                                drug1_name, drug2_name,
                                cl_names)

heatmap_combo_with_isoref_panel(dt_excess,
                                dt_isobolograms,
                                drug1_name, drug2_name,
                                cl_names,
                                iso_levels = c("0.25", "0.5"))

heatmap_combo_with_isoref_panel(dt_excess,
                                dt_isobolograms,
                                drug1_name, drug2_name,
                                cl_names,
                                normalization_type = "RV",
                                iso_levels = NULL,
                                colors_vec = c("darkcyan", "snow", "darkorange"))

}
\keyword{combo_plots}
