% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fitting_qc.R
\name{heatmap_control_mapping_qc}
\alias{heatmap_control_mapping_qc}
\title{Plot heatmap of mapping controls to treated for single-agent and combo data to control quality of the data}
\usage{
heatmap_control_mapping_qc(dt_treat, dt_controls)
}
\arguments{
\item{dt_treat}{data.table representation of the data in \code{RawTreated} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "RawTreated")}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}

\item{dt_controls}{data.table representation of the data in \code{Controls} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Controls")}
and \code{SummarizedExperiment} with chosen data type: single-agent or combo}
}
\value{
\code{pheatmap object} containing hetamap of mapping controls to treated
}
\description{
Plot heatmap of mapping controls to treated for single-agent and combo data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_treat <- gDRutils::convert_se_assay_to_dt(se, "RawTreated")
dt_controls <- gDRutils::convert_se_assay_to_dt(se, "Controls")

heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls)

heatmap_control_mapping_qc(dt_treat = dt_treat[1:1350, ],
                           dt_controls = dt_controls)

dt_treat_NA <- dt_treat[-c(1:135, 270:405),]
heatmap_control_mapping_qc(dt_treat = dt_treat_NA,
                           dt_controls = dt_controls)

dt_controls_NA <- dt_controls[-c(1:305, 611:763, 1221:1750),]
heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls_NA)

heatmap_control_mapping_qc(dt_treat = dt_treat,
                           dt_controls = dt_controls[1:3660, ])


}
\keyword{QC_plot}
