% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_with_anno_combo}
\alias{pheatmap_with_anno_combo}
\title{Plot pretty heatmap with annotations for combo data}
\usage{
pheatmap_with_anno_combo(
  tab_response,
  metric_growth = "GR",
  metric = "hsa_score",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = c("royalblue3", "royalblue1", "grey95", "grey95", "firebrick1",
    "firebrick3"),
  no_breaks = 50,
  annotation_col = NULL,
  annotation_colors = NULL
)
}
\arguments{
\item{tab_response}{\code{data.table} containing drug response metrics
output from \code{\link[gDRutils]{convert_se_assay_to_dt}} for assay "scores" 
and combo \code{SummarizedExperiment}}

\item{metric_growth}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively depending on \code{metric_growth}), 
"bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap}

\item{no_breaks}{numeric number of breaks on scale}

\item{annotation_col}{\code{data.table} that specifies the annotations shown above the heatmap.
Each row defines the features for a specific row. The rows in the data and in the annotation
are matched using corresponding names from \code{CellLineName} column. 
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_colors}{named list for specifying \code{annotation_col} track colors manually;
note list is named wit annotation name (column names of \code{annotation_col} - 
without \code{CellLineName}), each list item is named vector with valid colour name for 
each value describe in \code{annotation_col}). Not described elements will be colored in default.}
}
\value{
heatmap for selected metric with annotation - if given
}
\description{
Plot pretty heatmap with annotations for combo data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
response_metrics <- gDRutils::convert_se_assay_to_dt(se = se, assay_name = "scores")

annotation_manual <- data.table::data.table(
  CellLineName = c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
  mut_A = c(1, 1, 1, 0, 0),
  mut_B = c("yes", "yes", "no", "no", "no")
)

annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90")
)


pheatmap_with_anno_combo(tab_response = response_metrics)
pheatmap_with_anno_combo(tab_response = response_metrics, 
                         metric_growth = "RV",
                         metric = "bliss_score",
                         colors_vec = c("darkblue", "grey90", "darkred"),
                         annotation_col = annotation_manual)
pheatmap_with_anno_combo(tab_response = response_metrics, 
                         annotation_col = annotation_manual,
                         annotation_colors = annotation_map,
                         hm_title = get_hm_title("Combo Matrix - combo data"))
            
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\keyword{QC_plot}
