% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_with_anno_sa}
\alias{pheatmap_with_anno_sa}
\title{Plot pretty heatmap with annotations for single-agent data}
\usage{
pheatmap_with_anno_sa(
  dt_metrics,
  dt_metrics_capped = NULL,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = NULL,
  no_breaks = 50,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  distfun = compute_distances,
  annotation_row = NULL,
  annotation_col = NULL,
  annotation_colors = NULL
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{dt_metrics_capped}{\code{data.table} representing data from the \code{Metrics} assay,
the same as \code{dt_metrics} but with capped values with \code{\link[gDRutils]{cap_assay_infinities}}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean"); 
but the values from any numeric column can be displayed.}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap;
note that the first color will be assigned to the min value, and the last one - to the max}

\item{no_breaks}{numeric number of breaks on scale}

\item{cluster_rows}{logical flag whether rows should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{cluster_cols}{logical flag whether columns should be clustered;
the dendrogram will not be shown for the matrix with any dimension greater than 200.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns;
used for the dendrogram when \code{cluster_rows} or \code{cluster_cols} is set to TRUE
the default is \code{\link[gDRplots]{compute_distances}} using Spearman method.}

\item{annotation_row}{\code{data.table} that specifies the annotations shown on left side of the heatmap.
Each row defines the features for a specific row. The rows in the data and in the annotation
are matched using corresponding names from the required  \code{DrugName} column.
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_col}{\code{data.table} that specifies the annotations shown above the heatmap.
Each row defines the features for a specific column. The columns in the data and in the annotation
are matched using corresponding names from the required  \code{CellLineName} column.
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_colors}{named list for specifying \code{annotation_col} and \code{annotation_row} 
track colors manually; note list is named with annotation name (column names of \code{annotation_row} - 
without \code{DrugName} and column names of \code{annotation_col} - 
without \code{CellLineName}), each list item is named vector with valid color name for 
each value described in \code{annotation_row} and in \code{annotation_col} - respectively.
Not described elements will be colored in default.}
}
\value{
A named list with elements:
\itemize{
  \item \code{data} a list containing the information visualized in the heatmap:
    \itemize{
      \item \code{matrix} data shown in the heatmap for the selected metric.
      \item \code{annotation_row} a table with row annotations (for \code{DrugName}), if provided.
      \item \code{annotation_col} a table with column annotations (for \code{CellLineName}), if provided.
    }
  \item \code{pheatmap} object containing the heatmap itself.
}
}
\description{
Plot pretty heatmap with annotations for single-agent data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")
dt_averaged <- gDRutils::convert_se_assay_to_dt(se = se,
                                                assay_name = "Averaged")
dt_metrics_capped <-
  gDRutils::cap_assay_infinities(
    conc_assay_dt = dt_metrics,
    assay_dt = dt_averaged,
    experiment_name = gDRutils::get_supported_experiments("sa"),
    col = "xc50",
    capping_fold = 5)

output <- pheatmap_with_anno_sa(dt_metrics = dt_metrics)
hm_0 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_0[["gtable"]])

output <- pheatmap_with_anno_sa(dt_metrics = dt_metrics,)
                                dt_metrics_capped = dt_metrics_capped)
hm_1 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_1[["gtable"]])

annotation_manual_col <-
  unique(dt_metrics[, c("CellLineName", "Tissue"), with = FALSE])
annotation_manual_row <-
  unique(dt_metrics[, c("DrugName", "drug_moa"), with = FALSE])
annotation_map <-
  get_ann_color_map(unique(dt_metrics[, c("Tissue", "drug_moa"), with = FALSE]))

output <- pheatmap_with_anno_sa(dt_metrics = dt_metrics,
                                normalization_type = "RV",
                                metric = "x_mean",
                                colors_vec = c("darkblue", "grey90"),
                                annotation_row = annotation_manual_row,
                                annotation_col = annotation_manual_col,
                                annotation_colors = annotation_map)
hm_2 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_2[["gtable"]])

annotation_manual <- data.table::data.table(
  CellLineName =
    c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
  mut_A = c(1, 1, 1, 0, 0),
  mut_B = c("yes", "yes", "no", "no", "no")
)
annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90")
)

output <- pheatmap_with_anno_sa(dt_metrics = dt_metrics,
                                annotation_col = annotation_manual,
                                annotation_colors = annotation_map,
                                hm_title = get_hm_title(
                                  normalization_type = "GR",
                                  metric = "hsa_score",
                                  dataset_name = "Combo Matrix - combo data"))
hm_3 <- output[["heatmap"]]
ggpubr::as_ggplot(hm_3[["gtable"]])

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\keyword{pheat_ann}
