% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pheatmap.R
\name{pheatmap_with_anno_sa}
\alias{pheatmap_with_anno_sa}
\title{Plot pretty heatmap with annotationsfor single-agent data}
\usage{
pheatmap_with_anno_sa(
  tab_response,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  hm_title = NA,
  colors_vec = c("firebrick2", "white"),
  no_breaks = 50,
  annotation_row = NULL,
  annotation_col = NULL,
  annotation_colors = NULL
)
}
\arguments{
\item{tab_response}{\code{data.table} containing drug response metrics
outputted by \code{\link[gDRutils]{convert_se_assay_to_dt}} for the "Metrics" assay
and single-agent \code{SummarizedExperiment}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")}

\item{fit_source}{string source name for metrics}

\item{hm_title}{string plot title}

\item{colors_vec}{character vector of colors (valid name or hex) used in heatmap;
note that the first color will be assigned to the min value, and the last one - to the max}

\item{no_breaks}{numeric number of breaks on scale}

\item{annotation_row}{\code{data.table} that specifies the annotations shown on left side of the heatmap.
Each row defines the features for a specific row. The rows in the data and in the annotation
are matched using corresponding names from \code{DrugName} column. 
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_col}{\code{data.table} that specifies the annotations shown above the heatmap.
Each row defines the features for a specific column. The columns in the data and in the annotation
are matched using corresponding names from \code{CellLineName} column. 
Note that color schemes takes into account if variable is continuous or discrete.}

\item{annotation_colors}{named list for specifying \code{annotation_col} track colors manually;
note list is named with annotation name (column names of \code{annotation_row} - 
without \code{DrugName} and column names of \code{annotation_col} - 
without \code{CellLineName}), each list item is named vector with valid color name for 
each value described in \code{annotation_row} and in \code{annotation_col} - respectively.
Not described elements will be colored in default.}
}
\value{
A named list with elements:
\itemize{
  \item{\code{matrix} with data shown in the heatmap for the selected metric ,}
  \item{\code{heatmap} heatmap itself,}
  \item{\code{annotation_row} table with row annotation (for \code{DrugName}) - if given,}
  \item{\code{annotation_col} table with column annotation (for \code{CellLineName}) - if given.}
}
}
\description{
Plot pretty heatmap with annotationsfor single-agent data
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
response_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                                     assay_name = "Metrics")

output <- pheatmap_with_anno_sa(tab_response = response_metrics)
output[["heatmap"]]

annotation_manual_col <-
  unique(response_metrics[,.SD, .SDcols = c("CellLineName", "Tissue")])
annotation_manual_row <-
  unique(response_metrics[,.SD, .SDcols = c("DrugName", "drug_moa")])
annotation_map <-
  get_ann_color_map(unique(response_metrics[,.SD, .SDcols = c("Tissue", "drug_moa")]))

output <- pheatmap_with_anno_sa(tab_response = response_metrics,
                                normalization_type = "RV",
                                metric = "x_mean",
                                colors_vec = c("darkblue", "grey90"),
                                annotation_row = annotation_manual_row,
                                annotation_col = annotation_manual_col,
                                annotation_colors = annotation_map)
output[["heatmap"]]

annotation_manual <- data.table::data.table(
  CellLineName =
    c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
  mut_A = c(1, 1, 1, 0, 0),
  mut_B = c("yes", "yes", "no", "no", "no")
)
annotation_map <- list(
  mut_A = c("1" = "coral", "0" = "cadetblue"),
  mut_B = c("yes" = "black", "no" = "grey90")
)

output <- pheatmap_with_anno_sa(tab_response = response_metrics,
                                annotation_col = annotation_manual,
                                annotation_colors = annotation_map,
                                hm_title = get_hm_title(
                                  normalization_type = "GR",
                                  metric = "hsa_score",
                                  dataset_name = "Combo Matrix - combo data"))
output[["heatmap"]]

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\keyword{pheat_ann}
