% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_boxplot_meta}
\alias{plot_boxplot_meta}
\title{Plot boxplot for metric values grouped by metadata from DepMap}
\usage{
plot_boxplot_meta(
  dt_response,
  dt_depmap,
  selected_feat_meta_col,
  with_1_item_grp = TRUE,
  max_x_lbl_length = 60,
  with_inf = FALSE
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap - for one metadata;
(rows are samples, columns are metadata levels);
outputted by \code{\link{prep_dt_depmap_meta}}}

\item{selected_feat_meta_col}{string with the name of the selected metadata from \code{dt_depmap}
(will be used as a plot title)}

\item{with_1_item_grp}{logical flag indicating whether to show group with only one item}

\item{max_x_lbl_length}{numeric value for the maximum number of characters in the x-axis label}

\item{with_inf}{a logical flag indicating whether infinite values should be shown on boxplots}
}
\value{
\code{ggplot} object containing boxplots for variable levels
}
\description{
Plot boxplot for metric values grouped by metadata from DepMap
}
\examples{
dt_response <- data.table::data.table(
  CellLineName = paste0("CL", 1:10),
  xc50 = stats::rnorm(10)
)
dt_depmap <- data.table::data.table(
  CCLEName = paste0("CL", 1:10),
  Lung = as.integer(c(1, 0, 1, 0, 1, 0, 0, 0, 0, 0)),
  Breast = as.integer(c(0, 1, 0, 1, 0, 0, 0, 0, 0, 0)),
  Other = as.integer(c(0, 0, 0, 0, 0, 1, 1, 1, 1, 1))
)
plot_boxplot_meta(dt_response, dt_depmap,
                  selected_feat_meta_col = "PrimaryDisease")

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
