% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_metric.R
\name{plot_boxplot_metric_combo}
\alias{plot_boxplot_metric_combo}
\title{Plot box plots for metric for single-agent data grouped by selected variable}
\usage{
plot_boxplot_metric_combo(
  dt_scores,
  group_var,
  normalization_type = "GR",
  metric = "hsa_score",
  fit_source = "gDR",
  grouped_flag = FALSE,
  colored_pts_flag = FALSE,
  colors_vec = NULL
)
}
\arguments{
\item{dt_scores}{\code{data.table} representing data from the \code{scores} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "scores")}
and combo \code{SummarizedExperiment}}

\item{group_var}{string name of group variable; one of: \code{"CellLineName"} or \code{"DrugName"}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively 
depending on \code{normalization_type}), "bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}

\item{grouped_flag}{a logical flag whether the boxplots should be grouped and 
colored by \code{Tissue} for \code{group_var} set as \code{"CellLineName"} 
and \code{drug_moa} - for \code{"DrugName"}}

\item{colored_pts_flag}{a logical flag whether the points should be colored by grouped variable -
for \code{group_var} equal \code{"CellLineName"} points will be colored by \code{"DrugName"}
and similarly vice versa}

\item{colors_vec}{character vector with colors (name or hex value) to color boxplots; 
for \code{grouped_flag} set as FALSE only first from vector will be used.}
}
\value{
\code{ggplot} object containing boxplots for selected combo metric 
  grouped by \code{group_var}
}
\description{
Plot box plots for metric for single-agent data grouped by selected variable
}
\keyword{combo_plots}
