% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_metric.R
\name{plot_boxplot_metric_combo_by_drugs}
\alias{plot_boxplot_metric_combo_by_drugs}
\title{Plot box plots for metric for combo data grouped by drug names}
\usage{
plot_boxplot_metric_combo_by_drugs(
  dt_scores,
  normalization_type = "GR",
  metric = "hsa_score",
  fit_source = "gDR",
  colored_pts_flag = FALSE,
  colors_vec = NULL
)
}
\arguments{
\item{dt_scores}{\code{data.table} representing data from the \code{scores} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "scores")}
and combo \code{SummarizedExperiment}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively 
depending on \code{normalization_type}), "bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}

\item{colored_pts_flag}{a logical flag whether the points should be colored by grouped variable -
for \code{group_var} equal \code{"CellLineName"} points will be colored by \code{"DrugName"}
and similarly vice versa}

\item{colors_vec}{character vector with colors (name or hex value) to color boxplots; 
for \code{grouped_flag} set as FALSE only first from vector will be used.}
}
\value{
\code{ggplot} object containing boxplots for selected combo metric
   grouped by drug names
}
\description{
Plot box plots for metric for combo data grouped by drug names
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]

dt_scores <- gDRutils::convert_se_assay_to_dt(se = se,
                                              assay_name = "scores")

plot_boxplot_metric_combo_by_drugs(dt_scores)

plot_boxplot_metric_combo_by_drugs(dt_scores,
                                   normalization_type = "RV",
                                   colors_vec = "gold")

plot_boxplot_metric_combo_by_drugs(dt_scores,
                                   metric = "bliss_score",
                                   normalization_type = "RV",
                                   colored_pts_flag = TRUE)

}
\keyword{combo_plots}
