% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_metric.R
\name{plot_boxplot_metric_combo_by_grp}
\alias{plot_boxplot_metric_combo_by_grp}
\title{Plot box plots for metric for combo data grouped by selected variable}
\usage{
plot_boxplot_metric_combo_by_grp(
  dt_scores,
  selection_var,
  selection_name,
  group_var,
  group_names = NULL,
  normalization_type = "GR",
  metric = "hsa_score",
  fit_source = "gDR",
  named_n = 5,
  named_n_mode = "top",
  grouped_flag = FALSE,
  colors_vec = NULL
)
}
\arguments{
\item{dt_scores}{\code{data.table} representing data from the \code{scores} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "scores")}
and combo \code{SummarizedExperiment}}

\item{selection_var}{string name of selected main variable - one value from column
\code{"CellLineName"} or \code{"DrugName"}}

\item{selection_name}{string name of selected variable value from column \code{selection_var}
to filter data for plotting}

\item{group_var}{string name of group variable; should to be not numeric variable from \code{dt_metrics}
different than \code{selection_var} and not containing unique values for each row}

\item{group_names}{character vector with names to subset from column \code{group_var};
if \code{NULL} then all values will be plotted}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively
depending on \code{normalization_type}) or "bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}

\item{named_n}{number of points to label based on the highest or lowest \code{metric} values;
if \code{group_var} is \code{"DrugName"}, points are labeled by \code{"CellLineName"}
and similarly vice versa}

\item{named_n_mode}{string determines whether the labels are applied to the highest or lowest
values of \code{metric}; one of: \code{"top"} or \code{"bottom"}}

\item{grouped_flag}{logical flag whether the boxplots should be colored by \code{group_var}}

\item{colors_vec}{character vector with colors (name or hex value) to color boxplots;
for \code{grouped_flag} set as \code{FALSE} only first from vector will be used}
}
\value{
\code{ggplot} object containing boxplots for selected combo metric
  grouped by selected variable
}
\description{
Plot box plots for metric for combo data grouped by selected variable
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("combo")]]

dt_scores <- gDRutils::convert_se_assay_to_dt(se = se,
                                              assay_name = "scores")
invisible(dt_scores[, Tissue_grp := data.table::fifelse(Tissue == "tissue_w",
                                                        "tissue_w",
                                                        "tissue_other")])

plot_boxplot_metric_combo_by_grp(dt_scores,
                                 selection_var = "DrugName",
                                 selection_name = c("drug_001", "drug_021"),
                                 group_var = "Tissue")

plot_boxplot_metric_combo_by_grp(dt_scores,
                                 selection_var = "DrugName",
                                 selection_name = c("drug_001", "drug_021"),
                                 group_var = "Tissue_grp",
                                 grouped_flag = TRUE,
                                 colors_vec = c("darkblue", "deeppink"))

plot_boxplot_metric_combo_by_grp(dt_scores,
                                 selection_var = "DrugName",
                                 selection_name = c("drug_001", "drug_021"),
                                 group_var = "Tissue",
                                 grouped_flag = TRUE)

plot_boxplot_metric_combo_by_grp(dt_scores,
                                 selection_var = "DrugName",
                                 selection_name = c("drug_001", "drug_021"),
                                 group_var = "Tissue",
                                 colors_vec = c("darkblue", "orange"))

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{combo_plots}
