% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_metric.R
\name{plot_boxplot_metric_sa_by_CLs}
\alias{plot_boxplot_metric_sa_by_CLs}
\title{Plot box plots for metric for single-agent data grouped by cell line names}
\usage{
plot_boxplot_metric_sa_by_CLs(
  dt_metrics,
  normalization_type = "GR",
  metric = "xc50",
  fit_source = "gDR",
  grouped_flag = FALSE,
  colors_vec = NULL,
  with_inf = FALSE
)
}
\arguments{
\item{dt_metrics}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the metric;
one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
"x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean");
but the values from any numeric colum can be displayed.}

\item{fit_source}{string source name for metrics}

\item{grouped_flag}{a logical flag whether the boxplots should be grouped and 
colored by \code{Tissue}}

\item{colors_vec}{character vector with colors (name or hex value) to color boxplots}

\item{with_inf}{a logical flag whether the infinity valuse should be shown on boxplots or not}
}
\value{
\code{ggplot} object containing boxplots for selected single-agent grouped by cellline names
}
\description{
Plot box plots for metric for single-agent data grouped by cell line names
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")

plot_boxplot_metric_sa_by_CLs(dt_metrics)

plot_boxplot_metric_sa_by_CLs(dt_metrics,
                              grouped_flag = TRUE)
                              
plot_boxplot_metric_sa_by_CLs(dt_metrics,
                              metric = "x_AOC",
                              colors_vec = "gold")

plot_boxplot_metric_sa_by_CLs(
  dt_metrics,
  metric = "x_max",
  grouped_flag = TRUE)

plot_boxplot_metric_sa_by_CLs(
  dt_metrics,
  metric = "x_mean",
  grouped_flag = TRUE,
  colors_vec = c("deeppink", "darkcyan", "orange", "darkblue"))

}
\keyword{single-agent_plots}
