% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_boxplot_num}
\alias{plot_boxplot_num}
\title{Plot boxplot for categorical features}
\usage{
plot_boxplot_num(
  dt_response,
  dt_depmap,
  selected_feat,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features).
outputted by \code{\link{prep_dt_depmap_feat}}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap (will be used as a plot title)
that has 0-1 values only (categorical character but without relation one-to-one for ids and feature)}
}
\value{
\code{ggplot} object containing boxplots for variable levels
}
\description{
Plot boxplot for categorical features
}
\examples{
mae_prism <- gDRutils::get_synthetic_data("prism")
se_prism <- mae_prism[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se_prism, "Metrics")
dt_response <- prep_dt_response_metric_sa(
  dt_metrics = dt_metrics,
  d_name = unique(dt_metrics[["DrugName"]])[1],
  normalization_type = "RV", metric = "x_mean")
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
obj_depmap_meta <- prep_dt_depmap_meta(meta_data_path, metadata_col = "OncotreeLineage")
plot_boxplot_num(
  dt_response = dt_response,
  dt_depmap = obj_depmap_meta[["dt_depmap"]],
  selected_feat = names(obj_depmap_meta[["dt_depmap"]])[3])

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
