% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_boxplot_num}
\alias{plot_boxplot_num}
\title{Plot boxplot for categorical features}
\usage{
plot_boxplot_num(
  dt_response,
  dt_depmap,
  selected_feat,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link[gDRplots]{prep_dt_response_metric_sa}},
\code{\link[gDRplots]{prep_dt_response_dose_sa}}, \code{\link[gDRplots]{prep_dt_response_scores}}
or \code{\link[gDRplots]{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features). 
outputted by \code{\link[gDRplots]{prep_dt_depmap_feat}}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap (will be used as a plot title)
that has 0-1 values only (categorical character but without relation one-to-one for ids and feature)}
}
\value{
\code{ggplot} object containing boxplots for variable levels
}
\description{
Plot boxplot for categorical features
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
