% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_boxplot_num_panel}
\alias{plot_boxplot_num_panel}
\title{Plot panel with boxplots}
\usage{
plot_boxplot_num_panel(
  dt_response,
  dt_depmap,
  selected_feats,
  selected_feat_meta_col = NULL,
  ncol = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features).
outputted by \code{\link{prep_dt_depmap_feat}}}

\item{selected_feats}{character vector with names of selected features from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap (will be used as a plot title)
that has 0-1 values only (categorical character but without relation one-to-one for ids and feature)}

\item{ncol}{number of plot column in panel}
}
\value{
\code{ggplot} object containing boxplots for variable levels (0-1)
}
\description{
Plot panel with boxplots
}
\examples{
dt_response <- data.table::data.table(
  CellLineName = paste0("CL", 1:10),
  xc50 = stats::rnorm(10)
)
dt_depmap <- data.table::data.table(
  CCLEName = paste0("CL", 1:10),
  MutA = as.integer(c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0)),
  MutB = as.integer(c(0, 1, 0, 1, 0, 1, 0, 1, 0, 1))
)
plot_boxplot_num_panel(dt_response, dt_depmap,
                       selected_feats = c("MutA", "MutB"))

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
