% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chemical_genomics.R
\name{plot_cgs_ranking}
\alias{plot_cgs_ranking}
\title{Plot Chemical Genomics Screen GSEA Results}
\usage{
plot_cgs_ranking(
  results,
  cl_name,
  metric,
  padj_threshold = 0.1,
  top_results_no_sig = 5,
  max_results_with_sig = 15
)
}
\arguments{
\item{results}{A list object returned from `analyze_cgs`.}

\item{cl_name}{A string specifying the cell line included in the \code{results} to prepare a visualization.}

\item{metric}{A string specifying the metric included in the \code{results} to prepare a visualization.}

\item{padj_threshold}{A numeric value specifying the threshold for filtering significant GSEA results
based on adjusted p-value.}

\item{top_results_no_sig}{A numeric value specifying the number of top results to plot
if there are no significant values.}

\item{max_results_with_sig}{A numeric value specifying the maximum number of results
to plot when there are more than this number of significant values.}
}
\value{
A ggplot2 object with cgs results
}
\description{
Generates a ggplot2 visualization of chemical genomics screening data, highlighting GSEA results.
}
\examples{
dt_metrics <- qs::qread(system.file("testdata/cgs_data.qs", package = "gDRplots"))
results <- analyze_cgs(dt_metrics, metrics = c("xc50"), cl_name = "CellLineName_1")
plot_cgs_ranking(results,
  cl_name = "CellLineName_1",
  metric = "xc50",
  padj_threshold = 0.1,
  top_results_no_sig = 5,
  max_results_with_sig = 15)
}
\author{
Bartosz Czech \email{bartosz.czech@contractors.roche.com}
}
\keyword{cgs_plots}
