% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_genomics.R
\name{plot_cgs_ranking}
\alias{plot_cgs_ranking}
\title{Plot Chemical Genomics Screen GSEA Results}
\usage{
plot_cgs_ranking(results, cell_line, metric, yrange = 0.7)
}
\arguments{
\item{results}{A list object returned from `analyze_cgs`.}

\item{cell_line}{A character string specifying the cell line to prepare data for.}

\item{metric}{A character string specifying the metric to prepare data for.}

\item{yrange}{A numeric value to adjust the y-axis range. Defaults to 0.7}
}
\value{
A ggplot2 object. The plot is also printed to the console.
}
\description{
Generates a ggplot2 visualization of chemical genomics screening data, highlighting GSEA results.
}
\examples{
metrics_data <- qs::qread(system.file("testdata/cgs_data.qs", package = "gDRplots"))
results <- analyze_cgs(metrics_data, metrics = c("xc50"), cell_line = "A549")
plot_cgs_ranking(results, cell_line = "A549", metric = "xc50")
}
