% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sa-response_curve.R
\name{plot_dose_response_sa}
\alias{plot_dose_response_sa}
\title{Plot drug response curves for single-agent data}
\usage{
plot_dose_response_sa(
  dt_metrics,
  dt_average,
  selection_name,
  group_var,
  group_names = NULL,
  normalization_type = "GR",
  colors_vec = NULL,
  plot_averaged_flag = TRUE,
  plot_fit_flag = TRUE,
  fit_source = "gDR"
)
}
\arguments{
\item{dt_metrics}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{dt_average}{data.table representing data from the \code{Averaged} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
and single-agent \code{SummarizedExperiment}}

\item{selection_name}{string name of selected main variable - one value from column 
\code{"CellLineName"} or \code{"DrugName"}}

\item{group_var}{string name of group variable; one of: \code{"CellLineName"} or \code{"DrugName"}}

\item{group_names}{character vector with names to subset from se (the same dim as \code{group_var});
if \code{NULL} then all values will be plotted}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{colors_vec}{character vector with colors for \code{group_names} - name or hex value}

\item{plot_averaged_flag}{logical flag whether plot points with average values}

\item{plot_fit_flag}{logical flag whether plot points with fitted values}

\item{fit_source}{string source name for metrics}
}
\value{
plot with dose-response curves
}
\description{
Plot drug response curves for single-agent data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
selected_drug <- "drug_002"
group_var <- "CellLineName"
dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
celline_names <- unique(dt_metrics[[group_var]])[1:3]

plot_dose_response_sa(dt_metrics = dt_metrics,
                      dt_average = dt_average,
                      selection_name = selected_drug,
                      group_var = group_var,
                      group_names = celline_names)
            
selected_cellline <- "cellline_HB"
group_var <- "DrugName"
dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
group_names <- unique(dt_metrics[[group_var]])[1:3]

plot_dose_response_sa(dt_metrics = dt_metrics,
                      dt_average = dt_average,
                      selection_name = selected_cellline,
                      group_var = group_var,
                      group_names = group_names)
                      
}
\keyword{single-agent_plots}
