% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sa-response_curve.R
\name{plot_dose_response_sa_qc}
\alias{plot_dose_response_sa_qc}
\title{Plot drug response curves for single-agent data to check quality of data}
\usage{
plot_dose_response_sa_qc(
  dt_metrics,
  dt_average,
  cl_name,
  d_name,
  metric_growth = "GR",
  fit_source = "gDR"
)
}
\arguments{
\item{dt_metrics}{data.table representation of the data in \code{Metrics} assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}}

\item{dt_average}{data.table representation of the data in \code{Averaged} assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}}

\item{cl_name}{string cell line to be plotted (Cell Line Name)}

\item{d_name}{string vector with cell line to be plotted (Drug Name)}

\item{metric_growth}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{fit_source}{string source name for metrics}
}
\value{
plot with dose-response curves
}
\description{
Plot drug response curves for single-agent data to check quality of data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[1]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
cl_name <- dt_metrics[["CellLineName"]][1]
d_name <- dt_metrics[["DrugName"]][1]


plot_dose_response_sa_qc(dt_metrics = dt_metrics,
                         dt_average = dt_average,
                         cl_name = cl_name,
                         d_name = d_name)

}
\keyword{QC_plot}
