% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plate.R
\name{plot_plate_stack_info}
\alias{plot_plate_stack_info}
\title{Plot plate data}
\usage{
plot_plate_stack_info(dt_plate)
}
\arguments{
\item{dt_plate}{The data table containing plate information}
}
\value{
A named list of ggplot objects for each barcode
}
\description{
Plot plate data
}
\examples{
test_data <- data.table::data.table(
  WellColumn = rep(1:12, each = 8),
  clid = c("A", "A", "B", "B"),
  WellRow = rep(LETTERS[1:8], times = 12),
  Gnumber = c(rep("untreated", 48), sample(1:5, size = 48, replace = TRUE)),
  Gnumber_2 = c(rep("untreated", 48), sample(1:5, size = 48, replace = TRUE)),
  Concentration = sample(seq(0, 2.5, 1.25), 12, replace = TRUE),
  Concentration_2 = sample(seq(0, 2.5, 1.25), 12, replace = TRUE),
  ReadoutValue = runif(96, min = 0, max = 100),
  Barcode = rep(c("A", "B"), 48)
  )
plot_plate_stack_info(test_data)[["A"]]
}
\keyword{QC_plot}
