% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_scatter_with_corr}
\alias{plot_scatter_with_corr}
\title{Plot scatter with correlation}
\usage{
plot_scatter_with_corr(
  dt_response,
  dt_depmap,
  selected_feat,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features).
outputted by \code{\link{prep_dt_depmap_feat}}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap
(will be used as a plot title)}
}
\value{
\code{ggplot} object containing a scatter plot with correlation
}
\description{
Plot scatter with correlation
}
\examples{
dt_response <- data.table::data.table(
  CellLineName = paste0("CL", 1:10),
  xc50 = stats::rnorm(10)
)
dt_depmap <- data.table::data.table(
  CCLEName = paste0("CL", 1:10),
  GeneExpr = stats::rnorm(10)
)
plot_scatter_with_corr(dt_response, dt_depmap, selected_feat = "GeneExpr")

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
