% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_scatter_with_corr}
\alias{plot_scatter_with_corr}
\title{Plot scatter with correlation}
\usage{
plot_scatter_with_corr(
  dt_response,
  dt_depmap,
  selected_feat,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link[gDRplots]{prep_dt_response_metric_sa}},
\code{\link[gDRplots]{prep_dt_response_dose_sa}}, \code{\link[gDRplots]{prep_dt_response_scores}}
or \code{\link[gDRplots]{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features). 
outputted by \code{\link[gDRplots]{prep_dt_depmap_feat}}}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap 
(will be used as a plot title)}
}
\value{
a scatter plot with correlation
}
\description{
Plot scatter with correlation
}
\keyword{prism_plots}
