% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_scatter_with_corr}
\alias{plot_scatter_with_corr}
\title{Plot scatter with correlation}
\usage{
plot_scatter_with_corr(dt_response, dt_depmap, selected_feat)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link[gDRplots]{.prep_dt_response_metric_sa}},
\code{\link[gDRplots]{.prep_dt_response_dose_sa}}, \code{\link[gDRplots]{.prep_dt_response_scores}}
or \code{\link[gDRplots]{.prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one
feature or one metadata; (rows are samples, columns are features or meta).}

\item{selected_feat}{string with name of selected feature from \code{dt_depmap}}
}
\value{
a scatter plot with correlation
}
\description{
Plot scatter with correlation
}
\keyword{prism_plots}
