% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_scatter_with_corr_panel}
\alias{plot_scatter_with_corr_panel}
\title{Plot panel with scatter with correlation}
\usage{
plot_scatter_with_corr_panel(
  dt_response,
  dt_depmap,
  selected_feats,
  selected_feat_meta_col = NULL
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link[gDRplots]{prep_dt_response_metric_sa}},
\code{\link[gDRplots]{prep_dt_response_dose_sa}}, \code{\link[gDRplots]{prep_dt_response_scores}}
or \code{\link[gDRplots]{prep_dt_response_metric_diff}},}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap - for one feature;
(rows are samples, columns are features). 
outputted by \code{\link[gDRplots]{prep_dt_depmap_feat}}}

\item{selected_feats}{character vector with names of selected features from \code{dt_depmap}}

\item{selected_feat_meta_col}{string with the name of a feature column in DepMap 
(will be used as a plot title)}
}
\value{
\code{ggplot} object containing panel of scatter plot with correlation for selected features
}
\description{
Plot panel with scatter with correlation
}
\examples{
Y <- matrix(seq(0.5, 2, length.out = 50), nrow = 50,
            dimnames = list(sprintf("row_\%s", 1:50), "met_1"))
X <- matrix(
  withr::with_seed(42, sample(c(NA, seq(0.35, 23.5, 1.25)), 50*20, replace = TRUE)), nrow = 50,
  dimnames = list(sprintf("row_\%s", 1:50), sprintf("feat_\%s", 1:20)))
tab_assoc <- calc_assoc(X, Y)
plot_volcano_assoc(tab_assoc,
                   selected_feat_meta_col = "feat_XY",
                   selected_metric = "met_RV")

}
\keyword{prism_plots}
