% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-normalization_qc.R
\name{plot_var_stat_qc}
\alias{plot_var_stat_qc}
\title{Lollipop plot for metric single-agent data to check quality of data}
\usage{
plot_var_stat_qc(dt_assay, cl_name, metric = "r2", metric_growth = "GR")
}
\arguments{
\item{dt_assay}{data.table representation of the data in assay
output from \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}}

\item{cl_name}{string cell line name to be plotted (Cell Line Name)}

\item{metric}{string with variable name to be plotted; it has to be in \code{dt_assay}}

\item{metric_growth}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
lollipop plot with stat value for each drug
}
\description{
Lollipop plot for metric single-agent data to check quality of data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
cl_name <- dt_metrics[["CellLineName"]][1]

plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name)
                         
plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name,
                 metric = "r2",
                 metric_growth = "RV")
                         
plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name,
                 metric = "x_AOC",
                 metric_growth = "RV")
                         
}
\keyword{QC_plot}
