% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fitting_qc.R
\name{plot_var_stat_qc}
\alias{plot_var_stat_qc}
\title{Lollipop plot for metric single-agent data to control quality of the data}
\usage{
plot_var_stat_qc(
  dt_assay,
  cl_name,
  metric = "r2",
  normalization_type = "GR",
  with_table = FALSE
)
}
\arguments{
\item{dt_assay}{data.table representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and single-agent \code{SummarizedExperiment}}

\item{cl_name}{string cell line name to be plotted (Cell Line Name)}

\item{metric}{string with variable name to be plotted; it has to be in \code{dt_assay}}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{with_table}{logical whether table with metric values should be shown next to the plot}
}
\value{
\code{ggplot} object containing lollipop plot with stat value for each drug
}
\description{
Lollipop plot for metric single-agent data to control quality of the data
}
\examples{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]

dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
cl_name <- dt_metrics[["CellLineName"]][1]

plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name)
                         
plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name,
                 metric = "r2",
                 normalization_type = "RV")
                         
plot_var_stat_qc(dt_assay = dt_metrics,
                 cl_name = cl_name,
                 metric = "x_AOC",
                 normalization_type = "RV",
                 with_table = TRUE)
                         
}
\author{
Bartosz Czech \email{czech.bartosz@external.gene.com}
}
\keyword{QC_plot}
