% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_volcano_assoc}
\alias{plot_volcano_assoc}
\title{Volcano plot with association}
\usage{
plot_volcano_assoc(
  dt_assoc,
  condition_txt,
  correlation_txt,
  q_cutoff = 0.05,
  named_p_top = 10,
  max_N = NULL
)
}
\arguments{
\item{dt_assoc}{\code{data.table} with the calculated linear association between depmap and metrics
outputted by \code{kaleidoscope::calc_assoc}}

\item{condition_txt}{string describing experiment condition 
(preferred: \code{"DrugName"}_\code{"Gnumber"}_\code{"drug_moa"}_\code{"Duration"})}

\item{correlation_txt}{string describing association 
(preferred: \code{metric}_\code{depmap feature of metdata})}

\item{q_cutoff}{numeric cutoff to identify statistically significant correlations}

\item{named_p_top}{numeric value for p-top statistically significant correlations to be labeled on the plot}

\item{max_N}{numeric value for limit maximum number of non-statistically significant points to plot; 
for default \code{NULL} all points will be plotted.}
}
\value{
a volcano plot with association
}
\description{
Volcano plot with association
}
\keyword{prism_plots}
