% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_volcano_assoc_panel}
\alias{plot_volcano_assoc_panel}
\title{Plot panel with volcano plot and according to the data type - scatter plots or box plots}
\usage{
plot_volcano_assoc_panel(
  dt_response,
  dt_depmap,
  selected_metric,
  selected_feat_meta_col
)
}
\arguments{
\item{dt_response}{\code{data.table} with the experimental response data (rows are samples)
for one metric outputted by one of functions: \code{\link{prep_dt_response_metric_sa}},
\code{\link{prep_dt_response_dose_sa}}, \code{\link{prep_dt_response_scores}}
or \code{\link{prep_dt_response_metric_diff}},
must have at least a column with \code{CellLineName} and a numeric column with metric values.}

\item{dt_depmap}{\code{data.table} with dependent variables data loaded from DepMap
where rows are samples, columns are features/metadata levels;
one of: data for one feature outputted by \code{\link{prep_dt_depmap_feat}} or data
or data for one metadata outputted by \code{\link{prep_dt_depmap_meta}}}

\item{selected_metric}{string name of the metric in \code{dt_response}}

\item{selected_feat_meta_col}{string with name of selected feature from \code{dt_depmap} or
the name of the selected metadata from \code{dt_depmap} - respectively}
}
\value{
A named list with elements:
\itemize{
  \item \code{assoc_data} table with association data
  \item \code{panel} \code{ggplot} object containing a panel with volcano plot and depending
     on data type: a scatter plots with correlation for top 4 variables or boxplots for variable levels
}
}
\description{
Plot panel with volcano plot and according to the data type - scatter plots or box plots
}
\examples{
\dontshow{if (nzchar(system.file("depmap_data", package = "gDRtestData"))) withAutoprint(\{ # examplesIf}
mae_prism <- gDRutils::get_synthetic_data("prism")
se_prism <- mae_prism[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se_prism, "Metrics")
dt_response <- prep_dt_response_metric_sa(
  dt_metrics = dt_metrics,
  d_name = unique(dt_metrics[["DrugName"]])[1],
  normalization_type = "RV", metric = "x_mean")
feat_data_path <- system.file("depmap_data", package = "gDRtestData")
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
obj_depmap_feat <- prep_dt_depmap_feat(feat_data_path, meta_data_path,
                                       feature_set = "OmicsCNGene")
plot_volcano_assoc_panel(
  dt_response = dt_response,
  dt_depmap = obj_depmap_feat[["dt_depmap"]],
  selected_metric = "RV_gDR_x_mean",
  selected_feat_meta_col = obj_depmap_feat[["selected_feat_meta_col"]])
\dontshow{\}) # examplesIf}
}
\author{
Janina Smoła \email{janina.smola@external.roche.com}
}
\keyword{prism_plots}
