% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prism.R
\name{plot_volcano_box_panel}
\alias{plot_volcano_box_panel}
\title{Plot panel with volcano plot and boxplot for metric values grouped by metadata from DepMap}
\usage{
plot_volcano_box_panel(
  dt_response,
  dt_depmap,
  selected_metric,
  selected_feat_meta_col
)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric
outputted by one of functions: \code{\link[gDRplots]{prep_dt_response_metric_sa}},
\code{\link[gDRplots]{prep_dt_response_dose_sa}}, \code{\link[gDRplots]{prep_dt_response_scores}}
or \code{\link[gDRplots]{prep_dt_response_metric_diff}}, 
must have at least a column with \code{CellLineName} and a numeric column with metric values.}

\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap - for one metadata;
(rows are samples, columns are metadata levels);  
outputted by \code{\link[gDRplots]{prep_dt_depmap_meta}}}

\item{selected_metric}{string name of the metric in \code{dt_response}}

\item{selected_feat_meta_col}{string with the name of the selected metadata from \code{dt_depmap}}
}
\value{
a panel with volcano plot and boxplot for matadata levels
}
\description{
Plot panel with volcano plot and boxplot for metric values grouped by metadata from DepMap
}
\keyword{prism_plots}
