% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_assoc_summary}
\alias{prep_assoc_summary}
\title{Prepare summary table with statistically significant associations}
\usage{
prep_assoc_summary(
  dir_path,
  ls_file,
  alpha = 0.05,
  n_stat_sig_row = 10,
  read_file_fun = readxl::read_excel,
  as_list = FALSE
)
}
\arguments{
\item{dir_path}{A string path to the directory containing files with associations data.}

\item{ls_file}{A character vector with names of files containing associations data.}

\item{alpha}{A numeric cutoff to identify statistically significant correlations}

\item{n_stat_sig_row}{A numeric value for limit the maximum number of statistically significant 
associations to subset form table}

\item{read_file_fun}{A function to read the data from file; default is \code{readxl::read_excel}}

\item{as_list}{A logical flag weather result should be return as a list or as a table.}
}
\value{
A \code{DT::datatable} object.
}
\description{
Prepare summary table with statistically significant associations
}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{internal}
