% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_double_table_chunk}
\alias{prep_double_table_chunk}
\title{Prepare markdown chunk based on a doubly nested list of tables}
\usage{
prep_double_table_chunk(
  tbl_list,
  chunk_name,
  dwn_list = NULL,
  header_level = 3,
  tabset_options = c("tabset", "tabset-dropdown"),
  sorting_opts = NULL
)
}
\arguments{
\item{tbl_list}{A doubly nested named list of tables. The outer list represents cell lines,
and the inner lists represent metrics. Names are used as headings.}

\item{chunk_name}{A string representing the name of the chunk.}

\item{dwn_list}{A list for download links, corresponding to the structure of tbl_list. 
If NULL, no download links are included.}

\item{header_level}{An integer specifying the header level for the markdown.}

\item{tabset_options}{A character vector specifying tabset options, such as "tabset" or 
"tabset-dropdown". If NULL, no tabset options are applied.}

\item{sorting_opts}{A vector specifying global sorting options for all tables. 
Column names can be preceded by "-" to indicate descending order.}
}
\value{
A list of character vectors. Each element corresponds to a cell line. Each character vector
  represents markdown code for the cell line's tabset.
}
\description{
Generates markdown code for displaying tables in a document using `knitr::knit()`.
Handles doubly nested lists, allowing for tabbed sections for cell lines and then metrics.
The inner header level (for metrics) is automatically set to one level greater than the outer header level.
}
\examples{
\dontrun{
nested_tables <- list(
  CellLine1 = list(MetricA = mtcars[1:5, ], MetricB = mtcars[6:10, ]),
  CellLine2 = list(MetricC = iris[1:5, ], MetricD = iris[6:10, ])
)
sorting_options <- c("cyl", "-hp") # Apply the same sorting to all tables
prep_double_table_chunk(nested_tables, "nested_tables", header_level = 2, tabset_options = "tabset", sorting_opts = sorting_options)
}

}
\keyword{internal}
