% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_assoc}
\alias{prep_dt_assoc}
\title{Prep table with calculated linear associations}
\usage{
prep_dt_assoc(dt_response, dt_depmap, selected_feat_meta_col = NULL)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples) for one metric}

\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap.
(rows are samples, columns are features or meta);
outputted by one of \code{\link{prep_dt_depmap_feat}} or
\code{\link{prep_dt_depmap_meta}}}

\item{selected_feat_meta_col}{string name of feature/meta column in DepMap}
}
\value{
A named list with elements, that may be input to \code{\link{plot_volcano_assoc}}
\itemize{
  \item \code{dt_assoc} \code{data.table} with calculated association values between
     feature/meta of DepMap and selected metric,
  \item \code{condition_info} string describing experiment condition (drugs),
  \item \code{selected_feat_meta_col} string name of feature/meta.
}
}
\description{
Prep table with calculated linear associations
}
\examples{
dt_response <- data.table::data.table(
  CellLineName = paste0("CL", 1:10),
  xc50 = stats::rnorm(10)
)
dt_depmap <- data.table::data.table(
  CCLEName = paste0("CL", 1:10),
  GeneExprA = stats::rnorm(10),
  GeneExprB = stats::rnorm(10)
)
prep_dt_assoc(dt_response, dt_depmap,
              selected_feat_meta_col = "OmicsSignaturesProfile")

}
\author{
Janina Smoła \email{janina.smola@contractors.roche.com}
}
\keyword{prism_plots}
