% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_assoc}
\alias{prep_dt_assoc}
\title{Prep table with calculated linear associations}
\usage{
prep_dt_assoc(dt_response, dt_depmap)
}
\arguments{
\item{dt_response}{\code{data.table} with experimental response data (rows are samples).}

\item{dt_depmap}{\code{data.table} with dependent variables data load from DepMap.
(rows are samples, columns are features or meta).}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{plot_volcano_assoc}}
}
\description{
Prep table with calculated linear associations
}
\keyword{prism_plots}
