% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_feat}
\alias{prep_dt_depmap_feat}
\title{Load DepMap merged data for one selected feature}
\usage{
prep_dt_depmap_feat(
  feat_data_path,
  meta_data_path,
  feature_set = "CRISPRGeneEffect",
  with_decoding = FALSE
)
}
\arguments{
\item{feat_data_path}{string with path to the directory containing the molecular
feature set file to load from DepMap.}

\item{meta_data_path}{string with path to metadata file describing all cancer models/cell lines
which are referenced by a dataset contained within the DepMap portal.
It is usually a file named \code{Model.csv} or \code{Model.csv.gz}.}

\item{feature_set}{string containing the name of the molecular feature set to load from DepMap.
This name should also correspond to the file containing the feature data
(without the extension, which is assumed to be \code{csv} or \code{csv.gz})}

\item{with_decoding}{logical whether the feature OmicsArmLevelCNA, OmicsSomaticMutationsMatrixHotspot
and OmicsSomaticMutationsMatrixDamaging should be encoded into a 0-1 scheme}
}
\value{
A named list with elements, that may be input to \code{\link{prep_dt_assoc}}
\itemize{
  \item \code{dt_depmap} \code{data.table} with feature data from DepMap (wide format),.
  \item \code{selected_feat_meta_col} string name of feature.
}
}
\description{
Load DepMap merged data for one selected feature
}
\examples{
feat_data_path <- system.file("depmap_data", package = "gDRtestData")
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
dt_depmap_feat <- prep_dt_depmap_feat(feat_data_path = feat_data_path,
                                      meta_data_path = meta_data_path)
}
\keyword{internal}
