% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_feat}
\alias{prep_dt_depmap_feat}
\title{Load DepMap merged data for one selected feature}
\usage{
prep_dt_depmap_feat(
  feat_data_path,
  meta_data_path,
  feature_set = "CRISPRGeneEffect.csv"
)
}
\arguments{
\item{feat_data_path}{string path to the dfolder with file the molecular feature set to load from DepMap.}

\item{meta_data_path}{string path to metadata file describing all cancer models/cell lines
which are referenced by a dataset contained within the DepMap portal. 
It is usually a file named \code{Model.csv}.}

\item{feature_set}{string file name of the molecular feature set to load from DepMap.}
}
\value{
A named list with elements, that may be input to \code{\link[gDRplots]{prep_dt_assoc}}
\itemize{
  \item \code{dt_depmap} \code{data.table} with feature data from DepMap (wide format),.
  \item \code{selected_feat_meta_col} string name of feature.
}
}
\description{
Load DepMap merged data for one selected feature
}
\examples{
\dontrun{
feat_data_path <- file.path(".", "depmapdata")
meta_data_path <- file.path(".", "Model.csv")
dt_depmap_feat <- prep_dt_depmap_feat(feat_data_path = feat_data_path,
                                      meta_data_path = meta_data_path)
}
}
\seealso{
\code{kaleidoscope::load_depmap_merged}
}
\keyword{internal}
