% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_feat}
\alias{prep_dt_depmap_feat}
\title{Load DepMap merged data for one selected feature}
\usage{
prep_dt_depmap_feat(feature_set = "CRISPRGeneEffect", prefix = "KO_")
}
\arguments{
\item{feature_set}{string name of the molecular feature set to load from DepMap.}

\item{prefix}{string prefixes to use for the each feature set in \code{feature_set};
has to be the same length as \code{feature_sets}}
}
\value{
A named list with elements, that may be input to \code{\link[gDRplots]{prep_dt_assoc}}
\itemize{
  \item \code{dt_depmap} \code{data.table} with feature data from DepMap (wide format),.
  \item \code{selected_feat_meta_col} string name of feature.
}
}
\description{
Load DepMap merged data for one selected feature
}
\examples{
\dontrun{
dt_depmap_feat <- prep_dt_depmap_feat() 
}

}
\seealso{
\code{kaleidoscope::load_depmap_merged}
}
\keyword{internal}
