% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_meta}
\alias{prep_dt_depmap_meta}
\title{Load DepMap merged data for one selected metadata}
\usage{
prep_dt_depmap_meta(meta_data_path, metadata_col = "PatientRace")
}
\arguments{
\item{meta_data_path}{string with path to metadata file describing all cancer models/cell lines}
}
\value{
A named list with elements, that may be input to \code{\link{prep_dt_assoc}}
\itemize{
  \item \code{dt_depmap} \code{data.table} with feature data from DepMap (wide format),
  \item \code{selected_feat_meta_col} string name of metadata column.
}
}
\description{
Load DepMap merged data for one selected metadata
}
\examples{
\dontshow{if (nzchar(system.file("depmap_data", package = "gDRtestData"))) withAutoprint(\{ # examplesIf}
meta_data_path <- system.file("depmap_data/Model.csv.gz", package = "gDRtestData")
dt_depmap_meta <- prep_dt_depmap_meta(meta_data_path)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
