% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_meta}
\alias{prep_dt_depmap_meta}
\title{Load DepMap merged data for one selected metadata}
\usage{
prep_dt_depmap_meta(metadata_col = "OncotreeLineage")
}
\arguments{
\item{metadata_col}{character vector with the metadata columns to load for DepMap cell lines}
}
\value{
A named list with elements, that may be input to \code{\link[gDRplots]{prep_dt_assoc}}
\itemize{
  \item \code{dt_depmap} \code{data.table} with feature data from DepMap (wide format),
  \item \code{selected_feat_meta_col} string name of metadata column..
}
}
\description{
Load DepMap merged data for one selected metadata
}
\examples{
\dontrun{
dt_depmap_meta <- prep_dt_depmap_meta() 
}
}
\seealso{
\code{kaleidoscope::load_depmap_merged}
}
\keyword{internal}
