% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_depmap_meta}
\alias{prep_dt_depmap_meta}
\title{Load DepMap merged data for one selected metadata}
\usage{
prep_dt_depmap_meta(metadata_col = "OncotreeLineage")
}
\arguments{
\item{metadata_col}{character vector with the metadata columns to load for DepMap cell lines}
}
\value{
\code{data.table} with meta data
}
\description{
Load DepMap merged data for one selected metadata
}
\examples{
\dontrun{
dt_depmap_meta <- prep_dt_depmap_meta() 
}
}
\seealso{
\code{\link[kaleidoscope]{load_depmap_list}}
}
\keyword{internal}
