% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-prism.R
\name{prep_dt_response_metric_diff}
\alias{prep_dt_response_metric_diff}
\title{Prep table with metric values for combination experiment}
\usage{
prep_dt_response_metric_diff(
  dt_metrics,
  d_name,
  d_name2,
  normalization_type = "RV",
  metric = "xc50",
  fit_source = "gDR",
  additional_cols = NULL
)
}
\arguments{
\item{dt_metrics}{\code{data.table} representing data from the \code{Metrics} assay,
outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
and combo \code{SummarizedExperiment}}

\item{d_name}{string with drug name to be plotted (identifiers \code{DrugName}), default is NULL}

\item{d_name2}{string with drug name to be plotted (identifiers \code{DrugName_2}), default is NULL}

\item{normalization_type}{string with normalization types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}

\item{metric}{string name of the combo metric;
one of: "hsa_score"("Bliss Excess GR" or "Bliss Excess RV" - respectively 
depending on \code{normalization_type}), "bliss_score" ("Bliss Score GR" or "Bliss Score RV")}

\item{fit_source}{string source name for metrics}
}
\value{
\code{data.table} with selected metric, input to \code{\link[gDRplots]{prep_dt_assoc}}
}
\description{
Prep table with metric values for combination experiment
}
\examples{
mae <- gDRutils::get_synthetic_data("combo_matrix_small")
se <- mae[[gDRutils::get_supported_experiments("combo")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
                                               assay_name = "Metrics")
d_name <- "drug_004"
d_name2 <- "drug_026"
dt_response <- prep_dt_response_metric_diff(dt_metrics, d_name, d_name2)
dt_response <- 
  prep_dt_response_metric_diff(dt_metrics, d_name, d_name2,
                               metric = c("xc50", "x_mean", "x_max"))

}
\keyword{prism_plots}
