% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_filename_path}
\alias{prep_filename_path}
\title{Prepare list of names or paths}
\usage{
prep_filename_path(
  plt_list,
  prefix = NULL,
  path_file = NULL,
  file_format = NULL
)
}
\arguments{
\item{plt_list}{A named list of plots. Names will be used as headings for plots/tab groups.
If unnamed, ordinal numbers will be used.  Can be nested lists for tabbed output.
If a nested list is provided, the inner lists should also be named.}

\item{prefix}{string}

\item{path_file}{string}

\item{file_format}{string}
}
\description{
Prepare list of names or paths
}
\examples{
\dontrun{
# Simple list of plots
plotlist <- lapply(unique(iris$Species), function(iris_name) {
  ggplot2::ggplot(iris[iris$Species == iris_name, c("Sepal.Length", "Sepal.Width")]) +
    ggplot2::geom_point(ggplot2::aes(x = Sepal.Length, y = Sepal.Width))
})

prep_filename_path(plt_list = plotlist,
                   prefix = "iris__",
                   path_file = file.path(".", "plots"))

# Nested list of plots for tabbed output
nested_plotlist <- list()
for (species in unique(iris$Species)) {
  nested_plotlist[[species]] <- list()
  nested_plotlist[[species]][["Sepal"]] <- 
    ggplot2::ggplot(iris[iris$Species == species, ],
                    ggplot2::aes(x = Sepal.Length, y = Sepal.Width)) + ggplot2::geom_point()
  nested_plotlist[[species]][["Petal"]] <- 
    ggplot2::ggplot(iris[iris$Species == species, ],
                    ggplot2::aes(x = Petal.Length, y = Petal.Width)) + ggplot2::geom_point()
}

prep_filename_path(plt_list = nested_plotlist, 
                   file_format = "png")
}

}
\seealso{
\code{\link[gDRplots]{prep_plot_chunk}}
}
\keyword{internal}
