% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{prep_hm_limits}
\alias{prep_hm_limits}
\title{Calculate limit}
\usage{
prep_hm_limits(num_vec, metric = "smooth", normalization_type = "GR")
}
\arguments{
\item{num_vec}{numeric vector}

\item{lower_cap}{numeric lower capping value}

\item{upper_cap}{numeric upper capping value}
}
\value{
capped limits (min and max) for given numeric vector
}
\description{
Calculate limit
}
\examples{
\dontrun{
vec <- c(-0.1, -0.3, 0, 0.5, Inf, NA)
prep_hm_limits(vec)
}

}
\keyword{internal}
