% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{prep_hm_limits}
\alias{prep_hm_limits}
\title{Calculate limit for combo heatmap with gDR assumptions}
\usage{
prep_hm_limits(num_vec, metric = "smooth", normalization_type = "GR")
}
\arguments{
\item{num_vec}{numeric vector}

\item{metric}{string name of combo exccess metric;
one of: "smooth", "hsa_excess", "bliss_excess"}

\item{normalization_type}{string with normalization_types to be selected
one of: "GR" ("GRvalue") or "RV" ("RelativeViability")}
}
\value{
capped limits (min and max) for given numeric vector
}
\description{
Calculate limit for combo heatmap with gDR assumptions
}
\examples{
\dontrun{
vec <- c(-0.1, -0.3, 0, 0.5, Inf, NA)
prep_hm_limits(vec)
prep_hm_limits(vec, metric = "hsa_excess")
}

}
\keyword{internal}
