% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{prep_hm_limits}
\alias{prep_hm_limits}
\title{Calculate limit}
\usage{
prep_hm_limits(num_vec, lower_cap = -0.25, upper_cap = 0.25)
}
\arguments{
\item{num_vec}{numeric vector}

\item{lower_cap}{numeric lower capping value}

\item{upper_cap}{numeric upper capping value}
}
\value{
limit for given numeric vector
}
\description{
Calculate limit
}
\examples{
\dontrun{
vec <- c(-5, -0.3, 0, Inf, NA)
prep_hm_limits(vec)
}

}
\keyword{internal}
