% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_nested_plot_chunk}
\alias{prep_nested_plot_chunk}
\title{Prepare markdown chunk based on the nested plots list}
\usage{
prep_nested_plot_chunk(plt_list, chunk_name, header_level = 2)
}
\arguments{
\item{plt_list}{named list with generated plots to be shown in tabs; list of plots in nested 
hierarchy, where last 4th level is plot and 3rd level is \code{normalization_type} describe by
one of: "GR" ("GR Value") or "RV" ("Relative Viability")#'}

\item{chunk_name}{string name of markdown chunk; preferable without spaces}

\item{header_level}{numeric level of markdown header - only for the first level}
}
\value{
list of character vector - input for \code{knitr::knit}
}
\description{
Function output should be generated with \code{knitr::knit(text = unlist(<result>))}
}
\examples{
\dontrun{
mae <- gDRutils::get_synthetic_data("small")
se <- mae[[gDRutils::get_supported_experiments("sa")]]
dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")

# help function
plot_col <- function(tab_plt, norm_type, col = "red") {
  tab_plt <- data.table::melt(
    data = tab_plt[normalization_type == norm_type][, c("rId", "xc50", "x_mean", "x_max")],
    id = "rId")
  plt <- ggplot2::ggplot(tab_plt, ggplot2::aes(x = variable, y = value)) +
    ggplot2::geom_col(fill = col)
  return(plt)
}

# creating nested list with plots
plotlist <- list()
ls_color <- c("darkred", "orange", "darkcyan")
for (drug in unique(dt_metrics$DrugName)) {
  for (cl in unique(dt_metrics$CellLineName)) {
    tab_plot <- dt_metrics[DrugName == drug & CellLineName == cl]
    
    plt_GR <- lapply(ls_color, function(col) plot_col(tab_plot, "RV", col))
    names(plt_GR) <- sprintf("\%s_\%s", "GR", ls_color)
    plt_RV <- lapply(ls_color, function(col) plot_col(tab_plot, "RV", col))
    names(plt_RV) <- sprintf("\%s_\%s", "RV", ls_color)
    
    plotlist[[drug]][[cl]][["RV"]] <- plt_RV
    plotlist[[drug]][[cl]][["GR"]] <- plt_GR
  }
}

prep_nested_plot_chunk(plotlist, "metric_value")

prep_nested_plot_chunk(plotlist, "metric_value")

}
}
\seealso{
\code{\link[knitr]{knit}}
}
\keyword{internal}
