% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{prep_plot_chunk}
\alias{prep_plot_chunk}
\title{Prepare markdown chunk based on the plots list}
\usage{
prep_plot_chunk(plt_list, chunk_name, header_level = 3)
}
\arguments{
\item{plt_list}{named list with generated plots to be shown in tabs}

\item{chunk_name}{string name of markdown chunk; preferable without spaces}

\item{header_level}{numeric level of markdown header}
}
\value{
list of character vector - input for \code{knitr::knit}
}
\description{
Function output should be generated with \code{knitr::knit(text = unlist(<result>))}
}
\examples{
plotlist <- lapply(unique(iris$Species), function(iris_name) {
  plot(iris[iris$Species == iris_name, c("Sepal.Length", "Sepal.Width")])
})
names(plotlist) <- unique(iris$Species)

prep_plot_chunk(plotlist, "iris")#'  

}
\seealso{
\code{\link[knitr]{knit}}
}
\keyword{internal}
