% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rmd.R
\name{save_plot}
\alias{save_plot}
\title{Save gDR plots to a specified path}
\usage{
save_plot(plt, path, format = "svg")
}
\arguments{
\item{plt}{A plot object; either a ggplot2 or pheatmap object.}

\item{path}{A string specifying the path where the plot should be saved.}

\item{format}{A string specifying the format for saving the plot; either "svg", "png", or "pdf". Default is "svg".}
}
\description{
Save gDR plots to a specified path
}
\examples{
tmp_dir <- file.path(tempdir(), "plot_dir")
dir.create(tmp_dir, showWarnings = FALSE)
p <- ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) + ggplot2::geom_point()
save_plot(plt = p, path = paste(tmp_dir, "mtcars_scatter", sep = "/"), format = "png")
}
\keyword{internal}
