% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-combo-heatmaps.R
\name{transform_log_conc}
\alias{transform_log_conc}
\title{Calculate limit}
\usage{
transform_log_conc(conc_vec)
}
\arguments{
\item{conc_vec}{numeric vector with concentration values}
}
\value{
numeric vector with log10 concentration values; log10 for concentration equal 0 (-Inf)
   is replaced with one step less in the dose dilution
}
\description{
Calculate limit
}
\examples{
\dontrun{
vec <- c(0, 0.003, 0.01, 0.03)
transform_log_conc(vec)
}

}
\keyword{internal}
